% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{adjacentSide}
\alias{adjacentSide}
\title{Simple heatmap helper to get a default adjacent side for another annotation}
\usage{
adjacentSide(side = c("top", "right", "bottom", "left"))
}
\arguments{
\item{side}{one of "top", "right", "bottom", or "left"}
}
\value{
character
}
\description{
Simple heatmap helper to get a default adjacent side for another annotation
}
\examples{
adjacentSide("top")
}
