% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_tax.R
\name{agg_tax}
\alias{agg_tax}
\title{Aggregate taxa if requested}
\usage{
agg_tax(ps, level, return = "all")
}
\arguments{
\item{ps}{phyloseq object}

\item{level}{NULL, 'given' or name of valid taxonomic rank (try phyloseq::rank_names(ps))}

\item{return}{choose which parts of list object to return}
}
\value{
list including phyloseq object and level argument value
}
\description{
This function is a simple wrapper around microbiome::aggregate_taxa.
agg_tax allows you to pass NULL, 'given' or 'asv', to the level argument which will NOT aggregate the taxa.
If you pass any valid rank name it will aggregate taxa with microbiome::aggregate_taxa.
}
\examples{
library(microbiome)
data("dietswap", package = "microbiome")
agg_tax(ps = dietswap, level = NULL)
agg_tax(ps = dietswap, level = "Phylum")
}
