% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrgation.R
\name{aggRows}
\alias{aggRows}
\alias{aggCols}
\title{Aggregate rows or columns by a statistic.}
\usage{
aggRows(
  x,
  f,
  fun = colMeans,
  verbose = TRUE,
  .parallel = FALSE,
  .paropts = furrr::furrr_options(seed = TRUE),
  ...
)

aggCols(
  x,
  f,
  fun = rowMeans,
  verbose = TRUE,
  .parallel = FALSE,
  .paropts = furrr::furrr_options(seed = TRUE),
  ...
)
}
\arguments{
\item{x}{a data frame or a matrix.}

\item{f}{a vector that defines elements to be aggregated, e.g. a vector
of gene symbols.}

\item{fun}{a function that takes a matrix and returns a vector of column
(for \code{aggRows()}) or row (for \code{aggCols()}) statistics.}

\item{verbose}{logical, should messages and execution time be printed?}

\item{.parallel}{should the computation be run in parallel? experimental!}

\item{.paropts}{options of the parallel workers, as specified by
\code{\link[furrr]{furrr_options}}.}

\item{...}{additional arguments passed to \code{fun}.}
}
\value{
a data frame or a matrix, dependent on the input. Rows are named
after unique elements of \code{f}
}
\description{
The functions aggregate rows or columns of a data frame or matrix that
correspond to the same feature (e.g. microarray probes or transcript
variants that map to the same gene) by a statistic such as arithmetic mean,
geometric mean, harmonic mean, median, minimum, maximum or sum.
}
