% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps-sampleAnnotation.R
\name{anno_cat}
\alias{anno_cat}
\title{Create colored rectangle annotations for categorical data}
\usage{
anno_cat(
  x,
  which,
  renamer = identity,
  col = distinct_palette(),
  width = NULL,
  height = NULL,
  box_col = "white",
  box_lwd = 0.5,
  border_col = NA,
  border_lwd = 1,
  legend = TRUE,
  legend_title = ""
)
}
\arguments{
\item{x}{data vector, treated as categorical}

\item{which}{Whether it is a column annotation or a row annotation?}

\item{renamer}{function renaming variable values for legend}

\item{col}{colors vector, at least as long as unique(x), optionally named by x levels}

\item{width}{grid unit object or NULL}

\item{height}{grid unit object or NULL}

\item{box_col}{colour of boxes around individual cells}

\item{box_lwd}{line width of boxes around individual cells}

\item{border_col}{colour of border around all cells}

\item{border_lwd}{line width of border around all cells}

\item{legend}{generate legend for this annotation
(attached as attribute of heatmap, and not automatically included in plot)}

\item{legend_title}{title for legend, if drawn}
}
\value{
AnnotationFunction
}
\description{
Similar to anno_simple but with individual boxes!
}
\examples{
library(ComplexHeatmap)
# draw the annotation without a heatmap, you will never normally do this!
vp <- viewport(width = 0.75, height = 0.75)

grid::grid.newpage()
pushViewport(vp)
cats <- letters[1:4]
draw(anno_cat(cats, which = "row"))

grid::grid.newpage()
pushViewport(vp)
draw(
  anno_cat(
    x = cats, col = structure(names = cats, 1:4), which = "column",
    box_col = "black", box_lwd = 5
  )
)

# developer note #
# list of annotations can be split and ordered (adding NULL makes a list)
# https://jokergoo.github.io/ComplexHeatmap-reference/book/a-list-of-heatmaps.html
# (section #4.8 concatenate-only-the-annotations)
grid::grid.newpage()
pushViewport(vp)
annoList <- rowAnnotation(
  hi = anno_cat(cats, which = "row", border_col = "black")
) +
  NULL
draw(object = annoList, row_split = c(1, 1:3), row_order = 4:1)
pushViewport(viewport(x = 0.6))
draw(anno_cat(cats, "row", legend_title = "abcd") \%>\% attr("Legend"))
}
