% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps-sampleAnnotation.R
\name{anno_cat_legend}
\alias{anno_cat_legend}
\title{Convenience function for generating levels of convenience.}
\usage{
anno_cat_legend(col, x = NULL, renamer = identity, title = "", ...)
}
\arguments{
\item{col}{vector of colors, named by all levels of data (e.g. x) or not named}

\item{x}{optional: vector of data to pair with unnamed col or check against named col}

\item{title}{title of legend}

\item{...}{
  Arguments passed on to \code{\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend}}
  \describe{
    \item{\code{labels}}{Labels corresponding to \code{at}. If it is not specified, the values of \code{at} are taken as labels.}
    \item{\code{nrow}}{For legend which is represented as grids, \code{nrow} controls number of rows of the grids if the grids are arranged into multiple rows.}
    \item{\code{ncol}}{Similar as \code{nrow}, \code{ncol} controls number of columns of the grids if the grids are arranged into multiple columns. Note at a same time only one of \code{nrow} and \code{ncol} can be specified.}
    \item{\code{by_row}}{Are the legend grids arranged by rows or by columns?}
    \item{\code{grid_height}}{The height of legend grid. It can also control the height of the continuous legend if it is horizontal.}
    \item{\code{grid_width}}{The width of legend grid. It can also control the width of the continuous legend if it is vertical.}
    \item{\code{gap}}{If legend grids are put into multiple rows or columns, this controls the gap between neighbouring rows or columns, measured as a \code{\link[grid]{unit}} object.}
    \item{\code{labels_gp}}{Graphic parameters for labels.}
    \item{\code{labels_rot}}{Text rotation for labels. It should only be used for horizontal continuous legend.}
    \item{\code{border}}{Color of legend grid borders. It also works for the ticks in the continuous legend.}
    \item{\code{type}}{Type of legends. The value can be one of \code{grid}, \code{points}, \code{lines} and \code{boxplot}.}
    \item{\code{direction}}{Direction of the legend, vertical or horizontal?}
    \item{\code{title_position}}{Position of title relative to the legend. \code{topleft}, \code{topcenter}, \code{leftcenter-rot} and \code{lefttop-rot} are only for vertical legend and \code{leftcenter}, \code{lefttop} are only for  horizontal legend.}
    \item{\code{title_gap}}{Gap between title and the legend body.}
  }}
}
\value{
a ComplexHeatmap Legend class object
}
\description{
Convenience function for generating levels of convenience.
}
\examples{
grid::grid.newpage()
draw(
  anno_cat_legend(
    col = c("ibd" = "blue", "nonibd" = "grey90"),
    renamer = toupper, title = "Hi there, I'm a title"
  )
)
}
