% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps-taxAnnotation.R
\name{anno_tax_box}
\alias{anno_tax_box}
\title{Helper to specify heatmap annotation for showing taxa abundance on boxplot}
\usage{
anno_tax_box(
  undetected = 0,
  only_detected = TRUE,
  trans = "compositional",
  zero_replace = 0,
  use_counts = TRUE,
  size = grid::unit(30, "mm"),
  border = TRUE,
  gp = grid::gpar(fill = "#CCCCCC"),
  ylim = NULL,
  extend = 0.05,
  outline = TRUE,
  box_width = 0.6,
  pch = 1,
  pointsize = grid::unit(0.5, "mm"),
  axis = TRUE,
  ...,
  data = NULL,
  taxa = NULL,
  which = NULL
)
}
\arguments{
\item{undetected}{the value above which taxa are classed as detected/present in a sample}

\item{only_detected}{only plot values for samples where the taxon abundance is > undetected}

\item{trans}{name of transformation suitable for tax_transform,
or a function calling tax_transform, and/or tax_scale,
(a function must take a phyloseq or ps_extra, and return one)}

\item{zero_replace}{zero_replace value for for tax_transform, ignored if trans is a function}

\item{use_counts}{try to retrieve counts from data object?}

\item{size}{width or height as a grid unit object}

\item{border}{Wether draw borders of the annotation region?}

\item{gp}{Graphic parameters for the boxes. The length of the graphic parameters should be one or the number of observations.}

\item{ylim}{Data ranges.}

\item{extend}{The extension to both side of \code{ylim}. The value is a percent value corresponding to \code{ylim[2] - ylim[1]}.}

\item{outline}{Whether draw outline of boxplots?}

\item{box_width}{Relative width of boxes. The value should be smaller than one.}

\item{pch}{Point style.}

\item{pointsize}{size of outlier points, as grid::unit() object}

\item{axis}{Whether to add axis?}

\item{...}{
  Arguments passed on to \code{\link[ComplexHeatmap:anno_boxplot]{ComplexHeatmap::anno_boxplot}}
  \describe{
    \item{\code{axis_param}}{parameters for controlling axis. See \code{\link[ComplexHeatmap]{default_axis_param}} for all possible settings and default parameters.}
  }}

\item{data}{OPTIONAL phyloseq or ps_extra,
only set this to override use of same data as in heatmap}

\item{taxa}{OPTIONAL selection vector of taxa (names, numbers or logical),
only set this if providing data argument to override default}

\item{which}{OPTIONAL indicating if it is a 'column' or a 'row' annotation,
only set this if providing data argument to override default}
}
\value{
function or ComplexHeatmap AnnotationFunction object
}
\description{
Use this as an argument to taxAnnotation(),
which itself is used by cor_heatmap and comp_heatmap as tax_anno argument.
}
\examples{
library("ComplexHeatmap")
data("ibd_phylo", package = "corncob")
psq <- tax_filter(ibd_phylo, min_prevalence = 5)
psq <- tax_mutate(psq, Species = NULL)
psq <- tax_fix(psq)
psq <- tax_agg(psq, rank = "Family")
taxa <- tax_top(psq, n = 15, rank = "Family")
# makes a function that takes data, taxa and which (at minimum)
fun <- anno_tax_box()
# manually specify the prevalence barplot function by giving it data etc.
heatmapAnnoFunction <- fun(data = psq, which = "column", taxa = taxa)
# draw the barplot without a heatmap, you will never normally do this!
vp <- viewport(width = 0.75, height = 0.75)
grid.newpage()
pushViewport(vp)
draw(heatmapAnnoFunction)

# let's change some style options and specify the data up front
grid::grid.newpage()
pushViewport(vp)
draw(anno_tax_box(
  data = psq, taxa = taxa, which = "row", pointsize = grid::unit(1, "mm"),
  gp = grid::gpar(fill = "red"), border = FALSE, box_width = 0.2
))

# clear drawings
grid::grid.newpage()
}
