% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps-varAnnotation.R
\name{anno_var_box}
\alias{anno_var_box}
\title{Helper to specify heatmap annotation for showing variable distributions}
\usage{
anno_var_box(
  fun = identity,
  size = grid::unit(30, "mm"),
  border = TRUE,
  gp = grid::gpar(fill = "#CCCCCC"),
  ylim = NULL,
  extend = 0.05,
  outline = TRUE,
  box_width = 0.6,
  pch = 1,
  pointsize = grid::unit(0.5, "mm"),
  axis = TRUE,
  ...,
  data = NULL,
  vars = NULL,
  which = NULL
)
}
\arguments{
\item{fun}{function applied to all variables, with apply()}

\item{size}{width or height as a grid unit object}

\item{border}{Wether draw borders of the annotation region?}

\item{gp}{Graphic parameters for the boxes. The length of the graphic parameters should be one or the number of observations.}

\item{ylim}{Data ranges.}

\item{extend}{The extension to both side of \code{ylim}. The value is a percent value corresponding to \code{ylim[2] - ylim[1]}.}

\item{outline}{Whether draw outline of boxplots?}

\item{box_width}{Relative width of boxes. The value should be smaller than one.}

\item{pch}{Point style.}

\item{pointsize}{size of outlier points, as grid::unit() object}

\item{axis}{Whether to add axis?}

\item{...}{
  Arguments passed on to \code{\link[ComplexHeatmap:anno_boxplot]{ComplexHeatmap::anno_boxplot}}
  \describe{
    \item{\code{axis_param}}{parameters for controlling axis. See \code{\link[ComplexHeatmap]{default_axis_param}} for all possible settings and default parameters.}
  }}

\item{data}{OPTIONAL phyloseq or ps_extra,
only set this to override use of same data as in heatmap}

\item{vars}{OPTIONAL selection vector of variable names,
only set this if providing data argument to override default}

\item{which}{OPTIONAL indicating if it is a 'column' or a 'row' annotation,
only set this if providing data argument to override default}
}
\value{
function or ComplexHeatmap AnnotationFunction object
}
\description{
Use this as an argument to varAnnotation(),
which itself is used by cor_heatmap as var_anno() argument.
}
\examples{
library(ComplexHeatmap)
set.seed(123)
fakeData <- as.data.frame.matrix(matrix(rnorm(500, 10, 3), ncol = 10))
names(fakeData) <- paste0("var_", 1:10)

# draw the boxplot without a heatmap, you will never normally do this!
vp <- viewport(width = 0.75, height = 0.75)
grid.newpage()
pushViewport(vp)
draw(
  anno_var_box(data = fakeData, vars = names(fakeData), which = "column")
)

grid.newpage()
pushViewport(vp)
draw(
  anno_var_box(
    data = fakeData, fun = function(x) log(x + 1),
    vars = rev(names(fakeData)),
    which = "row"
  )
)
}
