% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{anova_tester}
\alias{anova_tester}
\alias{nb_tester}
\title{One-way ANOVA and negative binomial testing.}
\usage{
anova_tester(data, split_fct, variable, confounder = NULL, ...)

nb_tester(
  data,
  split_fct,
  variable,
  confounder = NULL,
  dev_test = "Chisq",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variable}{dependent variable name.}

\item{confounder}{name of the confounding variable, NULL by default.}

\item{...}{additional arguments passed to \code{\link[stats]{aov}} or
\code{\link[MASS]{glm.nb}}.}

\item{dev_test}{name of the test used for analysis of deviance. One of:
'Rao', 'LRT', 'Chisq' (default), 'F', 'Cp'.}
}
\value{
a list of data frames with ANOVA results or results of analysis of
deviance (F/test statistic, degrees of freedom and p value) and
linear model coefficient estimates (estimates, test statistic, errors and
confidence intervals, pooled standard deviation and Cohen's d).
}
\description{
Simplified versions of one-way ANOVA and negative binomial (NB)
modeling suitable for whole-genome testing.
}
