% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_estimates.R
\name{avg_deviation}
\alias{avg_deviation}
\title{Average deviation from the mean or median.}
\usage{
avg_deviation(
  data,
  split_fct,
  variables,
  grand_center = c("mean", "median", "gmean", "hmean"),
  split_center = c("mean", "median", "gmean", "hmean"),
  ci = c("distr", "percentile", "bca"),
  conf_level = 0.95,
  adj_method = "BH"
)
}
\arguments{
\item{data}{a data frame.}

\item{split_fct}{name of the splitting factor.}

\item{variables}{names of numeric variables, for which deviations will be
computed.}

\item{grand_center}{statistic of the grand central tendency, see Details.}

\item{split_center}{statistic of the central tendency in subsets defined
by \code{split_fct}.}

\item{ci}{type of 95\% confidence intervals, as specified in Details.}

\item{conf_level}{width of the confidence interval.}

\item{adj_method}{method of multiple testing adjustment.
See: \code{\link[stats]{p.adjust}}.}
}
\value{
A data frame with the grand averages (\code{grand_center}), average deviations
from the grand average (\code{deviation_center}), standard deviations and
standard errors (\code{deviation_sd} and \code{deviation-sem}), t statistic, degrees of
freedom (\code{df}), limits of the 95\% confidence interval (\code{lower_ci} and
\code{upper_ci}), raw and multiple testing-adjusted one-sample T test p values
(\code{p_value} and \code{p_adjusted}) for variables split by the user-defined
splitting feature.
}
\description{
The function computes deviations from the grand measure of central tendency
and spits them by an user defined factor.
}
\details{
A bunch of metrics of the central tendency is available, both for
computation of the grand statistic (argument \code{grand_center}) and averaging
of the the deviations (argument \code{split_center}). They are: arithmetic mean
('mean', default), median ('median'), geometric mean ('gmean') and
harmonic mean ('hmean'). Note that, not all of those statistic are
appropriate for all types of numeric data: e.g. geometric mean for a mixture
of positive and negative values.
The function computes also the Student's t statistic for the difference
of the averaged deviations and the grand mean as well as p-value of a single
sample t test. Prior to using those values, checking for the normality
assumption is highly recommended.
Finally, 95\% confidence intervals are returned for three scenarios. For
\code{ci = 'distr'}, confidence intervals of the average deviation from the grand
mean are computed based on the critical values of the t statistic and
standard errors. By contrast, \code{ci = 'percentile'} and \code{ci = 'bca'} calculate
95\% CI for single deviations from the grand means with the percentile and
bias-corrected method, respectively.
}
