% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_disper.R
\name{beta_disper}
\alias{beta_disper}
\title{Wrapper for vegan::betadisper}
\usage{
beta_disper(
  data,
  variables,
  method = c("centroid", "median")[[1]],
  complete_cases = TRUE,
  verbose = TRUE,
  return = "all"
)
}
\arguments{
\item{data}{list output from calc_dist}

\item{variables}{list of variables to use as group}

\item{method}{centroid or median}

\item{complete_cases}{drop samples with NAs in any of the variables listed}

\item{verbose}{sends messages about progress if true}

\item{return}{what parts of return list to return, defaults to all parts}
}
\value{
list containing betadisper results list and input objects
}
\description{
Takes the output of calc_dist function. Or use with the result of the permanova function to ensure the results correspond to exactly the same input data.
Runs betadisper for all categorical variables in variables argument.
See help for vegan::betadisper.
}
\examples{
library(phyloseq)
library(vegan)
data("dietswap", package = "microbiome")

# add some missings to demonstrate automated removal
sample_data(dietswap)$sex[3:6] <- NA
# create a numeric variable to show it will be skipped with a warning
sample_data(dietswap)$timepoint <- as.numeric(sample_data(dietswap)$timepoint)

# straight to the betadisp
bd1 <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  calc_dist("aitchison") \%>\%
  beta_disper(variables = c("sex", "bmi_group", "timepoint"))
# quick vegan plotting methods
plot(bd1$beta_disper$sex$model, label.cex = 0.5)
boxplot(bd1$beta_disper$sex$model)

# compute distance and use for both permanova and beta_disper
testDist <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  calc_dist("bray")

PERM <- testDist \%>\%
  permanova(
    variables = c("sex", "bmi_group"),
    n_processes = 1, n_perms = 99
  )
str(PERM, max.level = 1)

bd <- PERM \%>\% beta_disper(variables = c("sex", "bmi_group"))
bd$beta_disper$bmi_group
}
