% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distances between pairs of samples in phyloseq object}
\usage{
calc_dist(
  data,
  dist = c("bray", "gunifrac", "unifrac", "wunifrac", "va-wunifrac", "aitchison",
    "euclidean")[1],
  return = "all",
  gunifrac_alpha = 0.5
)
}
\arguments{
\item{data}{phyloseq object or list object output from agg_tax}

\item{dist}{name of distance to calculate between pairs of samples}

\item{return}{choose which parts of list object to return}

\item{gunifrac_alpha}{setting alpha value only relevant if gunifrac distance used}
}
\value{
list with distance matrix, phyloseq object, and name of distance used
}
\description{
For use directly with phyloseq object or with list output of agg_tax.
Computes various unifrac distances, aitchison distance or any distance from vegdist.
Returns phyloseq and name of distance in addition to the distance matrix itself!
Result intended to be piped into ordin8 or permanova functions.
}
\details{
Note: Instead of calculating aitchison distance with this function, users should prefer directly applying microbiome::transform(ps, 'clr') and using pca/rda method of ordin8 to allow plotting loading vectors in plot_ordin8
}
\examples{
# bray curtis distance on genera-level features
data("dietswap", package = "microbiome")
bc <- dietswap \%>\%
  agg_tax("Genus") \%>\%
  calc_dist("aitchison")
class(bc)

# gunifrac distance using phyloseq input
data("esophagus", package = "phyloseq")
gunifrac <- esophagus \%>\%
  calc_dist("gunifrac", return = "distMat")

class(gunifrac)

}
