% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-accessors.R
\name{check_otutable_is_counts}
\alias{check_otutable_is_counts}
\title{Internal helper for ps_counts}
\usage{
check_otutable_is_counts(otu, warn)
}
\arguments{
\item{otu}{A phyloseq otu_table object.}

\item{warn}{A boolean or "error" string to control warning or error behavior.}
}
\description{
Internal helper for ps_counts
}
\keyword{internal}
