% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{colMedians}
\alias{colMedians}
\alias{colGmeans}
\alias{colHmeans}
\alias{colVars}
\alias{colSDs}
\alias{colGini}
\alias{distr_stats}
\title{Compute statistics for numeric columns of a data frame.}
\usage{
colMedians(x, na.rm = TRUE, .parallel = FALSE)

colGmeans(x, na.rm = TRUE, .parallel = FALSE)

colHmeans(x, na.rm = TRUE, .parallel = FALSE)

colVars(x, na.rm = TRUE, .parallel = FALSE)

colSDs(x, na.rm = TRUE, .parallel = FALSE)

colGini(x, na.rm = TRUE, .parallel = FALSE)

distr_stats(x, na.rm = TRUE, ..., .parallel = FALSE)
}
\arguments{
\item{x}{a numeric data frame.}

\item{na.rm}{logical: should NAs be removed?}

\item{.parallel}{logical, should the computation be done in parallel? Makes
sense for large datasets.}

\item{...}{extra arguments passed to \code{\link[caret]{nearZeroVar}}.}
}
\value{
\code{colMedians()}, \code{colGmeans()}, \code{colHmeans()}, \code{colVars()},
\code{colSDs()}, \code{colGini()} return numeric vectors with, respectively,
column medians, geometric means, harmonic means, variances, standard
deviations and Gini coefficients.
\code{distr_stats()} returns a data frame with variable names, their means,
variances, Gini coefficients, variance to mean ratios and the output of the
\code{caret} function \code{\link[caret]{nearZeroVar}}.
}
\description{
The functions compute column medians, variances, standard deviations and
Gini coefficients. \code{distr_stats} computes a bunch of distribution stats that
are helpful at selection of variant features for the further analysis, e.g.
differential gene expression.
}
