% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_similarity.R
\name{components.set_sim}
\alias{components.set_sim}
\alias{plot.set_sim}
\title{Reduction analysis and visualization of similarities between sets.}
\usage{
\method{components}{set_sim}(
  object,
  type = c("mds", "umap", "hcl"),
  k = 2,
  trans_fun = function(x) 1 - x,
  method = "complete",
  ...
)

\method{plot}{set_sim}(
  x,
  type = c("heat_map", "mds", "umap", "dendrogram"),
  trans_fun = function(x) 1 - x,
  method = "complete",
  point_size = 2,
  point_alpha = 1,
  point_color = "steelblue",
  point_wjitter = 0,
  point_hjitter = 0,
  order_by_similarity = c("none", "descending", "ascending"),
  tile_rim = "black",
  show_txt = TRUE,
  txt_size = 2.75,
  txt_color = "black",
  txt_alpha = 1,
  txt_face = "plain",
  labeller = identity,
  cust_theme = microViz::theme_micro(),
  ...
)
}
\arguments{
\item{object}{a similarity matrix returned by \code{\link{set_similarity}}.}

\item{type}{type of dimensionality reduction or plot.}

\item{k}{number of dimensions.}

\item{trans_fun}{function used to transform similarities to distances.}

\item{method}{algorithm of hierarchical clustering, see:
\code{\link[stats]{hclust}} for details.}

\item{...}{for \code{components()}: additional arguments passed to
\code{\link[stats]{cmdscale}}, \code{\link[umap]{umap}},
or \code{\link[stats]{hclust}};
for \code{plot()}: extra arguments passed to \code{\link{components.set_sim}} or
\code{\link[ggdendro]{ggdendrogram}}.}

\item{x}{a similarity matrix returned by \code{\link{set_similarity}}.}

\item{point_size}{size of the data points.}

\item{point_alpha}{alpha of the the points.}

\item{point_color}{color of the points.}

\item{point_wjitter}{horizontal jittering of the points.}

\item{point_hjitter}{vertical jittering of the points.}

\item{order_by_similarity}{ordering of the heat map tiles by descending of
ascending similarities. 'none' (default): the set order corresponds to the
order of the sets in the input matrix.}

\item{tile_rim}{color of rims of tiles of the heat map.}

\item{show_txt}{logical, should names of the sets be displayed in the plots?}

\item{txt_size}{size of the set names label text.}

\item{txt_color}{color of the text.}

\item{txt_alpha}{alpha of the text.}

\item{txt_face}{face of the text.}

\item{labeller}{function used to transform set names prior to displaying them
in the plot.}

\item{cust_theme}{custom \code{ggplot} theme.}
}
\value{
\code{components()} returns a data frame with the MDS or UMAP layout
dimensions, or a \code{hclust} class object. \code{plot()} produces a \code{ggplot} object.
}
\description{
\code{components()} and \code{plot()} methods for matrices of similarity between sets
generated by \code{\link{set_similarity}}.
\code{components()} performs dimensionality reduction of the similarity matrix
with MDS (\code{type = 'mds'}: multi-dimensional scaling) or UMAP
(\code{type = 'umap'}: uniform manifold approximation and projection),
or hierarchical clustering (\code{type = 'hcl'}).
\code{plot()} visualizes distances between the sets in form of heat map, MDS or
UMAP layouts, or as a dendrogram.
}
\details{
UMAP is computed with \code{\link[umap]{umap}}.
}
\references{
McInnes L, Healy J, Melville J. UMAP: Uniform Manifold Approximation and
Projection for Dimension Reduction. (2018)
Available at: https://arxiv.org/abs/1802.03426v3.

Konopka T. umap: Uniform Manifold Approximation and Projection. (2022)
Available at: https://cran.r-project.org/web/packages/umap/index.html
}
