% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_test.R
\name{cor_test}
\alias{cor_test}
\title{Simple wrapper around cor.test for y ~ x style formula input}
\usage{
cor_test(formula, data, ...)
}
\arguments{
\item{formula}{a formula in form y ~ x}

\item{data}{dataframe}

\item{...}{passed to cor.test}
}
\description{
Intended for use within the function tax_model
}
\examples{
data("shao19")
ps <- shao19 \%>\%
  ps_filter(family_role == "mother") \%>\%
  tax_filter(min_prevalence = 20) \%>\%
  tax_agg("family")

cors <- ps \%>\% tax_model(
  rank = "family", variables = list("age", "number_reads"), type = cor_test
)

}
