% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{count_binary}
\alias{count_binary}
\title{Frequency of binary data.}
\usage{
count_binary(data, variables = NULL, split_fct = NULL)
}
\arguments{
\item{data}{a data frame with the splitting factor and binary variables of
interest.}

\item{variables}{a character vector with names of the binary variables.
If \code{NULL}, the argument defaults to all variables in the data frame
except of \code{split_fct}.}

\item{split_fct}{name of the splitting factor variable. If \code{NULL}, global
frequencies in the data set will be quantified.}
}
\value{
a data frame with the variable names (\code{variable}), numbers of
1-coded events (\code{n}), total numbers of observations (\code{n_total}), and
percentage of events within all observations (\code{percent}).
}
\description{
Computes frequencies and percentages for binary data split by a factor.
}
\details{
The binary data can be provided in a 0/1 form or as two-level factors.
}
