% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{count_features}
\alias{count_features}
\alias{shared_features}
\title{Tally and identify shared elements in sets.}
\usage{
count_features(x)

shared_features(x, m = 2)
}
\arguments{
\item{x}{a list of character or numeric vectors.}

\item{m}{number of sets which share the common features.}
}
\value{
\code{shared_features()}: a vector with the features shared by the given
number of elements.
\code{count_features()}: a data frame with the following columns:
\itemize{
\item \code{element}: element names
\item \code{n}: number of sets, where the element was present
\item \code{n_total}: total number of sets
\item \code{percent}: percentage of sets, where the given element was present
}
}
\description{
The functions count and select elements shared by at least n elements of
a list.
As such they may be useful at identifying e.g. significantly regulated genes
or pathways in several data sets.
}
