% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{delta}
\alias{delta}
\title{Difference of averages.}
\usage{
delta(
  data,
  split_fct,
  variables = NULL,
  average_fun = colMeans,
  extreme = TRUE
)
}
\arguments{
\item{data}{a data frame.}

\item{split_fct}{name of the factor variable used for splitting.}

\item{variables}{a vector of names of variables for which the statistics
will be computed. If \code{NULL}, all variables except of \code{split_factor} are
used.}

\item{average_fun}{a function for averaging numeric variables.}

\item{extreme}{logical, should difference between the last and first level
of the splitting factor be returned. If \code{TRUE} (default), a single numeric
vector of differences of averages between the last and fist levels of the
splitting factor is returned. If \code{FALSE}, a list of numeric vector is
returned; each of the vector contains differences between the particular
level and the first level of the splitting factor.}
}
\value{
a numeric vector or a list of numeric vectors.
}
\description{
Calculates differences between avarages of numeric variables split by a
factor  variable.
}
