% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plotting.R
\name{draw_estimate_scatter}
\alias{draw_estimate_scatter}
\title{Scatter/Forest plot of estimates.}
\usage{
draw_estimate_scatter(
  data,
  label_variable,
  regulation_variable,
  p_variable = NULL,
  signif_level = 0.05,
  regulation_level = 0,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  central_stat = c("none", "mean", "median", "gmean", "hmean"),
  error_stat = c("none", "percentile_ci", "bca_ci", "iqr"),
  conf_level = 0.95,
  point_size = 2,
  point_alpha = 0.75,
  point_wjitter = 0,
  point_hjitter = 0.1,
  center_symbol_size = 3,
  stat_color = "orangered3",
  error_bar_size = 0.75,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "Regulation",
  y_lab = NULL,
  fill_lab = regulation_variable,
  plot_tag = NULL,
  cust_theme = microViz::theme_micro()
)
}
\arguments{
\item{data}{a data frame with analysis results.}

\item{label_variable}{name of the variable storing feature names.}

\item{regulation_variable}{name of the variable storing the estimate to
be plotted.}

\item{p_variable}{optional, name of the variable storing statistical test
results.}

\item{signif_level}{significance cutoff used for identification of
significant estimates, ignored if \code{p_variable = NULL}.}

\item{regulation_level}{estimate value cutoff used for identification of
significant estimates.}

\item{fill_scale}{colors coding for the estimate status.}

\item{central_stat}{measure of central tendency to be plotted: none, mean,
median, geometric mean ('gmean') or harmonic mean ('hmean').}

\item{error_stat}{measure of distribution to be plotted as a horizontal bar:
none, percentile confidence interval, BCA (bias-corrected and accelerated)
confidence interval , or interquartile range ('iqr').}

\item{conf_level}{width of the confidence interval.}

\item{point_size}{size of the data points.}

\item{point_alpha}{alpha of the data points.}

\item{point_wjitter}{data point jittering width.}

\item{point_hjitter}{data point jittering height.}

\item{center_symbol_size}{size of the symbol representing the central
tendency measure.}

\item{stat_color}{color of the symbols of the central tendency measure and
the error bars.}

\item{error_bar_size}{size of the error bar.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{fill_lab}{fill scale title.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{custom \code{ggplot} theme.#'}
}
\value{
a \code{ggplot} object.
}
\description{
Draws a scatter plot of estimates specified by the user. Point color codes
for significance and regulation status. As an option, a central statistic
over all estimates such as mean or median with an error measure
(SD, SEM, etc.) may be displayed.
}
