% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{extract_mutations}
\alias{extract_mutations}
\title{Binarize somatic mutation data.}
\usage{
extract_mutations(
  data,
  sample_id = "Tumor_Sample_Barcode",
  gene_id = "Hugo_Symbol",
  as_numeric = TRUE
)
}
\arguments{
\item{data}{a data frame with detailed mutation information obtained e.g.
from cBioportal.}

\item{sample_id}{name of the variable in \code{data} that stores sample
identifiers.}

\item{gene_id}{name of the variable in \code{data} that stores gene identifiers
such as official HUGO symbols or Entrez IDs.}

\item{as_numeric}{logical, should absence/presence of mutations be 0/1 coded?}
}
\value{
a data frame. The variable \code{sample_id} that stores sample
identifiers. The variables named after gene identifiers retrieved from the
\code{gene_id} variable of data store binarized indexes of presence of at least
one somatic mutation.
}
\description{
Converts a data frame with detailed mutation information
(e.g. nucleotide sequence, variant classification) into a simplified
binarized form.
In the output, presence of at least one mutation in a particular gene is
coded as 1 or 'mutated'. Absence of mutations in a gene is
coded as 0 or 'WT'.
Gene identifiers are stored in columns, sample identifiers in rows.
}
\details{
The function bottlenecks will run in parallel if a parallel backend is
declared with \code{\link[future]{plan}}.
}
