% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{fast_num_stats}
\alias{fast_num_stats}
\title{Fast computation of descriptive distribution statistics for numeric features.}
\usage{
fast_num_stats(
  data,
  split_fct,
  variables = NULL,
  format = c("median", "mean", "full"),
  signif_digits = 2
)
}
\arguments{
\item{data}{a data frame as describe above.}

\item{split_fct}{name of the factor variable used for splitting.}

\item{variables}{a vector of names of variables for which the statistics
will be computed. If \code{NULL}, all variables except of \code{split_factor} are
used.}

\item{format}{format of the output data frame cells. \code{median} (default):
median, interquartile range, and range; \code{mean}: mean, standard deviation,
and range; \code{full}: combination of the \code{mean} and \code{median} formats.}

\item{signif_digits}{significant digits used for the output.}
}
\value{
a data frame.
}
\description{
The function takes a data frame with numeric variables and a splitting
variable in factor format, and returns a bunch of statistics of
distribution of the numeric variables such as median with interquartile
range and range, or mean, standard deviation and range.
Numbers of complete observations are returned as well.
The output is publication styled in a way that is accepted with most
scientific journals.
}
\details{
If any \code{NA} are detected, cells of the output data frame are appendend with
the information on numbers of complete observations.
}
