% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{gene_sem}
\alias{gene_sem}
\alias{go_sem}
\title{Semantic similarity between GO terms or genes.}
\usage{
gene_sem(
  x,
  semData,
  measure = "Wang",
  drop = "IEA",
  combine = "BMA",
  as_dist = TRUE,
  pad_na = NULL,
  .parallel = FALSE
)

go_sem(
  x,
  semData,
  measure = "Wang",
  as_dist = TRUE,
  pad_na = NULL,
  .parallel = FALSE
)
}
\arguments{
\item{x}{a vector of Entrez IDs, gene symbols or genes or GO IDs, specific
for the \code{semData} argument.}

\item{semData}{a GOSemSimDATA object, created by e.g.
\code{\link[GOSemSim]{godata}}.}

\item{measure}{distance measure, please consult
\code{\link[GOSemSim]{geneSim}} or \code{\link[GOSemSim]{goSim}}.}

\item{drop}{evidence rules for dropping annotations, please consult
\code{\link[GOSemSim]{geneSim}} or \code{\link[GOSemSim]{goSim}}.}

\item{combine}{a method for combining semantic similarity scores, please
consult \code{\link[GOSemSim]{geneSim}} or \code{\link[GOSemSim]{goSim}}}

\item{as_dist}{logical, if \code{TRUE} an object of the \code{dist} class is returned.}

\item{pad_na}{\code{NULL} or a numeric value used to replace any NAs in the output
matrix. If \code{NULL}, NAs are left as they are.}

\item{.parallel}{logical, should the analysis be run in parallel?}
}
\value{
a numeric matrix or an object of the \code{dist} class containing pairwise
distances or similarities between the genes or GO terms-
}
\description{
The \code{gen_sem()} and \code{GO_sem()} functions compute semantic similarities
between genes and gene ontology (GO) terms using the interface provided
by the \code{GOSemSim} package. Given a vector of gene Entrez ID or GO IDs,
a similarity or distance matrix is returned.
}
\details{
Please refer to the
GOSemSim package manual, and to the downstream functions
\code{\link[GOSemSim]{geneSim}} and \code{\link[GOSemSim]{goSim}}.
The parallel backend is provided by the \code{furrr}
package tools and may not work with some platforms. Distances are computed
with the \code{distance = 1 - similarity} formula.
}
\references{
Yu G, Li F, Qin Y, Bo X, Wu Y, Wang S. GOSemSim: an R package for measuring
semantic similarity among GO terms and gene products. Bioinformatics (2010)
26:976–978. doi:10.1093/BIOINFORMATICS/BTQ064
}
