% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{Gmean}
\alias{Gmean}
\alias{Hmean}
\alias{Gini}
\alias{perCI}
\alias{bcaCI}
\alias{freqRatio}
\alias{percUnique}
\title{Distribution statistics for numeric vectors and matrices.}
\usage{
Gmean(x, na.rm = TRUE)

Hmean(x, na.rm = TRUE)

Gini(x, unbiased = TRUE, na.rm = TRUE)

perCI(x, conf_level = 0.95, na.rm = TRUE)

bcaCI(x, conf_level = 0.95, na.rm = TRUE)

freqRatio(x)

percUnique(x)
}
\arguments{
\item{x}{a numeric vector or a numeric matrix.}

\item{na.rm}{logical, should \code{NAs} ne removed prior to the statistic
computation?}

\item{unbiased}{logical, should estimated of the population Gini coefficient
be returned (multiplied by \code{n/(n - 1)})?}

\item{conf_level}{a single numeric with the width of the confidence interval.}
}
\value{
a single numeric or a numeric vector, see Details.
}
\description{
A bunch of functions that compute extra numeric statistics not provided by
base R such as geometric and harmonic means or BCA confidence intervals.
}
\details{
\code{Gmean()}, \code{Hmean()} and \code{Gini()} compute geometric mean, harmonic mean,
and Gini coefficient, respectively.
\code{perCI()} and \code{bcaCI()} return confidence intervals calculated with the
percentile and BCA (bias corrected and accelerated) method.
\code{freqRatio()} computes the ratio of frequencies of the first most frequent
element of a vector to the second most frequent one.
\code{percUnique()} returns the percentage of unique observations.
The \code{freqRatio()} and \code{percUnique()} functions skip silently any \code{NA} in the
vector.
}
\references{
DiCiccio TJ, Efron B. Bootstrap confidence intervals.
https://doi.org/101214/ss/1032280214 (1996) 11:189–228.
doi:10.1214/SS/1032280214
}
