% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{GOana}
\alias{GOana}
\title{Gene Ontology enrichment analysis.}
\usage{
GOana(
  de,
  universe,
  ontology = c("all", "BP", "CC", "MF"),
  adj_method = "BH",
  ...
)
}
\arguments{
\item{de}{a vector with Entrez IDs of differentially regulated genes.}

\item{universe}{an obligatory vector of Entrez IDs for all investigated genes.}

\item{ontology}{specifies ontology of the GO term to be returned. If \code{all}
(default), all possible GO terms are returned.}

\item{adj_method}{method of adjustment for multiple testing, \code{BH}
(Benjamini-Hochberg or FDR) by default.}

\item{...}{extra arguments passed to \code{\link[limma]{goana}}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{go_id}: GO term identifier
\item \code{term}: name of the GO term
\item \code{n_go_dge}: number of the differentially regulated genes associated with
the particular GO term
\item \code{n_go_total}: total number of genes associated with the particular GO term
\item \code{n_dge}: number of the differentially regulated genes.
\item \code{n_total}: number of all analyzed genes
\item \code{or}: OR for the GO term enrichment over the frequency in the universe
\item \code{p_value}: raw, uncorrected p value of the enrichment, as returned
by \code{\link[limma]{goana}}.
\item \code{p_adjusted}: p value corrected for multiple testing
}
}
\description{
The function performs Gene Ontology (GO) enrichment analysis with the
\code{limma} function \code{\link[limma]{goana}} and enhances its results by
computing odds ratios (OR) for enrichment of the differentially regulated
genes associated with the particular GO term over the genome or universe
frequency.
}
\details{
The OR is computed with the following formula:

\deqn{OR = \frac{N_{GO DGE} \times N_{universe}}{N_{GO} \times N_{DGE}}}

where \eqn{N_{GO DGE}} stands for the number of differentially regulated
genes associated with the particular GO term, \eqn{N_{universe}} is the
number of all investigated genes, \eqn{N_{GO}} is the number of all genes
associated with the particular GO, and \eqn{N_{DGE}} represents the total
number of differentially regulated genes.
As such, the OR metric may be used as a measure of effect size of the
enrichment. See: \code{\link[effectsize]{interpret_oddsratio}} for details.
}
