% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_colors}
\alias{heat_colors}
\title{Easy palettes for ComplexHeatmap}
\usage{
heat_colors(
  palette = "Green-Orange",
  breaks = c(-1, 0, 1),
  rev_palette = FALSE
)
}
\arguments{
\item{palette}{named palette from colorspace::hcl_palettes(type = "diverging") or a vector of 3 colour names (low, middle, high)}

\item{breaks}{numeric vector of values for colour scale breaks (low, middle, high)}

\item{rev_palette}{reverse the colorspace palette}
}
\description{
Pass a named diverging colorspace hcl palette to circlize::colorRamp2
}
