% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_grid}
\alias{heat_grid}
\title{set options for drawing gridlines on heatmaps}
\usage{
heat_grid(
  col = "white",
  alpha = 1,
  lty = 1,
  lwd = 0.5,
  lex = 1,
  lineend = "round",
  linejoin = "round"
)
}
\arguments{
\item{col}{Colour for lines and borders.}

\item{alpha}{Alpha channel for transparency}

\item{lty}{Line type}

\item{lwd}{Line width}

\item{lex}{Multiplier applied to line width}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}
}
\description{
set options for drawing gridlines on heatmaps
}
