% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_numbers}
\alias{heat_numbers}
\title{helps for drawing numbers on heatmap tiles}
\usage{
heat_numbers(
  decimals = 1,
  fmt = "\%.1f",
  fontsize = 7,
  col = "black",
  fontface = "plain",
  ...
)
}
\arguments{
\item{fmt}{number print format}

\item{fontsize}{fontsize specification,}

\item{what}{what numbers to print (currently only "values" is supported)}
}
\description{
helps for drawing numbers on heatmap tiles
}
