% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{tax_anno}
\alias{tax_anno}
\alias{anno_prev}
\alias{anno_abund}
\title{Create list describing taxa annotation}
\usage{
tax_anno(undetected = 0, which = "row", prev = 1, abund = 2, ...)

anno_prev(data, taxa, undetected = 0, which = "row", ...)

anno_abund(data, taxa, undetected = 0, which = "row", ...)
}
\arguments{
\item{undetected}{value above which taxa are considered present/detected in a sample}

\item{which}{"row" or "column" annotation?}

\item{prev}{order in which prevalence annotation shown (number, or NA)}

\item{abund}{order in which abundance annotation shown (number, or NA)}

\item{...}{further args passed HeatmapAnnotation}
}
\description{
main usefulness of tax_anno function is to give sane defaults and prompt user which annotation options are available
}
