% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cor_heatmap}
\alias{cor_heatmap}
\alias{heat_colors}
\alias{heat_numbers}
\alias{taxAnnotation}
\title{Microbe-to-sample-data correlation heatmap}
\usage{
cor_heatmap(
  data,
  taxa = NA,
  tax_anno = NULL,
  vars = NA,
  cor = c("pearson", "kendall", "spearman")[1],
  cor_use = "everything",
  colors = heat_colors(),
  numbers = heat_numbers(),
  taxa_which = "row",
  seriation_method = "OLO_ward",
  seriation_dist = "euclidean",
  seriation_method_col = seriation_method,
  seriation_dist_col = seriation_dist,
  tax_transform = "identity",
  ...
)

heat_colors(
  palette = "Green-Orange",
  breaks = c(-1, 0, 1),
  rev_palette = FALSE
)

heat_numbers(what = "values", fmt = "\%.1f", fontsize = 7, ...)

taxAnnotation(
  data,
  taxa,
  undetected = 0,
  which = "row",
  prev = TRUE,
  abun = TRUE,
  ...
)
}
\arguments{
\item{data}{phyloseq or phyloseq extra}

\item{taxa}{list of taxa to include, or NA for all}

\item{tax_anno}{taxAnnotation output describing taxa distribution or NULL}

\item{vars}{selection of variable names from sample_data}

\item{cor}{name of correlation method}

\item{cor_use}{passed to cor(use = cor_use)}

\item{colors}{output of heat_colors() to set heatmap fill color scheme}

\item{numbers}{output of heat_numbers() to draw numbers on heatmap cells}

\item{taxa_which}{row or column (which represents a taxon)}

\item{seriation_method}{method to order the rows (in seriation::seriate)}

\item{seriation_dist}{distance to use in seriation_method (if needed)}

\item{seriation_method_col}{method to order the columns (in seriation::seriate)}

\item{seriation_dist_col}{distance to use in seriation_method_col (if needed)}

\item{tax_transform}{transformation applied to otu_table before correlating}

\item{...}{passed to grid::gpar() for grid.text}

\item{palette}{named palette from colorspace::hcl_palettes(type = "diverging") or a vector of 3 colour names (low, middle, high)}

\item{breaks}{numeric vector of values for colour scale breaks (low, middle, high)}

\item{rev_palette}{reverse the colorspace palette}

\item{what}{what numbers to print (currently only "values" is supported)}

\item{fmt}{number print format}

\item{fontsize}{fontsize specification,}
}
\value{
ComplexHeatmap heatmap
}
\description{
Correlate microbial abundances to numeric sample_data variables

Pass a named diverging colorspace hcl palette to circlize::colorRamp2



taxAnnotation is used for cor_heatmap, comp_heatmap & tax_model_heatmap (will be)
}
\examples{
library(dplyr)
data("dietswap", package = "microbiome")
# create a couple of numerical variables to use
psq <- dietswap \%>\%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0),
    african = if_else(nationality == "AFR", true = 1, false = 0)
  )
psq <- tax_filter(ps = psq, min_prevalence = 1 / 10, min_sample_abundance = 1 / 10)
psq <- tax_agg(psq, "Genus") \%>\% ps_get()


taxa <- sample(microbiome::top_taxa(psq)[1:50], size = 30)

p <- cor_heatmap(psq, taxa, tax_anno = taxAnnotation(psq, taxa, undetected = 50))
p

p2 <- cor_heatmap(psq, taxa, tax_anno = taxAnnotation(psq, taxa, undetected = 50, which = "column"), taxa_which = "column")
p2
}
