% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrgation.R
\name{integrate_expression}
\alias{integrate_expression}
\title{Collapse values of duplicated features by a function.}
\usage{
integrate_expression(
  data,
  annotation,
  fun = colMeans,
  transpose = TRUE,
  trans_fun = identity,
  ...
)
}
\arguments{
\item{data}{a numeric matrix with features in rows and samples in columns.}

\item{annotation}{a data frames with assignment of the features in data
(first column) to gene, protein or other symbols/identifiers (second column).}

\item{fun}{a function to handle column of duplicated features,
\code{colMean} by default.}

\item{transpose}{logical, should the output be returned as a data frame with
observation in rows and features in columns?}

\item{trans_fun}{a function to transform the output numeric variables,
defaults to \code{identity}.}

\item{...}{extra arguments passed to \code{\link{aggRows}} and \code{fun}.}
}
\value{
If \code{transpose = FALSE}, a numeric matrix is returned with row names
corresponding to the second column of the \code{annotation} data frame, and
columns corresponding to the samples.
If \code{transpose = TRUE}, a data frame is returned with samples in the rows and
features in columns and named after the content of the second column of
\code{annotation}, sample identifiers are stored in the \code{sample_id} column of
the output.
}
\description{
The function collapses duplicated features such as genes or proteins by a
user-provided function that computes returns a vector of numeric values
for a matrix or data frame columns (such as \code{\link[base]{colMeans}} or
\code{\link{colMedians}}).
}
\details{
The function takes a matrix or a data frames with defined row and column
names.
Features such as genes of proteins are placed in rows, samples or
observations are provided in columns. Note: row names should contain unique
feature identifiers such as microarray probe IDs.
The assignment of feature IDs to gene symbols or other names is defined by
a data frame with two columns:
\itemize{
\item a columns with the feature identifiers corresponding to the row names of
the matrix.
\item a column with gene, protein or other names.
}

This data frame is provided as the \code{annotation} argument.
}
