% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microViz.R
\docType{package}
\name{microViz}
\alias{microViz-package}
\alias{microViz}
\title{microViz: microbiome data analysis and visualization}
\description{
microViz provides functions for statistics and visualization of microbiome sequencing data.
microViz wraps, extends and complements popular microbial ecology packages like phyloseq, vegan, and microbiome.

Check out the website for tutorials and illustrated help pages.

\url{https://david-barnett.github.io/microViz/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://david-barnett.github.io/microViz}
  \item \url{https://github.com/david-barnett/microViz}
}

}
\author{
David Barnett
(\href{https://orcid.org/0000-0003-1961-7206}{ORCID})
(\href{https://github.com/david-barnett}{GitHub})
}
