% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{minimum_shift}
\alias{minimum_shift}
\alias{minimum_shift.default}
\alias{minimum_shift.matrix}
\alias{minimum_shift.data.frame}
\title{Offset numeric vectors be their minimum.}
\usage{
minimum_shift(x, ...)

\method{minimum_shift}{default}(x, ...)

\method{minimum_shift}{matrix}(x, ...)

\method{minimum_shift}{data.frame}(x, ...)
}
\arguments{
\item{x}{an object: numeric vector, matrix or data frame.}

\item{...}{exra arguments passed to the methods.}
}
\value{
a numeric vector, matrix or data frame, depending on the input.
}
\description{
Offsets a numeric vectors or columns of a numeric matrix or a data frame by
the respective minimal values.
This results in non-negative values with minimum corresponding to 0.
Any \code{NA}s are removed silently for minimum calculation but kept in the output.
}
