% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{minMax}
\alias{minMax}
\alias{minMax.default}
\alias{minMax.matrix}
\alias{minMax.data.frame}
\title{Minimum - maximum scaling.}
\usage{
minMax(x, ...)

\method{minMax}{default}(x, ...)

\method{minMax}{matrix}(x, feature_type = c("columns", "rows"), ...)

\method{minMax}{data.frame}(x, variables = NULL, ...)
}
\arguments{
\item{x}{a numeric vector, numeric matrix or a data frame.}

\item{...}{extra arguments passed to methods.}

\item{feature_type}{specifies if columns (default) or rows are subjected to
scaling.}

\item{variables}{a vector of names of numeric variables to be normalized.
If \code{NULL}, all variables of a data frame are scaled. Please note, that
variable selection makes the computation slower.}
}
\value{
a numeric vector, numeric matrix or a data frame. Row and column
names are preserved.
}
\description{
Computes minimum - maximum scaled values of a numeric vectors, columns or
rows of a numeric matrix, or columns of a data frame.
}
\details{
The scaled values of a vector \code{x} are computed with the following formula:

\deqn{Sc(x_i) = \frac{x_i - min(x)}{max(i) - min(i)}}
}
