% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\name{new_psExtraInfo}
\alias{new_psExtraInfo}
\title{Constructor for psExtraInfo list objects (internal use)}
\usage{
new_psExtraInfo(
  tax_agg = character(),
  tax_trans = character(),
  tax_scale = character(),
  dist_method = character(),
  ord_info = new_psExtraOrdInfo()
)
}
\arguments{
\item{tax_agg}{character string naming rank at which psExtra was aggregated}

\item{tax_trans}{character vector naming transformation(s) applied to taxa in psExtra}

\item{tax_scale}{character vector}

\item{dist_method}{character vector}

\item{ord_info}{psExtraOrdInfo list}
}
\value{
psExtraInfo S3 class list
}
\description{
Constructor for psExtraInfo list objects (internal use)
}
\keyword{internal}
