% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\name{new_psExtraOrdInfo}
\alias{new_psExtraOrdInfo}
\title{Constructor for info list class used in ord_info element of psExtraInfo}
\usage{
new_psExtraOrdInfo(
  method = character(),
  constraints = character(),
  conditions = character()
)
}
\arguments{
\item{method}{character string naming ordination method}

\item{constraints}{character vector naming any constraint variables}

\item{conditions}{character vector naming any condition variables}
}
\value{
psExtraOrdInfo S3 class list
}
\description{
Constructor for info list class used in ord_info element of psExtraInfo
}
\keyword{internal}
