% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{opposite_side}
\alias{opposite_side}
\title{Simple heatmap internal helper to get opposite side}
\usage{
opposite_side(side = c("top", "right", "bottom", "left"))
}
\arguments{
\item{side}{one of "top", "right", "bottom", or "left"}
}
\value{
character
}
\description{
Simple heatmap internal helper to get opposite side
}
