% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_arrows.R
\name{Ordination-arrows}
\alias{Ordination-arrows}
\alias{vec_constraint}
\alias{vec_tax_sel}
\alias{vec_tax_all}
\title{Create ordination plot vector styling lists}
\usage{
vec_constraint(
  linewidth = 1,
  alpha = 0.8,
  colour = "brown",
  arrow = grid::arrow(length = grid::unit(0.005, units = "npc"), type = "closed", angle =
    30),
  lineend = "round",
  linejoin = "mitre",
  ...
)

vec_tax_sel(
  linewidth = 0.5,
  alpha = 1,
  colour = "black",
  arrow = grid::arrow(length = grid::unit(0.005, units = "npc"), type = "closed", angle =
    30),
  lineend = "round",
  linejoin = "mitre",
  ...
)

vec_tax_all(linewidth = 0.5, alpha = 0.25, arrow = NULL, ...)
}
\arguments{
\item{linewidth}{width of vector}

\item{alpha}{opacity of vector}

\item{colour}{colour of vector}

\item{arrow}{arrow style specified with grid::arrow() or NULL for no arrow}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{...}{further arguments passed to geom_segment}
}
\value{
list
}
\description{
Used by ord_plot, see examples there.
}
