% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_explore.R
\name{ord_explore}
\alias{ord_explore}
\title{Interactively explore compositions of ordinated samples}
\usage{
ord_explore(
  ord,
  seriate_method = "OLO_ward",
  tax_transform_for_ordering = "identity",
  ...
)
}
\arguments{
\item{ord}{list output of ord_calc}

\item{seriate_method}{seriation method to order phyloseq samples by similarity}

\item{tax_transform_for_ordering}{transform tax before ordering with ps_seriate}

\item{...}{additional arguments passed to ord_plot}
}
\value{
nothing, opens html viewer
}
\description{
Once running: click and drag to draw box over 2 or more samples to view their compositions.
You can style the ordination plot points using the options on the left panel.
}
\examples{
library(phyloseq)
library(dplyr)
data("enterotype")

ps <- enterotype
ord1 <- ps \%>\%
  dist_calc("bray") \%>\%
  ord_calc("PCoA")

# ord_explore(ord = ord1)
}
