% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_explore.R
\name{ord_explore_palet_fun}
\alias{ord_explore_palet_fun}
\title{Create fixed palette for ord_explore: tax_name = colour}
\usage{
ord_explore_palet_fun(ps, tax_level)
}
\arguments{
\item{ps}{phyloseq object}

\item{tax_level}{tax_level at which to create fixed palette}
}
\value{
named vector of colours
}
\description{
Create fixed palette for ord_explore: tax_name = colour
}
