% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_sort_ord.R
\name{tax_sort_ord}
\alias{tax_sort_ord}
\alias{ord_order_taxa}
\title{Order taxa in phyloseq by their loading vectors}
\usage{
tax_sort_ord(ps, ord, axes = 1:2, scaling = 2)

ord_order_taxa(ord, axes = 1:2, scaling = 2)
}
\arguments{
\item{ps}{phyloseq object to be sorted}

\item{ord}{psExtra with ordination object}

\item{axes}{which axes to use for sorting? numerical vector of length 1 or 2}

\item{scaling}{Type 2, or type 1 scaling. For more info,
see \url{https://sites.google.com/site/mb3gustame/constrained-analyses/rda}.
Either "species" or "site" scores are scaled by (proportional) eigenvalues,
and the other set of scores is left unscaled (from ?vegan::scores.cca)}
}
\description{
\code{tax_sort_ord} reorders taxa in a phyloseq object based on the relative
length of their taxa scores / "loading" vector lengths on 1 or 2 ordination axes.

\code{ord_order_taxa} gets the taxa names in order from the ordination
contained in a psExtra object. This is used internally by \code{tax_sort_ord}.
}
\seealso{
\itemize{
\item These functions were created to support ordering of taxa bars on \code{ord_plot_iris}
\item \code{ps_sort_ord} for ordering samples in phyloseq by ordination
}
}
