% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova}
\alias{permanova}
\title{Calculate PERMANOVA after dist_calc}
\usage{
permanova(
  data,
  variables = NULL,
  interactions = NULL,
  complete_cases = TRUE,
  n_processes = 1,
  n_perms = 999,
  seed = NULL,
  by = "margin",
  verbose = TRUE,
  return = "all",
  ...
)
}
\arguments{
\item{data}{list output from dist_calc}

\item{variables}{character vector of variables to include in model or character representation of the right-hand side of a formula, e.g "varA + varB + varA:varB"}

\item{interactions}{optional argument to define any interactions between variables, written in the style of e.g. "var_a * var_b"}

\item{complete_cases}{if TRUE, drops observations if they contain missing values (otherwise stops if missings are detected)}

\item{n_processes}{how many parallel processes to use? (on windows this uses parallel::makePSOCKcluster)}

\item{n_perms}{how many permutations? e.g. 9999 less is faster but more is better!}

\item{seed}{set a random number generator seed to ensure you get the same results each run}

\item{by}{passed to adonis2 \code{by} argument: what type of sums of squares to calculate? "margin" or "terms"}

\item{verbose}{sends messages about progress if TRUE}

\item{return}{what parts of return list to return, defaults to all parts}

\item{...}{additional arguments are passed directly to adonis2 (e.g. strata, add, sqrt.dist etc.)}
}
\value{
list containing permanova results and input objects
}
\description{
This function is a wrapper around vegan's adonis2 function. See ?vegan::adonis2 for more insight.
Test for the statistical significance of (independent) associations between variables in your phyloseq sample_data,
and a microbiota distance matrix you have already calculated with dist_calc.
The variables argument will be collapsed into one string (if length > 1) by pasting together, separated by "+".
Any interaction terms described in the interactions argument will be pasted onto the end of the pasted variables argument.
Alternatively, you can supply the complete right hand side of the formula yourself e.g variables = "varA + varB + varC\emph{varD"
Watch out, if any of your variable names contain characters that would normally separate variables in a formula then
you should rename the offending variable (e.g. avoid any of "+" "}" "|" or ":" ) otherwise permanova will split that variable into pieces.
}
\examples{
library(phyloseq)
library(vegan)
data("dietswap", package = "microbiome")

# add some missings to demonstrate automated removal
sample_data(dietswap)$sex[3:6] <- NA

# compute distance
testDist <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  dist_calc("bray")

PERM <- testDist \%>\%
  permanova(
    seed = 1,
    variables = c("sex", "bmi_group"),
    n_processes = 1,
    n_perms = 99 # only 99 perms used in examples for speed (use 999, 9999 or more!)
  )
str(PERM, max.level = 1)

# try permanova with interaction terms
PERM2 <- testDist \%>\%
  permanova(
    seed = 1,
    variables = "nationality + sex * bmi_group",
    n_processes = 1, n_perms = 99
  )
PERM2$permanova

# specify the same model in alternative way
PERM3 <- testDist \%>\%
  permanova(
    seed = 1,
    variables = c("nationality", "sex", "bmi_group"),
    interactions = "sex * bmi_group",
    n_processes = 1, n_perms = 99
  )
PERM3$permanova

identical(PERM3, PERM2) # TRUE

# take the same distance matrix used for the permanova and plot an ordination
PERM2 \%>\%
  ord_calc("PCoA") \%>\%
  ord_plot(color = "bmi_group")
# this trick ensures any samples dropped from the permanova for having missing values
# in the covariates are NOT included in the corresponding ordination plot

}
