% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_plotting.R
\name{plot_bionco}
\alias{plot_bionco}
\title{Oncoplots for binary data.}
\usage{
plot_bionco(
  data,
  variables = NULL,
  split_fct = NULL,
  variable_classification = NULL,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "observation",
  y_lab = "variable",
  fill_lab = "status",
  hide_x_axis_text = FALSE,
  y_text_face = "plain",
  cust_theme = microViz::theme_micro(),
  color_scale = c("white", "steelblue"),
  color_labels = c("WT", "altered"),
  x_rug_scale = c("number", "percentage"),
  y_rug_scale = c("number", "percentage"),
  x_rug_fill = "gray50",
  x_rug_color = x_rug_fill,
  y_rug_color = "white",
  y_rug_fill = color_scale[2],
  one_plot = TRUE,
  rel_heights = c(0.75, 0.25),
  rel_widths = c(0.75, 0.25),
  legend_position = c("inside", "right", "bottom", "top", "left"),
  plot_margins = ggplot2::margin(t = -35, r = -35, b = -10, l = 0, unit = "pt"),
  title_panel_ratio = c(0.1, 0.9),
  legend_panel_ratio = c(0.1, 0.9)
)
}
\arguments{
\item{data}{a data frame with binary 0/1 numeric data or two-level factors,
where 0 codes for absence and 1 codes for presence of the event of interest.}

\item{variables}{names of variables in \code{data} that store the binary event
indexes. If \code{NULL}, all variables will be used.}

\item{split_fct}{an optional name of a variable used for classification of
observation and column faceting of the plots.}

\item{variable_classification}{a two-variable data frame or a
\code{\link[clustTools]{clust_analysis}} object that defines categories of
the variables used for row faceting of the plots.}

\item{plot_title}{plot title}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{fill_lab}{title of the fill legend.}

\item{hide_x_axis_text}{logical, should the X axis text be hidden?}

\item{y_text_face}{font face of the Y axis text. An italic font face may be
useful e.g. for gene symbols.}

\item{cust_theme}{custom \code{ggplot} theme applied to the plots.}

\item{color_scale}{a two-element vector storing the colors for the
absence and presence of the event, respectively.}

\item{color_labels}{a two-element character vector with labels for the
event absence and presence, respectively.}

\item{x_rug_scale}{statistic of frequency of the event per observation,
presented in the X axis rug bar plot.}

\item{y_rug_scale}{statistic of frequency of the event per variable,
presented in the Y axis rug bar plot.}

\item{x_rug_fill}{fill color for bars of the X axis rug plot.}

\item{x_rug_color}{color for the bars of the X axis rug plot.}

\item{y_rug_color}{color for the bars of the Y axis rug plot.}

\item{y_rug_fill}{fill color for bars of the Y axis rug plot.}

\item{one_plot}{logical, should the oncoplot and the rug plots be stitched
into one panel?}

\item{rel_heights}{relative heights of the output plot panel. The defaults
are expected to work fine in most cases.}

\item{rel_widths}{relative widths of the output plot panel. The defaults
are expected to work fine in most cases.}

\item{legend_position}{placement of the oncoplot fill color legend in the
plot panel. By default it is placed in the right bottom corner (
\code{legend_position = 'inside'}).}

\item{plot_margins}{margins of the rug plots, which may be used for fine
tuning of their position in the output panel. The defaults are expected to
work fine in most cases.}

\item{title_panel_ratio}{ratio of the title/subtitle banner width to the
width of the output panel.}

\item{legend_panel_ratio}{ratio of the legend size to the size of the output
panel. Ignored if \code{legend_position = 'inside'}.}
}
\value{
a \code{ggplot} graphic object or, if \code{one_plot = FALSE}, a list of three
\code{ggplot} objects with the main plot and X and Y axis rug bar plots.
}
\description{
Generates oncoplots also known as waterfall plots for binary data.
}
\details{
The function takes a data frame that stores binary, i.e. 0/1 numeric data or
two-level factors, as the first argument and generates a two-color heat map
with observations in the X axis and features in the Y axis. The variables are
ordered by frequency of the event. The observations are arranged by
the frequency of event in the first variable with the highest frequency of
the  event, followed by the second one, till the n-th variable with the
lowest event frequency.

The user may facet the plot by providing a name of the splitting variable
(\code{split_fct}, faceting of the columns) and a two-column data frame
as \code{variable_classification} (faceting of the rows). In the
\code{variable_classification} data frame, the first column stores variable names
and the second one codes for the variable subset assignment. Alternatively,
the user may provide a \code{\link[clustTools]{clust_analysis}} object with
results of a clustering analysis as the \code{variable_classification} argument.

By default, i.e. when \code{one_plot = TRUE}, the function returns a single panel
of three plots: the oncoplot, an Y axis rug bar plot with overall
frequencies of the event per variable, and an X axis rug bar plot with
overall frequencies of the event per observation. Those frequencies may be
represented as absolute numbers or percentages as defined by \code{x_rug_scale}
and \code{y_rug_scale} arguments. If \code{one_plot = FALSE}, the function returns
a list of thoe three plots.
}
