% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_plotting.R
\name{plot_bistack}
\alias{plot_bistack}
\title{Stack plots for binary data.}
\usage{
plot_bistack(
  data,
  variables = NULL,
  split_fct = NULL,
  variable_classification = NULL,
  scale = c("number", "percentage"),
  reverse_status = FALSE,
  show_frequencies = FALSE,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "events, # observations",
  y_lab = "variable",
  fill_lab = "status",
  cust_theme = theme_micro(),
  color_scale = c("white", "steelblue"),
  color_labels = c("WT", "altered"),
  bar_rim_color = "gray60",
  signif_digits = 2,
  text_size = 2.75,
  text_color = "black",
  facet_scales = c("free_y", "free_x", "free"),
  facet_space = c("free_y", "free_x", "free"),
  ...
)
}
\arguments{
\item{data}{a data frame with binary 0/1 numeric data or two-level factors,
where 0 codes for absence and 1 codes for presence of the event of interest.}

\item{variables}{names of variables in \code{data} that store the binary event
indexes. If \code{NULL}, all variables will be used.}

\item{split_fct}{an optional name of a variable used for classification of
observation and column faceting of the plots.}

\item{variable_classification}{a two-variable data frame or a
\code{\link[clustTools]{clust_analysis}} object that defines categories of
the variables used for row faceting of the plots.}

\item{scale}{statistic of frequency to be presented in the plot: absolute
number (default) or frequency.}

\item{reverse_status}{logical, should the levels of the status be reversed?
By default, i.e. when \code{reverse_status = FALSE}, the event coded as 1 is
plotted first.}

\item{show_frequencies}{logical, should labels with the frequencies be
displayed in the plot?}

\item{plot_title}{plot title}

\item{plot_subtitle}{plot subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{fill_lab}{title of the fill legend.}

\item{cust_theme}{custom \code{ggplot} theme applied to the plots.}

\item{color_scale}{a two-element vector storing the colors for the
absence and presence of the event, respectively.}

\item{color_labels}{a two-element character vector with labels for the
event absence and presence, respectively.}

\item{bar_rim_color}{color of rims of the bars.}

\item{signif_digits}{number of significant digits used for rounding of the
text labels. Ignored if \code{show_frequencies = FALSE} or \code{scale = 'number'}.}

\item{text_size}{size of the frequency labels.}

\item{text_color}{color of the text in the frequency labels.}

\item{facet_scales}{a string that indicates which scales can vary between
the plot facets. By default, \code{facet_scales = 'free_y'}, which means that
only the Y axis scale will differ. See \code{\link[ggplot2]{facet_grid}} for
details.}

\item{facet_space}{a string that indicates which spaces can vary between
the plot facets. By default, \code{facet_space = 'free_y'}, which means that
only the Y axis scale spaces will differ. See
\code{\link[ggplot2]{facet_grid}} for details.}

\item{...}{extra arguments that control appearance of the frequency labels
passed to \code{\link[ggplot2]{geom_label}}.}
}
\description{
The function draws a classical stack plot with counts or percentages of the
event in binary data.
}
\details{
The function takes a data frame that stores binary, i.e. 0/1 numeric data or
two-level factors, as the first argument and generates a two-color stack
plot. The variables are ordered by frequency of the event.

The user may facet the plot by providing a name of the splitting variable
(\code{split_fct}, faceting of the columns) and a two-column data frame
as \code{variable_classification} (faceting of the rows). In the
\code{variable_classification} data frame, the first column stores variable names
and the second one codes for the variable subset assignment. Alternatively,
the user may provide a \code{\link[clustTools]{clust_analysis}} object with
results of a clustering analysis as the \code{variable_classification} argument.
}
