% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_bubble}
\alias{plot_bubble}
\title{Draw a bubble plot.}
\usage{
plot_bubble(
  data,
  x_variable,
  y_variable,
  size_variable,
  color_variable = size_variable,
  label_variable = NULL,
  point_color = "steelblue",
  color_scale = c(low = "steelblue", mid = "white", high = "firebrick"),
  midpoint = NULL,
  size_type = c("size", "area"),
  size_range = c(0.2, 4),
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = x_variable,
  y_lab = y_variable,
  size_lab = NULL,
  color_lab = NULL,
  cust_theme = theme_micro(),
  txt_hjust = -1.4,
  txt_vjust = 0.5,
  txt_size = 2.75,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{x_variable}{variable to be presented in the X axis of the plot.}

\item{y_variable}{variable to be presented in the Y axis of the plot.}

\item{size_variable}{name of the variable defining the bubble size.}

\item{color_variable}{name of the variable defining the bubble color.}

\item{label_variable}{variable to be presented as text next to the data
points.}

\item{point_color}{color of the points, relevant only if color variable is
not provided.}

\item{color_scale}{a character vector of length 3, which defines the lower,
middle and upper point color of the fill scale.}

\item{midpoint}{the middle point of the color scale, optional.}

\item{size_type}{type of size scaling: \code{size} makes bubble radius correspond
to the \code{size_variable}, \code{area} makes bubble area relates to the
\code{size_variable}.}

\item{size_range}{a numeric vector of size 2, which defines the minimum and
maximum of the size scale.}

\item{plot_title}{plot_title.}

\item{plot_subtitle}{plot_subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{size_lab}{title of the size scale.}

\item{color_lab}{title of the color scale.}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{txt_hjust}{horizontal justification of the text.}

\item{txt_vjust}{vertical justification of the text.}

\item{txt_size}{text size.}

\item{...}{extra arguments passed to the size and color scales, e.g.
\code{\link[ggplot2]{scale_fill_gradient2}}.}
}
\value{
a \code{ggplot} object.
}
\description{
A general form of a bubble plot: variables are presented in the Y axis,
splitting factor levels are shown in the x axis, size and fill of the bubbles
code for additional variables and, as an option, can be labeled with a text.
}
