% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plotting.R
\name{plot_common_effect}
\alias{plot_common_effect}
\title{Counts of significant effects and average effect sizes in multi-data set
analysis.}
\usage{
plot_common_effect(
  data,
  label_variable,
  regulation_variable,
  regulation_status,
  ns_level = "ns",
  average_fun = colMeans,
  cutoff = 0,
  palette = c(upregulated = "firebrick", activated = "firebrick", positive = "firebrick",
    downregulated = "steelblue", inhibited = "steelblue", negative = "steelblue",
    regulated = "plum4", significant = "plum4", ns = "gray70"),
  show_whiskers = TRUE,
  whisker_size = 0.75,
  probs = c(0.025, 0.975),
  point_size = 2,
  point_alpha = 0.75,
  point_wjitter = 0,
  point_hjitter = 0,
  show_txt = TRUE,
  txt_size = 2.75,
  txt_color = NULL,
  txt_face = "plain",
  labeller = identity,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = regulation_variable,
  y_lab = NULL,
  status_lab = regulation_status,
  plot_tag = NULL,
  cust_theme = theme_micro()
)
}
\arguments{
\item{data}{a list of data frames with analysis results.}

\item{label_variable}{name of the variable storing feature names.}

\item{regulation_variable}{name of the variable storing the estimate to
be plotted.}

\item{regulation_status}{name of the variable storing regulation status index
(e.g. activated/inhibited/ns or regulated/ns) in factor form. This variable
levels will be represented by color of the data points.}

\item{ns_level}{level of the \code{regulation_status} variable that represent
non-significant regulation. \code{ns} by default.}

\item{average_fun}{function used for calculation of the average
regulation estimates over the data sets.}

\item{cutoff}{cutoff value of the count of data sets with significant
effects. Useful for identification of common significant effects shared by
multiple data sets.}

\item{palette}{colors coding for the regulation status. Most common
options are provided as a default.}

\item{show_whiskers}{logical, should whiskers representing a quantile range
be displayed in the plot? Defaults to false.}

\item{whisker_size}{size of the percentile whiskers.}

\item{probs}{a numeric vector of two quantiles defining the range for the
whiskers. Ignored if \code{show_whiskers = FALSE}.}

\item{point_size}{size of the data points.}

\item{point_alpha}{alpha of the data points.}

\item{point_wjitter}{data point jittering width.}

\item{point_hjitter}{data point jittering height.}

\item{show_txt}{logical, should labels with names of the data sets be
displayed?}

\item{txt_size}{size of the label text.}

\item{txt_color}{color of the text. If \code{NULL}, the color is specified by
levels of the splitting factor.}

\item{txt_face}{font face of the text labels.}

\item{labeller}{a function transforming the data set names to the text labels
in the plot. Defaults to \code{identity}, i.e. the names are displayed as the are.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{status_lab}{fill and color scale title.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{custom \code{ggplot} theme.}
}
\value{
a \code{ggplot} object.
}
\description{
Draws a scatter plot of average effect sizes (or regulation; X axis) and
counts of data sets with significant regulation (Y axis).
}
