% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_plotting.R
\name{plot_common_estimates}
\alias{plot_common_estimates}
\title{Scatter/Forest plot of estimates in a multi-data set analysis.}
\usage{
plot_common_estimates(
  data,
  label_variable,
  regulation_variable,
  regulation_status,
  palette = c(upregulated = "firebrick", activated = "firebrick", positive = "firebrick",
    downregulated = "steelblue", inhibited = "steelblue", negative = "steelblue",
    regulated = "plum4", significant = "plum4", ns = "gray80"),
  show_average = TRUE,
  average_fun = colMeans,
  show_whiskers = TRUE,
  whisker_size = 0.75,
  probs = c(0.025, 0.975),
  point_size = 2,
  point_alpha = 0.75,
  point_wjitter = 0,
  point_hjitter = 0.1,
  average_shape = 23,
  average_size = 3,
  stat_color = "orangered3",
  stat_rim = stat_color,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = regulation_variable,
  y_lab = NULL,
  status_lab = regulation_status,
  plot_tag = NULL,
  cust_theme = theme_micro()
)
}
\arguments{
\item{data}{a list of data frames with analysis results.}

\item{label_variable}{name of the variable storing feature names.}

\item{regulation_variable}{name of the variable storing the estimate to
be plotted.}

\item{regulation_status}{name of the variable storing regulation status index
(e.g. activated/inhibited/ns or regulated/ns) in factor form. This variable
levels will be represented by color of the data points.}

\item{palette}{colors coding for the regulation status. Most common
options are provided as a default.}

\item{show_average}{logical, should averages over all data srts be displayed?}

\item{average_fun}{function used for calculation of the average
regulation estimates over the data sets.}

\item{show_whiskers}{logical, should whiskers representing a quantile range
be displayed in the plot? Defaults to false.}

\item{whisker_size}{size of the percentile whiskers.}

\item{probs}{a numeric vector of two quantiles defining the range for the
whiskers. Ignored if \code{show_whiskers = FALSE}.}

\item{point_size}{size of the data points.}

\item{point_alpha}{alpha of the data points.}

\item{point_wjitter}{data point jittering width.}

\item{point_hjitter}{data point jittering height.}

\item{average_shape}{code of shape of the average symbol.}

\item{average_size}{size of the symbol representing the central
tendency measure.}

\item{stat_color}{color of the symbols of the central tendency measure and
the error bars.}

\item{stat_rim}{color of the rim of the average statistic symbol.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle.}

\item{x_lab}{X axis title.}

\item{y_lab}{Y axis title.}

\item{status_lab}{fill and color scale title.}

\item{plot_tag}{plot tag.}

\item{cust_theme}{custom \code{ggplot} theme.}
}
\value{
a \code{ggplot} object.
}
\description{
Draws a scatter plot of estimates specified by the user. Point color codes
for significance and regulation status. As an option, a central statistic
over all estimates such as mean or median with a quantile range may be
displayed.
}
