% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_bar.R
\name{plot_comp_bar}
\alias{plot_comp_bar}
\title{Pretty & flexible barplots of sample composition}
\usage{
plot_comp_bar(
  ps,
  n_taxa = 10,
  tax_level = "Genus",
  tax_order = "abundance",
  palette = c("lightgrey", rev(distinct_palette(n_taxa))),
  sample_order = "aitchison",
  label = "SAMPLE",
  groups = NA,
  horizontal = TRUE,
  bar_width = 1,
  bar_outline_colour = "black",
  drop_unused_vars = TRUE,
  seriate_method = "OLO_ward"
)
}
\arguments{
\item{ps}{phyloseq object}

\item{n_taxa}{how many distinct taxa to colour in the plot (otherwise "other")}

\item{tax_level}{taxonomic aggregation level (from rank_names(ps))}

\item{tax_order}{or what? taxonomy grouping?}

\item{palette}{palette for taxa fill colours}

\item{sample_order}{any distance measure in calc_dist that does not require a phylogenetic tree (or a variable?)}

\item{label}{could also consider arbitrary annotation with extra info, like in complex heatmap}

\item{groups}{splits dataset by this variable (must be categorical) and uses patchwork to assemble multiple plots?}

\item{horizontal}{typically you have more samples than taxa and more vertical space than horizontal, so TRUE makes sense as default}

\item{bar_width}{default 1 avoids random gapping otherwise seen with many samples (set to something less than 1 to introduce gaps between fewer samples)}

\item{bar_outline_colour}{line colour separating taxa and samples}

\item{drop_unused_vars}{speeds up ps_melt but might limit future plot customisation options}

\item{seriate_method}{name of any ordering method suitable for distance matrices (see ?seriation::seriate)}
}
\value{
ggplot or list of harmonised ggplots
}
\description{
No gaps, unless you want them. Grouped bars and bar outlines if you want them.
Consistent, distinct, fairly pleasant colouring.
}
\examples{
library(microbiome)
data(dietswap)

# make a list of 2 harmonised composition plots (grouped by sex)
p <- plot_comp_bar(dietswap, n_taxa = 15, bar_outline_colour = "black",
                   sample_order = 'aitchison', groups = 'sex')

# plot them side by side with patchwork package
patch <- patchwork::wrap_plots(p, ncol = 2, guides = 'collect')
patch

# beautifying tweak #
# modify one plot in place (flip the order of the samples in the 2nd plot)
# notice that the scaling is for the x-axis
# (that's because coord_flip is used inside plot_comp_bar)
patch[[2]] <- patch[[2]] + scale_x_discrete(limits = rev)
# Explainer: rev() function takes current limits and reverses them.
# You could also pass a completely arbitrary order, naming all samples

# you can theme all plots with the & operator
patch & theme(axis.text.y = element_text(size = 5), legend.text = element_text(size = 6))
# See https://patchwork.data-imaginist.com/index.html


}
