% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_density}
\alias{plot_density}
\title{Draw regulation statistics for selected genes.}
\usage{
plot_density(
  data,
  regulation_variable,
  label_variable,
  group_variable = NULL,
  fill_color = "steelblue",
  fill_scale = NULL,
  fill_labels = waiver(),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Regulation",
  y_lab = "Feature density",
  cust_theme = theme_micro(),
  shape_alpha = 1,
  show_points = FALSE,
  point_size = 2,
  point_alpha = 1,
  point_hjitter = 0.1,
  point_wjitter = 0,
  show_txt = FALSE,
  top_regulated = 5,
  txt_size = 2.5,
  txt_color = NULL,
  txt_face = "plain",
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{regulation_variable}{name of the variable storing the regulation/effect
size data.}

\item{label_variable}{variable storing the gene/object names.}

\item{group_variable}{a variable which defines groups, in which the estimate
density will be plotted. If provided, the plot is faceted by levels of this
variable. If \code{group_variable = NULL}, the density in the whole data set is
plotted.}

\item{fill_color}{fill color for the density shapes. Ignored if
\code{group_variable} is provided.}

\item{fill_scale}{colors for the fill scale corresponding to
levels of \code{group_variable}.}

\item{fill_labels}{labels for the fill scale.}

\item{fill_title}{title of the point fill legend.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text.}

\item{x_lab}{x axis title.}

\item{y_lab}{Y axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{shape_alpha}{alpha/opacity of the density shapes.}

\item{show_points}{logical, should data points be displayed in the plot?}

\item{point_size}{size of the data points.}

\item{point_alpha}{alpha/opacity of the data points.}

\item{point_hjitter}{height of jittering of the data points.}

\item{point_wjitter}{width of jittering of the data points.}

\item{show_txt}{logical, should data points for features with the most
extreme regulation be labeled?}

\item{top_regulated}{top n regulated features to be presented in the plot.}

\item{txt_size}{size of the text.}

\item{txt_color}{color of the text. If \code{NULL}, the text color corresponds to
levels of the grouping variable.}

\item{txt_face}{font face of the text.}

\item{...}{additional arguments passed to \code{\link[ggplot2]{geom_density}}.}
}
\value{
a \code{ggplot} object.
}
\description{
Function \code{plot_density()} draws density of regulation estimates:
in the whole data set or in groups defined by a variable.
}
