% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_forest}
\alias{plot_forest}
\alias{plot_top}
\title{Draw regulation statistics for selected genes.}
\usage{
plot_forest(
  data,
  regulation_variable,
  label_variable,
  p_variable,
  signif_level = 0.05,
  regulation_level = 0,
  lower_ci_variable = NULL,
  upper_ci_variable = NULL,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Regulation",
  cust_theme = microViz::theme_micro(),
  signif_digits = 2,
  show_txt = FALSE,
  show_ci_txt = FALSE,
  txt_size = 2.75,
  txt_hjust = 0.5,
  txt_vjust = -0.8
)

plot_top(
  data,
  regulation_variable,
  label_variable,
  p_variable,
  signif_level = 0.05,
  regulation_level = 0,
  lower_ci_variable = NULL,
  upper_ci_variable = NULL,
  top_regulated = 10,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Regulation",
  cust_theme = microViz::theme_micro(),
  signif_digits = 2,
  show_txt = FALSE,
  show_ci_txt = FALSE,
  txt_size = 2.75,
  txt_hjust = 0.5,
  txt_vjust = -0.8
)
}
\arguments{
\item{data}{a data frame.}

\item{regulation_variable}{name of the variable storing the regulation/effect
size data.}

\item{label_variable}{variable storing the gene/object names.}

\item{p_variable}{variable storing the p values.}

\item{signif_level}{significance threshold, p = 0.05 be default.}

\item{regulation_level}{the regulation cutoff for identification of the
significant effects, 0 difference versus control by default.}

\item{lower_ci_variable}{variable storing the lower CI values.}

\item{upper_ci_variable}{variable storing the upper CI values.}

\item{fill_scale}{regulation colors, a vector of three elements: for
upregulated, downregulated and non-significant items, respectively.}

\item{fill_title}{title of the point fill legend.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text.}

\item{x_lab}{x axis title.}

\item{cust_theme}{custom ggplot theme.}

\item{signif_digits}{significant digits used for rounding of the
displayed numeric variables.}

\item{show_txt}{logical, label the points with the estimate value?}

\item{show_ci_txt}{logical, should the CI be included in the text label?}

\item{txt_size}{text size.}

\item{txt_hjust}{horizontal justification of the text label.}

\item{txt_vjust}{vertical justification of the text label.}

\item{top_regulated}{top n regulated genes/objects to be presented in the plot.}
}
\value{
a ggplot object.
}
\description{
Draws a Forest plot with the regulation estimates and,
optionally confidence intervals for all items presented in a data frame
(\code{plot_forest}) or top up- and downregulated fratures (\code{plot_top}).
}
