% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordin8.R
\name{plot_ordin8}
\alias{plot_ordin8}
\title{Customisable ggplot of ordin8 ordination result}
\usage{
plot_ordin8(
  data,
  axes = 1:2,
  constraint_label_length = 1,
  constraint_label_style = list(size = 2.5, colour = "#D92027", alpha = 0.8),
  plot_taxa = FALSE,
  taxon_label_length = 1,
  taxon_label_style = list(size = 2.5, colour = "black", alpha = 0.8),
  taxon_renamer = function(x) identity(x),
  plot_samples = TRUE,
  auto_title = TRUE,
  ...
)
}
\arguments{
\item{data}{list object output from ordin8}

\item{axes}{which axes to plot: numerical vector of length 2}

\item{constraint_label_length}{relative length of line segment drawn for any constraints (relative to default length which is defined by correlation with each drawn axis)}

\item{constraint_label_style}{list of fixed aesthetics (colour, size etc) for the constraint labels themselves}

\item{plot_taxa}{if ordin8 method was "RDA" draw the taxa loading vectors (see details)}

\item{taxon_label_length}{relative length of line segment drawn for any constraints}

\item{taxon_label_style}{list of fixed aesthetics (colour, size etc) for the taxon labels}

\item{plot_samples}{if TRUE, plot sample points with geom_point}

\item{auto_title}{if TRUE, add a crude title and subtitle with info about the ordination}

\item{...}{pass aesthetics arguments for sample points, drawn with geom_point using aes_string}
}
\value{
ggplot
}
\description{
Ordination visualisation. Utilises results of \code{ordin8}.
Some inspiration was from here ref: \url{https://ourcodingclub.github.io/2018/05/04/ordination.html}
}
\details{
How to specify the plot_taxa argument:
\itemize{
\item integer vector e.g. 1:3 --> plot labels for top 3 taxa (by line length)
\item single numeric value e.g. 0.75 --> plot labels for taxa with line length > 0.75
\item character vector e.g. c('g__Bacteroides', 'g__Veillonella') --> plot labels for the exactly named taxa
}
}
\examples{
library(phyloseq)
library(vegan)
library(microbiome)
data("dietswap", package = "microbiome")

# create a couple of numerical variables to use as constraints or conditions
sample_data(dietswap)$female <-
  dplyr::if_else(sample_data(dietswap)$sex == "female", true = 1, false = 0)
sample_data(dietswap)$weight <-
  dplyr::recode(sample_data(dietswap)$bmi_group, obese = 3, overweight = 2, lean = 1)

# compute and plot ordinations for demonstration of conditioning
unconstrained_aitchison_pca <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ordin8(method = "RDA")

unconstrained_aitchison_pca \%>\%
  plot_ordin8(colour = "bmi_group") +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ordin8(method = "RDA", conditions = "weight") \%>\%
  plot_ordin8(colour = "bmi_group") +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

constrained_aitchison_rda <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ordin8(method = "RDA", constraints = c("weight", "female"))

constrained_aitchison_rda \%>\%
  plot_ordin8(colour = "bmi_group", constraint_label_length = 2) +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

# ggplot allows additional customisation of the resulting plot
p <- constrained_aitchison_rda \%>\%
  plot_ordin8(
    colour = "bmi_group", constraint_label_length = 2,
    plot_taxa = 1:3, taxon_label_length = 5
  ) +
  lims(x = c(-5, 6), y = c(-5, 5)) +
  scale_colour_brewer(palette = "Set1")

p + stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))
p + stat_density2d(aes(colour = bmi_group))

# it is possible to facet only unconstrained ordination plots (with plot_taxa = FALSE)
unconstrained_aitchison_pca \%>\%
  plot_ordin8(color = "sex") +
  facet_wrap("sex") +
  stat_density2d(aes(colour = sex)) +
  guides(colour = FALSE)

unconstrained_aitchison_pca \%>\%
  plot_ordin8(color = "bmi_group", plot_samples = FALSE) +
  facet_wrap("sex") +
  stat_density2d_filled(show.legend = FALSE)
}
