% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sign}
\alias{plot_sign}
\title{Regulation bar plot.}
\usage{
plot_sign(
  data,
  regulation_variable,
  p_variable,
  signif_level = 0.05,
  regulation_level = 1.5,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  fill_title = "",
  bar_alpha = 0.75,
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Gene",
  y_lab = "Regulation",
  show_trend = TRUE,
  cust_theme = microViz::theme_micro(),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{regulation_variable}{name of the variable storing the regulation/effect
size data.}

\item{p_variable}{name of the variable storing p values.}

\item{signif_level}{significance threshold, p = 0.05 be default.}

\item{regulation_level}{the regulation cutoff for identification of the
significant effects, 1.5 difference versus control by default.}

\item{fill_scale}{color palette for coding significance/regulation.
Three colors required: for upregulation, downregulation and non-significance.}

\item{fill_title}{fill legend title.}

\item{bar_alpha}{plot bar alpha.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text. If NULL, the numbers of significantly
regulated genes is displayed.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{show_trend}{logical, should a LOESS trend for
regulation be displayed in the plot?}

\item{cust_theme}{custom ggplot theme.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_smooth}}.}
}
\value{
a ggplot object.
}
\description{
Draws a bar plot with the regulation/effect size values.
Bar color codes for the regulation sign and significance.
}
