% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_signifcant}
\alias{plot_signifcant}
\title{Plot top p values as a bar plot.}
\usage{
plot_signifcant(
  data,
  p_variable,
  label_variable,
  signif_level = 0.05,
  top_significant = 10,
  fill_scale = c(significant = "coral3", ns = "gray60"),
  x_lab = expression("-log"[10] * " p"),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = microViz::theme_micro()
)
}
\arguments{
\item{data}{a data frame.}

\item{p_variable}{variable storing the p values.}

\item{label_variable}{variable storing the gene/object names.}

\item{signif_level}{significance threshold, p = 0.05 be default.}

\item{top_significant}{top n significant genes/objects to be presented in
the plot.}

\item{fill_scale}{regulation colors, a vector of two elements: for
significant and non-significant items.}

\item{x_lab}{x axis title.}

\item{fill_title}{title of the point fill legend.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text.}

\item{cust_theme}{custom ggplot theme.}
}
\value{
a ggplot object.
}
\description{
Plots top n p values as a bar plot.
}
