% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_significant}
\alias{plot_significant}
\alias{plot_regulated}
\title{Plot top p values or regulation estimates as bar plots.}
\usage{
plot_significant(
  data,
  p_variable,
  label_variable,
  signif_level = 0.05,
  top_significant = 10,
  fill_scale = c(significant = "coral3", ns = "gray60"),
  x_lab = expression("-log"[10] * " p"),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = theme_micro()
)

plot_regulated(
  data,
  regulation_variable,
  label_variable,
  p_variable,
  signif_level = 0.05,
  regulation_level = 0,
  top_regulated = 10,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  x_lab = "Regulation",
  cust_theme = theme_micro(),
  show_txt = FALSE,
  signif_digits = 2,
  txt_color = "black",
  txt_size = 2.75,
  txt_offset = 0.05,
  txt_vjust = 0.5
)
}
\arguments{
\item{data}{a data frame.}

\item{p_variable}{variable storing the p values.}

\item{label_variable}{variable storing the gene/object names.}

\item{signif_level}{significance threshold, p = 0.05 be default.}

\item{top_significant}{top n significant genes/objects to be presented in
the plot.}

\item{fill_scale}{significance or regulation colors, a vector of two or
three elements.}

\item{x_lab}{x axis title.}

\item{fill_title}{title of the point fill legend.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text.}

\item{cust_theme}{custom ggplot theme.}

\item{regulation_variable}{variable storing the regulation estimates.}

\item{regulation_level}{regulation cutoff.}

\item{top_regulated}{top n strongest regulated genes/objects to be presented
in the plot.}

\item{show_txt}{logical, should values of the regulation estimates be
presented in the plot?}

\item{signif_digits}{significant digits for regulation estimates presented
in the plot.}

\item{txt_color}{text color.}

\item{txt_size}{text size.}

\item{txt_offset}{horizontal offsetting of the text relative to the
regulation variable values.}

\item{txt_vjust}{vertical justification of the text.}
}
\value{
a \code{ggplot} object.
}
\description{
Plots top n p values or regulation estimates as a bar plot.
}
