% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_two_euler}
\alias{plot_two_euler}
\title{Two-set Euler/Venn plot.}
\usage{
plot_two_euler(
  data,
  show_quantities = TRUE,
  show_labels = TRUE,
  fill_scale = c("orangered3", "aquamarine3"),
  fill_title = "Set",
  alpha = 0.5,
  n_prefix = "n = ",
  txt_size = 2.75,
  txt_color = "black",
  txt_face = "plain",
  plot_title = NULL,
  plot_subtitle = NULL,
  cust_theme = theme_micro(),
  ...
)
}
\arguments{
\item{data}{a named list with two vectors; duplicated elements in the vectors
are silently removed.}

\item{show_quantities}{logical, should N numbers of the intersecting and
non-intersecting parts be shown in the plots?}

\item{show_labels}{logical, should names of the sets be displayed in the plot?}

\item{fill_scale}{a two element character vector with colors for the sets.}

\item{fill_title}{fill legend title.}

\item{alpha}{alpha of the circles/ellipses.}

\item{n_prefix}{text to be displayed before the N numbers. Omitted if
\code{show_quantities = FALSE} or set to \code{NULL}.}

\item{txt_size}{size of the text.}

\item{txt_color}{text color.}

\item{txt_face}{face of the text.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle. #'}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{...}{additional arguments passed to \code{\link[eulerr]{euler}}.}
}
\value{
An Euler plot in \code{ggplot} format.
}
\description{
Creates a two-set Euler plot, i.e. a Venn plot, where areas of the
intersecting and non-intersecting elements correspond to sizes of the
intersecting and non-intersecting sets.
}
\details{
The functions provides a handy tool for simple Euler plots in \code{ggplot}
format.
Internally, it employs \code{\link[eulerr]{euler}} function to calculate
parameters of circles/ellipses.
The idea comes from https://gist.github.com/danlooo/ - many thanks!
}
