% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Draw a volcano plot.}
\usage{
plot_volcano(
  data,
  regulation_variable,
  p_variable,
  signif_level = 0.05,
  regulation_level = 1.5,
  fill_scale = c(upregulated = "firebrick", downregulated = "steelblue", ns = "gray60"),
  x_lab = regulation_variable,
  y_lab = expression("-log"[10] * " p"),
  top_significant = 0,
  top_regulated = 0,
  label_variable = NULL,
  label_type = c("label", "text"),
  txt_size = 2.75,
  txt_color = "black",
  txt_face = 1,
  fill_title = "",
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_tag = NULL,
  cust_theme = microViz::theme_micro(),
  point_alpha = 0.8,
  point_hjitter = 0,
  point_wjitter = 0,
  show_hline = TRUE,
  show_vlines = c("both", "right", "left")
)
}
\arguments{
\item{data}{a data frame.}

\item{regulation_variable}{name of the variable storing the regulation/effect
size data.}

\item{p_variable}{name of the variable storing p values.}

\item{signif_level}{significance threshold, p = 0.05 be default.}

\item{regulation_level}{the regulation cutoff for identification of the
significant effects, 1.5 difference versus control by default.}

\item{fill_scale}{color palette for coding significance/regulation.
Three colors required: for upregulation, downregulation and non-significance.}

\item{x_lab}{x axis title.}

\item{y_lab}{y axis title.}

\item{top_significant}{top n significant genes/objects to be labeled in
the plot with their names.}

\item{top_regulated}{top n most regulated genes/objects to be labeled in
the plot with their names. Ignored if \code{top_significant} is provided.}

\item{label_variable}{variable storing the gene/object names.}

\item{label_type}{type of the gene/object label: 'label' (default)
or plain 'text'.}

\item{txt_size}{size of the gene/object name text.}

\item{txt_color}{gene/object name text color.}

\item{txt_face}{face of the gene/object name text.}

\item{fill_title}{title of the point fill legend.}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot subtitle.}

\item{plot_tag}{plot tag text. If NULL, the numbers of significantly
regulated genes is displayed.}

\item{cust_theme}{custom ggplot theme.}

\item{point_alpha}{plot point alpha.}

\item{point_hjitter}{height jittering of the points, defaults to 0.}

\item{point_wjitter}{width jittering of the points, defaults to 0.}

\item{show_hline}{logical, should the horizontal line representing the
significance cutoff be presented in the plot?}

\item{show_vlines}{specifies which vertical lines representing the regulation
cutoffs will be plotted.}
}
\value{
a \code{ggplot} object.
}
\description{
Generates a Volcano plot with the effect size/regulation
statistic value on the X axis and -log10 p value on the y axis. Regulation
and significance is coded by the point fill color. Top n significant
genes/objects may be labeled with their names.
}
