% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_wordcloud}
\alias{plot_wordcloud}
\title{Plot a word cloud.}
\usage{
plot_wordcloud(
  data,
  label_variable,
  split_fct = NULL,
  size_variable = NULL,
  color_variable = NULL,
  size_type = c("size", "area"),
  size_range = c(2, 6),
  color_scale = c("steelblue", "black", "firebrick"),
  midpoint = NULL,
  size_lab = size_variable,
  color_lab = color_variable,
  cust_theme = ggplot2::theme_void(),
  wrap = FALSE,
  len = 3,
  fraction_rotated = 0,
  nrow = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{label_variable}{name of the variable storing the text to be plotted.}

\item{split_fct}{optional, name of a variable that defines dataset subsets.
Each subset will be presented as a separate facet of the plot.}

\item{size_variable}{optional, name of a variable numeric whose value will be
size coded.}

\item{color_variable}{optional, name of a variable numeric whose value will be
color coded.}

\item{size_type}{type of size scaling: \code{size} makes font size correspond to
the \code{size_variable}, \code{area} makes font text related to the \code{size_variable}.}

\item{size_range}{a numeric vector of size 2, which defines the minimum and
maximum of the size scale.}

\item{color_scale}{a character vector of length 3, which defines the lower,
middle and upper point color of the color scale.}

\item{midpoint}{optional, the middle point of the color scale.}

\item{size_lab}{title of the size scale.}

\item{color_lab}{title of the color scale.}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{wrap}{logical, should each text be split into multiple lines? May
be useful in case of long labels, e.g. descriptions of biological processes.}

\item{len}{number of words per line. Relevant only if \code{wrap_text = TRUE}.}

\item{fraction_rotated}{fraction of words to be rotated with a 90 degree
angle. May improve visibility.}

\item{nrow}{number of rows in a faceted plot. Ignored if \code{split_fct = NULL}.}

\item{...}{extra arguments passed to
\code{\link[ggwordcloud]{geom_text_wordcloud}} or
\code{\link[ggwordcloud]{geom_text_wordcloud_area}}}
}
\value{
a \code{ggplot} object.
}
\description{
Plots a word cloud e.g. of gene or GO term names. Font color and size may
code for additional parameters such as p value or fold regulation.
}
\details{
The function internally employs
\code{\link[ggwordcloud]{geom_text_wordcloud}} and
\code{\link[ggwordcloud]{geom_text_wordcloud_area}} from the \code{wordcloud}
package.
}
