% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap-utils.R
\name{prev}
\alias{prev}
\title{Calculate prevalence from numeric vector}
\usage{
prev(x, undetected = 0)
}
\arguments{
\item{x}{numeric vector (of taxon counts or proportions)}

\item{undetected}{value above which a taxon is considered present or detected}
}
\value{
numeric value
}
\description{
Useful as helper for taxon prevalence calculation
}
\examples{
prev(c(0, 0, 1, 2, 4))
prev(c(0, 0, 1, 2, 4), undetected = 1.5)
}
