% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\name{print.psExtraInfo}
\alias{print.psExtraInfo}
\title{Print method for psExtraInfo object}
\usage{
\method{print}{psExtraInfo}(
  psExtraInfo,
  which = c("tax_agg", "tax_trans", "tax_scale", "dist_method", "ord_info")
)
}
\arguments{
\item{psExtraInfo}{psExtraInfo object}

\item{which}{which elements of psExtraInfo list to print}
}
\description{
Print method for psExtraInfo object
}
