% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\name{print.psExtraOrdInfo}
\alias{print.psExtraOrdInfo}
\title{Print method for psExtraOrdInfo list}
\usage{
\method{print}{psExtraOrdInfo}(psExtraOrdInfo)
}
\description{
Print method for psExtraOrdInfo list
}
