% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_calc_diversity.R
\name{ps_calc_diversity}
\alias{ps_calc_diversity}
\title{Calculate diversity index and add to phyloseq sample data}
\usage{
ps_calc_diversity(
  ps,
  rank,
  index = "shannon",
  exp = FALSE,
  varname = paste0(ifelse(exp, "exp_", ""), paste0(index, "_", rank))
)
}
\arguments{
\item{ps}{phyloseq}

\item{rank}{taxonomic rank name, or "unique"}

\item{index}{name of diversity index from microbiome::diversity().
One of: "inverse_simpson", "gini_simpson", "shannon", "fisher", "coverage"}

\item{exp}{exponentiate the result? (i.e. return e^index) - see details}

\item{varname}{name of the variable to be added to phyloseq sample data}
}
\value{
phyloseq
}
\description{
Wrapper around microbiome::diversity() function.
Takes and returns a phyloseq object. Calculates an alpha diversity index at
a given taxonomic rank. Returns phyloseq unaggregated, with an additional
variable. Variable name is by default created by pasting the index and rank.
}
\details{
Don't filter taxa before calculating alpha diversity.

See the following resources for a discussion of exponentiated diversity indices
http://www.loujost.com/Statistics\%20and\%20Physics/Diversity\%20and\%20Similarity/Diversity\%20of\%20a\%20single\%20community.htm
http://www.loujost.com/Statistics\%20and\%20Physics/Diversity\%20and\%20Similarity/EffectiveNumberOfSpecies.htm
}
\examples{
data(ibd, package = "microViz")
ibd \%>\%
  ps_filter(abx == "abx") \%>\%
  tax_fix() \%>\%
  ps_calc_diversity("Genus", index = "shannon", exp = TRUE) \%>\%
  ps_calc_diversity("Family", index = "inverse_simpson") \%>\%
  tax_transform(rank = "Genus", transform = "clr") \%>\%
  ord_calc("PCA") \%>\%
  ord_plot(
    colour = "exp_shannon_Genus", size = "inverse_simpson_Family"
  ) +
  ggplot2::scale_colour_viridis_c()

}
