% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_calc_dominant.R
\name{ps_calc_dominant}
\alias{ps_calc_dominant}
\title{Calculate dominant taxon in each phyloseq sample}
\usage{
ps_calc_dominant(
  ps,
  rank = "unique",
  threshold = 0.3,
  n_max = 6,
  var = paste("dominant", rank, sep = "_")
)
}
\arguments{
\item{ps}{phyloseq object}

\item{rank}{taxonomic rank to calculate}

\item{threshold}{minimum proportion at which to consider a sample dominated by a taxon}

\item{n_max}{maximum number of taxa that can be listed as dominant taxa}

\item{var}{name of variable to add to phyloseq object sample data}
}
\value{
phyloseq object
}
\description{
Which taxon is most abundant in each sample of your phyloseq object?
This function adds this information as a new variable in your phyloseq
sample_data.
\itemize{
\item If the most abundant taxon is below the proportional abundance threshold,
the dominant taxon will be "none" for that sample
\item If there are more than n_max dominant taxa across all samples (not
including "none") the dominant taxon will be "other" for those samples
}
}
\details{
Thanks to Vitor Heidrich for the initial idea and draft implementation
}
\examples{
library(ggplot2)
ps <- corncob::ibd_phylo \%>\%
  tax_filter(min_prevalence = 3) \%>\%
  tax_fix() \%>\%
  phyloseq_validate()

ps \%>\%
  ps_filter(DiseaseState == "CD") \%>\%
  ps_calc_dominant(rank = "Genus") \%>\%
  comp_barplot(tax_level = "Genus", label = "dominant_Genus", n_taxa = 12) +
  coord_flip()
}
