% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_calc_richness.R
\name{ps_calc_richness}
\alias{ps_calc_richness}
\title{Calculate richness estimate and add to phyloseq sample data}
\usage{
ps_calc_richness(
  ps,
  rank,
  index = "observed",
  detection = 0,
  varname = paste0(index, "_", rank)
)
}
\arguments{
\item{ps}{phyloseq}

\item{rank}{taxonomic rank name, or "unique"}

\item{index}{"observed" or "chao1" - name of richness estimate from microbiome::richness()}

\item{detection}{Detection threshold. Used for the "observed" index.}

\item{varname}{name of the variable to be added to phyloseq sample data}
}
\value{
phyloseq
}
\description{
Wrapper around microbiome::richness() function.
Takes and returns a phyloseq object. Calculates a richness estimate at
a given taxonomic rank. Returns phyloseq unaggregated, with an additional
variable. Variable name is by default created by pasting the index and rank.
}
\details{
Don't filter taxa before calculating richness.

These richness indices are estimates.
For a discussion of the uncertainty and bias of these estimates
see e.g. work by Amy Willis https://doi.org/10.3389/fmicb.2019.02407
}
\examples{
data(ibd_phylo, package = "corncob")
ibd_phylo \%>\%
  ps_filter(abx == "abx") \%>\%
  tax_fix() \%>\%
  ps_calc_richness("Genus", index = "observed") \%>\%
  ps_calc_richness("Family", index = "chao1") \%>\%
  tax_transform(rank = "Genus", transform = "clr") \%>\%
  ord_calc("PCA") \%>\%
  ord_plot(
    colour = "observed_Genus", size = "chao1_Family"
  ) +
  ggplot2::scale_colour_viridis_c()

}
