% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-accessors.R
\name{ps_counts}
\alias{ps_counts}
\title{Get phyloseq with counts if available}
\usage{
ps_counts(data, warn = TRUE)
}
\arguments{
\item{data}{A phyloseq or psExtra object.}

\item{warn}{A boolean or "error" string to control warning or error behaviour (default: TRUE).}
}
\value{
A phyloseq object with counts if available.
}
\description{
Get phyloseq with counts if available
}
\keyword{internal}
