% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_drop_incomplete.R
\name{ps_drop_incomplete}
\alias{ps_drop_incomplete}
\title{Deselect phyloseq samples with sample_data missings}
\usage{
ps_drop_incomplete(ps, vars = NA, verbose = FALSE)
}
\arguments{
\item{ps}{phyloseq with sample_data}

\item{vars}{vector of variable names to check for missings (or NA, which uses all variables in sample data)}

\item{verbose}{message about number of samples dropped if not FALSE, and number per variable in vars if "max"}
}
\value{
phyloseq
}
\description{
Wrapper for stats::complete.cases function.
}
\examples{
library(phyloseq)
data("enterotype", package = "phyloseq")

enterotype
ps_drop_incomplete(enterotype)
ps_drop_incomplete(enterotype, vars = "Enterotype", verbose = TRUE)
ps_drop_incomplete(enterotype, vars = "Sample_ID", verbose = TRUE)
ps_drop_incomplete(enterotype, vars = c("Enterotype", "Sample_ID"))
ps_drop_incomplete(enterotype, verbose = "max")
}
