% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_extra-accessors.R
\name{ps_extra-accessors}
\alias{ps_extra-accessors}
\alias{ps_get}
\alias{dist_get}
\alias{ord_get}
\alias{info_get}
\alias{perm_get}
\alias{bdisp_get}
\title{Extract elements from ps_extra class}
\usage{
ps_get(ps_extra)

dist_get(ps_extra)

ord_get(ps_extra)

info_get(ps_extra)

perm_get(ps_extra)

bdisp_get(ps_extra)
}
\arguments{
\item{ps_extra}{ps_extra class object}
}
\value{
element of ps_extra class object
}
\description{
Extract elements from ps_extra class
}
\examples{
library(phyloseq)
data("esophagus")

}
