% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_filter.R
\name{ps_filter}
\alias{ps_filter}
\title{Filter phyloseq samples by sample_data variables}
\usage{
ps_filter(ps, ..., .target = "sample_data", .keep_all_taxa = FALSE)
}
\arguments{
\item{ps}{phyloseq object}

\item{...}{passed directly to dplyr::filter (see examples and ?dplyr::filter)}

\item{.target}{which slot of phyloseq to mutate, currently only "sample_data" supported}

\item{.keep_all_taxa}{if FALSE (the default) taxa which are no longer present in the dataset after filtering are removed}
}
\value{
phyloseq object (with filtered sample_data)
}
\description{
Wrapper for dplyr::filter function. All dplyr functionality
}
\examples{
library(phyloseq)
library(microbiome)
library(dplyr)

data("enterotype")
enterotype
sample_data(enterotype)[1:10, 1:5]

ps1 <- ps_filter(enterotype, SeqTech != "Sanger")
ps1
sample_data(ps1)[1:10, 1:5]

ps2 <- enterotype \%>\% ps_filter(across(everything(), ~ !is.na(.)))
ps2
sample_data(ps2)[1:8, 1:8]

# function will give warning if some of the otu_values are negative
# unless you set .keep_all_taxa = TRUE
enterotype \%>\%
  microbiome::transform("clr") \%>\%
  ps_filter(SeqTech == "Sanger", .keep_all_taxa = TRUE)
}
