% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_otu2samdat.R
\name{ps_otu2samdat}
\alias{ps_otu2samdat}
\title{Copy phyloseq otu_table data to sample_data}
\usage{
ps_otu2samdat(ps, taxa = NULL)
}
\arguments{
\item{ps}{phyloseq with sample_data}

\item{taxa}{list of taxa_names to copy to sample_data, or NULL (which selects all with \code{phyloseq::taxa_names()})}
}
\value{
phyloseq with augmented sample_data
}
\description{
Copy phyloseq otu_table data to sample_data
}
\examples{
library(dplyr)
library(phyloseq)
library(microbiome)
data("dietswap")

ps <- dietswap \%>\% ps_otu2samdat("Akkermansia")
sample_variables(ps)

# or if you do not specify any taxa, all are copied
ps_all <- dietswap \%>\% ps_otu2samdat()
sample_variables(ps_all)[1:15]

# this could be useful for colouring ordination plots, for example
ps \%>\%
  ps_mutate(log_akkermansia = log(Akkermansia)) \%>\%
  calc_dist("bray") \%>\%
  ordin8("PCoA") \%>\%
  plot_ordin8(
    colour = "log_akkermansia",
    size = 3, shape = "nationality"
  )
}
