% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_reorder.R
\name{ps_reorder}
\alias{ps_reorder}
\title{Set order of samples in phyloseq object}
\usage{
ps_reorder(ps, sample_order)
}
\arguments{
\item{ps}{phyloseq}

\item{sample_order}{names or current numerical indices of samples in desired order}
}
\value{
phyloseq
}
\description{
Manually set order of samples by specifying samples names in desired order.
}
\details{
Ordering of samples in a phyloseq is controlled from the otu_table slot!
}
\examples{
library(phyloseq)
data("dietswap", package = "microbiome")

dietswap \%>\%
  sample_data() \%>\%
  head(8)

new_order <- rev(sample_names(dietswap))
dietswap \%>\%
  ps_reorder(new_order) \%>\%
  sample_data() \%>\%
  head(8)

# random ordering with numbers
set.seed(1000)
random_order <- sample(1:nsamples(dietswap))
dietswap \%>\%
  ps_reorder(random_order) \%>\%
  sample_data() \%>\%
  head(8)
}
\seealso{
\code{\link{ps_arrange}} for arranging samples by sample_data variables (or otu_table)

\code{\link{ps_seriate}} for arranging samples by microbiome similarity

\code{\link{ps_filter}} for keeping only some samples, based on sample_data
}
