% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_select.R
\name{ps_select}
\alias{ps_select}
\title{Select phyloseq sample_data using dplyr::select syntax}
\usage{
ps_select(ps, ...)
}
\arguments{
\item{ps}{phyloseq with sample_data}

\item{...}{passed straight to dplyr::select}
}
\value{
phyloseq object
}
\description{
Simple selection of phyloseq sample_data variables, might be useful for printing reduced sample_data, or inside other functions
}
\examples{
library(phyloseq)
library(dplyr)
data("enterotype", package = "phyloseq")

head(sample_data(enterotype))

enterotype \%>\%
  ps_select(!contains("Sample")) \%>\%
  sample_data() \%>\%
  head()
}
