% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_seriate.R
\name{ps_seriate}
\alias{ps_seriate}
\title{Arrange samples in a phyloseq by microbiome similarity}
\usage{
ps_seriate(
  ps,
  method = "OLO_ward",
  dist = "bray",
  tax_transform = "identity",
  add_variable = FALSE,
  rank = NA
)
}
\arguments{
\item{ps}{phyloseq object}

\item{method}{seriation method for ordering samples, from seriation::seriate}

\item{dist}{distance method for dist_calc (only used if required for particular seriation method!)}

\item{tax_transform}{transformation to apply before seriation or any distance calculation}

\item{add_variable}{add a variable to the sample data indicating seriation order}

\item{rank}{taxonomic rank to aggregate at, before seriation, NA for no aggregation}
}
\value{
phyloseq
}
\description{
Uses seriation methods from seriation::seriate and often dist_calc (depending on if seriation method requires a distance matrix)
}
\examples{
library(phyloseq)
data("dietswap", package = "microbiome")

dietswap \%>\%
  sample_data() \%>\%
  head(8)

dietswap \%>\%
  tax_agg("Genus") \%>\%
  .$ps \%>\%
  ps_seriate(method = "OLO_ward", dist = "bray") \%>\%
  sample_data() \%>\%
  head(8)
}
\seealso{
\code{\link{ps_arrange}} \code{\link{ps_reorder}}
}
