% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-accessors.R
\name{psExtra-accessors}
\alias{psExtra-accessors}
\alias{ps_get}
\alias{dist_get}
\alias{ord_get}
\alias{info_get}
\alias{perm_get}
\alias{bdisp_get}
\alias{tax_models_get}
\alias{tax_stats_get}
\alias{taxatree_models_get}
\alias{taxatree_stats_get}
\alias{otu_get}
\alias{tt_get}
\alias{samdat_tbl}
\title{Extract elements from ps_extra class}
\usage{
ps_get(psExtra, ps_extra)

dist_get(psExtra, ps_extra)

ord_get(psExtra, ps_extra)

info_get(psExtra, ps_extra)

perm_get(psExtra, ps_extra)

bdisp_get(psExtra, ps_extra)

tax_models_get(psExtra)

tax_stats_get(psExtra)

taxatree_models_get(psExtra)

taxatree_stats_get(psExtra)

otu_get(data, taxa = NA, samples = NA, counts = FALSE)

tt_get(data)

samdat_tbl(data, sample_names_col = ".sample_name")
}
\arguments{
\item{psExtra}{psExtra S4 class object}

\item{ps_extra}{deprecated! don't use this}

\item{data}{phyloseq or psExtra}

\item{taxa}{subset of taxa to return, NA for all (default)}

\item{samples}{subset of samples to return, NA for all (default)}

\item{counts}{should otu_get ensure it returns counts? if present in object}

\item{sample_names_col}{name of column where sample_names are put.
if NA, return data.frame with rownames (sample_names)}
}
\value{
element(s) from psExtra object (or NULL)
}
\description{
\itemize{
\item \code{ps_get}         returns phyloseq
\item \code{info_get}       returns psExtraInfo object
\item \code{dist_get}       returns distance matrix (or NULL)
\item \code{ord_get}        returns ordination object (or NULL)
\item \code{perm_get}       returns adonis2() permanova model (or NULL)
\item \code{bdisp_get}      returns results of betadisper() (or NULL)
\item \code{otu_get}        returns phyloseq otu_table matrix with taxa as columns
\item \code{tt_get}         returns phyloseq tax_table
\item \code{tax_models_get} returns list generated by tax_model or NULL
\item \code{tax_stats_get}  returns dataframe generated by tax_models2stats or NULL
\item \code{taxatree_models_get} returns list generated by taxatree_models or NULL
\item \code{taxatree_stats_get} returns dataframe generated by taxatree_models2stats or NULL
\item \code{samdat_tbl}     returns phyloseq sample_data as a tibble
with sample_names as new first column called .sample_name
}
}
\examples{
data("dietswap", package = "microbiome")
psx <- tax_transform(dietswap, "identity", rank = "Genus")
psx

ps_get(psx)
info_get(psx)

dist_get(psx) # this psExtra has no dist_calc result
ord_get(psx) # this psExtra has no ord_calc result
perm_get(psx) # this psExtra has no dist_permanova result
bdisp_get(psx) # this psExtra has no dist_bdisp result

# these can be returned from phyloseq objects too
otu_get(psx, taxa = 6:9, samples = c("Sample-9", "Sample-1", "Sample-6"))
tt_get(psx) \%>\% head()
samdat_tbl(psx)
}
