% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\docType{class}
\name{psExtra-class}
\alias{psExtra-class}
\title{Define psExtra class S4 object}
\description{
Define psExtra class S4 object
}
\section{Slots}{

\describe{
\item{\code{info}}{list.}

\item{\code{counts}}{otu_table.}

\item{\code{dist}}{dist.}

\item{\code{ord}}{ANY.}

\item{\code{permanova}}{ANY.}

\item{\code{bdisp}}{ANY.}

\item{\code{taxatree_models}}{list.}

\item{\code{taxatree_stats}}{data.frame.}

\item{\code{tax_models}}{list.}

\item{\code{tax_stats}}{data.frame.}
}}

\examples{
library(phyloseq)
data("shao19")

ps <- shao19 \%>\% ps_filter(infant_age == 12)
ps \%>\% tax_agg("genus")
}
