% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-class.R
\name{print.psExtraInfo}
\alias{print.psExtraInfo}
\title{Print method for psExtraInfo object}
\usage{
\method{print}{psExtraInfo}(
  x,
  ...,
  which = c("tax_agg", "tax_trans", "tax_scale", "dist_method", "ord_info")
)
}
\arguments{
\item{x}{psExtraInfo object}

\item{...}{ignored}

\item{which}{which elements of psExtraInfo list to print}
}
\description{
Print method for psExtraInfo object
}
