% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{re_adjust}
\alias{re_adjust}
\title{Adjust for multiple testing.}
\usage{
re_adjust(data, p_variable = "p_value", method = "BH", simplify_p = TRUE)
}
\arguments{
\item{data}{a data frame with the testing results.}

\item{p_variable}{name of the variable that stores raw p values to be
adjusted.}

\item{method}{adjustment method.}

\item{simplify_p}{logical, should adjusted p values below 0.001 be presented
in a p < 0.001 form?}
}
\value{
The genuine data frame appended with two columns:
\itemize{
\item \code{p_adjusted} with adjusted p values
\item \code{significance} with text indicating significance and p value, e.g.
'ns(p = 0.12).
}
}
\description{
Adjusts for multiple comparison. The input is a result of a statistic test
in a data frame format.
}
\details{
For available methods, please consult \code{\link[stats]{p.adjust}}.
}
