% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regulation_colors}
\alias{regulation_colors}
\title{Colors for regulation status.}
\usage{
regulation_colors()
}
\value{
a named character vector with colors for up- and downregulated, and
non-significant features used e.g. by Volcano and Forest plots.
}
\description{
Default colors representing regulation and significance status.
}
