% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_functions.R
\name{rowMedians}
\alias{rowMedians}
\alias{rowMedians.matrix}
\alias{rowMedians.data.frame}
\alias{rowMins}
\alias{rowMins.matrix}
\alias{rowMins.data.frame}
\alias{rowMax}
\alias{rowMax.matrix}
\alias{rowMax.data.frame}
\alias{rowGmeans}
\alias{rowGmeans.matrix}
\alias{rowGmeans.data.frame}
\alias{rowHmeans}
\alias{rowHmeans.matrix}
\alias{rowHmeans.data.frame}
\alias{rowVars}
\alias{rowVars.matrix}
\alias{rowVars.data.frame}
\alias{rowSDs}
\alias{rowSDs.matrix}
\alias{rowSDs.data.frame}
\alias{rowGini}
\alias{rowGini.matrix}
\alias{rowGini.data.frame}
\alias{rowFreqRatios}
\alias{rowFreqRatios.matrix}
\alias{rowFreqRatios.data.frame}
\alias{rowPercUniques}
\alias{rowPercUniques.matrix}
\alias{rowPercUniques.data.frame}
\alias{rowQuantiles}
\alias{rowQuantiles.matrix}
\alias{rowQuantiles.data.frame}
\alias{rowCI}
\alias{rowCI.matrix}
\alias{rowCI.data.frame}
\alias{row_stats}
\title{Compute distribution metrics for rows of a data frame or a matrix.}
\usage{
rowMedians(x, ...)

\method{rowMedians}{matrix}(x, na.rm = TRUE, ...)

\method{rowMedians}{data.frame}(x, na.rm = TRUE, ...)

rowMins(x, ...)

\method{rowMins}{matrix}(x, na.rm = TRUE, ...)

\method{rowMins}{data.frame}(x, na.rm = TRUE, ...)

rowMax(x, ...)

\method{rowMax}{matrix}(x, na.rm = TRUE, ...)

\method{rowMax}{data.frame}(x, na.rm = TRUE, ...)

rowGmeans(x, ...)

\method{rowGmeans}{matrix}(x, na.rm = TRUE, ...)

\method{rowGmeans}{data.frame}(x, na.rm = TRUE, ...)

rowHmeans(x, ...)

\method{rowHmeans}{matrix}(x, na.rm = TRUE, ...)

\method{rowHmeans}{data.frame}(x, na.rm = TRUE, ...)

rowVars(x, ...)

\method{rowVars}{matrix}(x, na.rm = TRUE, ...)

\method{rowVars}{data.frame}(x, na.rm = TRUE, ...)

rowSDs(x, ...)

\method{rowSDs}{matrix}(x, na.rm = TRUE, ...)

\method{rowSDs}{data.frame}(x, na.rm = TRUE, ...)

rowGini(x, unbiased, ...)

\method{rowGini}{matrix}(x, unbiased = TRUE, na.rm = TRUE, ...)

\method{rowGini}{data.frame}(x, unbiased = TRUE, na.rm = TRUE, ...)

rowFreqRatios(x, ...)

\method{rowFreqRatios}{matrix}(x, ...)

\method{rowFreqRatios}{data.frame}(x, ...)

rowPercUniques(x, ...)

\method{rowPercUniques}{matrix}(x, ...)

\method{rowPercUniques}{data.frame}(x, ...)

rowQuantiles(x, probs, ...)

\method{rowQuantiles}{matrix}(x, probs, na.rm = TRUE, ...)

\method{rowQuantiles}{data.frame}(x, probs, na.rm = TRUE, ...)

rowCI(x, conf_level, ...)

\method{rowCI}{matrix}(x, conf_level = 0.95, method = c("percentile", "bca"), na.rm = TRUE, ...)

\method{rowCI}{data.frame}(x, conf_level = 0.95, method = c("percentile", "bca"), na.rm = TRUE, ...)

row_stats(x, unbiased = TRUE, freqCut = 95/5, uniqueCut = 10, ...)
}
\arguments{
\item{x}{a numeric data frame or matrix.}

\item{...}{extra arguments passed to methods, currently none.}

\item{na.rm}{logical: should NAs be removed?}

\item{unbiased}{logical, should estimated of the population Gini coefficient
be returned (multiplied by \code{n/(n - 1)})?}

\item{probs}{a numeric vector of quantiles.}

\item{conf_level}{a single numeric with the width of the confidence interval.}

\item{method}{method of confidence interval calculation: percentile (default)
or BCA.}

\item{freqCut}{the cutoff of ratio of the most frequent element to the
second most frequent element. Used for identification of near-zero
variance variables.}

\item{uniqueCut}{the cutoff of percentage of unique observations. Used for
identification of near-zero variance variables.}
}
\value{
\code{rowMedians()}, \code{rowMins()}, \code{rowMax()}, \code{rowGmeans()},
\code{rowHmeans()}, \code{rowVars()}, \code{rowSDs()}, \code{rowGini()} return numeric vectors
with, respectively, row medians, row minima, row maxima, geometric means,
harmonic means, variances, standard deviations and Gini coefficients.
\code{rowQuantiles()}, and \code{rowCI()} return numeric matrices of quantiles and
confidence intervals of rows.
\code{rowFreqRatios()} returns ratios of frequencies of the first most common
element to the second one for each row; the function silently skips any
\code{NAs}.
\code{rowPercUniques()} returns percentages of unique observations in each row;
\code{NAs} are silently removed.

\code{row_stats()} returns a data frame with the following columns:
\itemize{
\item \code{variable}: if the input object has defined row names, they will be
displayed here, otherwise observation numbers.
\item \code{median}: row medians
\item \code{mean}: row means
\item \code{var}: row variances
\item \code{gini_coef}: Gini coefficients of the rows
\item \code{var_mean_ratio}: variance to mean ratios
\item \code{freqRatio}: ratios of frequencies of the first most common element to the
second most common element.
\item \code{percentUnique}: percentage of unique observations
\item \code{zeroVar}: logical, indicates if the row has only one unique value.
\item \code{nzv}: logical, indicates if the row passes the variability criteria
defined by the \code{freqCut} and \code{uniqueCut} arguments.
}
}
\description{
The functions compute row medians, variances, standard deviations, Gini
coefficients, quantiles, confidence intervals and metrics of uniqueness.
\code{row_stats} computes a bunch of distribution stats that
are helpful at selection of variant features for the further analysis, e.g.
differential gene expression.
}
\references{
DiCiccio TJ, Efron B. Bootstrap confidence intervals.
https://doi.org/101214/ss/1032280214 (1996) 11:189–228.
doi:10.1214/SS/1032280214
}
