% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psExtra-accessors.R
\name{samdatAsDataframe}
\alias{samdatAsDataframe}
\title{Internal helper that gets phyloseq sample_data as a plain dataframe}
\usage{
samdatAsDataframe(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A dataframe with sample data from the phyloseq object.
}
\description{
Internal helper that gets phyloseq sample_data as a plain dataframe
}
\keyword{internal}
