% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shao19.R
\docType{data}
\name{shao19}
\alias{shao19}
\title{Gut microbiota relative abundance data from Shao et al. 2019}
\format{
large phyloseq object
}
\source{
Stunted microbiota and opportunistic pathogen colonization in caesarean-section birth.
\doi{10.1038/s41586-019-1560-1}
}
\usage{
shao19
}
\description{
A phyloseq dataset containing 1644 gut microbiome samples.
Fecal samples were collected from 596 infants. 175 mothers also provided fecal samples.
}
\section{Data acquisition}{

The processed (metaphlan3) relative abundance data were obtained using curatedMetagenomicData package version 3.4.2.
\url{https://waldronlab.io/curatedMetagenomicData/articles/available-studies.html}

A small amount of data cleaning was performed after download:
\enumerate{
\item to make sample metadata more user friendly
\item to fill gaps in the taxonomy table.
}
}

\keyword{datasets}
