% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{shared_features}
\alias{shared_features}
\title{Identify shared elements in sets.}
\usage{
shared_features(x, m = 2)
}
\arguments{
\item{x}{a list of character or numeric vectors.}

\item{m}{number of sets which share the common features.}
}
\value{
a vector with the features shared by the given number of
elements.
}
\description{
The function chooses elements shared by at least n elements of a list.
As such it may be useful at identifying e.g. significantly regulated genes
or pathways in several data sets.
}
