% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_vec}
\alias{split_vec}
\title{Split a vector into chunks of a defined size.}
\usage{
split_vec(x, len)
}
\arguments{
\item{x}{a vector.}

\item{len}{chunk length.}
}
\value{
a list of vector chunks.
}
\description{
Splits the vector into chuns of a defined size: the last one
can be shorter ('ceiling').
}
