% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{subset_distance}
\alias{subset_distance}
\title{Similarity and distance between data subsets.}
\usage{
subset_distance(
  data,
  variables,
  split_fct,
  normalize = TRUE,
  norm_center = c("mean", "median"),
  dist_FUN = dist,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector of names of the variables of interest.}

\item{split_fct}{name of the splitting factor, the variable defining the
data set subsets.}

\item{normalize}{logical, should the data frame variables be normalized prior
to plotting?}

\item{norm_center}{defines centering of the variable during scaling: mean
(default) or median. Ignored if \code{normalize = FALSE}.}

\item{dist_FUN}{a distance-computing function.}

\item{...}{additional arguments passed to \code{dist_FUN}.}
}
\description{
The function \code{subset_distance()} computes distances such as Euclidean,
Manhattan or cosine distance, or similarity measures between subsets of
a data set in respect to variables of interest.
}
\details{
Since there are multiple ways to calculate distances in R, the function
allows the user to work with their favorite one - provided it returns a
numeric matrix or a \code{dist} class instance. We can recommend
\code{\link[stats]{dist}} and \code{\link[clustTools]{calculate_dist}}.
The function returns a list of numeric distance matrices of
the \code{\link[clustTools]{cross_dist}} class, with multiple analytic and
visualization tools provided by the
\href{https://github.com/PiotrTymoszuk/clustTools}{clustTools} package.
}
