% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{t_tester}
\alias{t_tester}
\title{T test.}
\usage{
t_tester(data, split_fct, variable, ...)
}
\arguments{
\item{data}{a data frame.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variable}{dependent variable name.}

\item{...}{additional arguments passed to \code{\link[stats]{t.test}}.}
}
\value{
a data frame with the regulation statistic, confidence intervals,
t and p values
}
\description{
Test for the difference with a T test.
}
