% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_fill_unknowns.R
\name{tax_fill_unknowns}
\alias{tax_fill_unknowns}
\title{Replace unknown, NA, or short tax_table values}
\usage{
tax_fill_unknowns(
  x,
  min_length = 4,
  unknowns = c("unknown", paste0(c("p", "c", "o", "f", "g", "s"), "__")),
  levels = phyloseq::rank_names(x),
  suffix_rank = "classified",
  sep = " ",
  verbose = TRUE
)
}
\arguments{
\item{x}{phyloseq or tax_table (taxonomyTable) with at least two ranks}

\item{min_length}{replace strings shorter than this}

\item{unknowns}{also replace strings matching any in this vector}

\item{levels}{names of taxonomic levels to modify, defaults to all}

\item{suffix_rank}{"classified" (default) or "current", when replacing an entry, should the suffix be taken from the lowest classified rank for that taxon "classified", or the "current" unclassified rank?}

\item{sep}{character(s) separating new name and taxonomic rank level suffix (see suffix_rank)}

\item{verbose}{emit warnings when cannot replace with informative name?}
}
\value{
object same class as x
}
\description{
Identifies phyloseq tax_table values as unknown or uninformative and replaces them with the first informative value from a higher taxonomic rank.
\itemize{
\item Short values in phyloseq tax_table are typically empty strings or " ", or "g__" etc. so it is helpful to replace them. Set \code{min_length} = 0 to avoid filtering on length.
\item Values in \code{unknowns} are also removed, even if longer than \code{min_length}. It is up to the user to specify sensible values in \code{unknowns} if their dataset has other unwanted values.
\item NA values are also replaced.
}
}
\examples{
library(dplyr)
library(tibble)
library(phyloseq)
library(microbiome)

data(dietswap)
ps <- dietswap

# create unknowns to test filling
tt <- tax_table(ps)
ntax <- ntaxa(ps)
set.seed(123)
tt[sample(1:ntax, 30), 3] <- "g__"
tt[sample(1:ntax, 20), 2] <- "f__"
tt[sample(1:ntax, 10), 1] <- "p__"
tt[sample(1:ntax, 10), 3] <- "unknown"
# create a row with only NAs
tt[1, ] <- NA

tax_table(ps) <- tax_table(tt)

# tax_fill_unknowns defaults should solve most problems
ps
tax_table(ps) \%>\% head(50)

# this will replace "unknown"s as well as short values including "g__" and "f__"
tax_fill_unknowns(ps) \%>\%
  tax_table() \%>\%
  head(50)

# this will only replace values in Genus column, only replace short entries, and not "unknown"
ps \%>\%
  tax_fill_unknowns(unknowns = NULL, levels = "Genus") \%>\%
  tax_table() \%>\%
  head(50)

# Change rank suffix and separator settings
tax_fill_unknowns(ps, suffix_rank = "current", sep = " - ") \%>\%
  tax_table() \%>\%
  head(50)


# larger example tax_table shows 1000s rows still fast, from microbiomeutilities package
# library(microbiomeutilities)
# data("hmp2")
# tax_fill_unknowns(hmp2)
}
