% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_mutate.R
\name{tax_mutate}
\alias{tax_mutate}
\title{Modify or compute new taxonomic ranks in phyloseq}
\usage{
tax_mutate(ps, ...)
}
\arguments{
\item{ps}{phyloseq object with a tax_table, or just a tax_table}

\item{...}{passed straight to dplyr::mutate (see examples and dplyr::mutate help)}
}
\value{
phyloseq object with modified tax_table
}
\description{
Add or overwrite tax_table ranks.
Use dplyr::mutate() syntax.
}
\examples{
library(phyloseq)
library(dplyr)
data("dietswap", package = "microbiome")

# compute new rank
tax_mutate(dietswap, loud_genus = toupper(Genus)) \%>\%
  tt_get() \%>\%
  head()

# overwrite a current rank
tax_mutate(dietswap, Genus = toupper(Genus)) \%>\%
  tt_get() \%>\%
  head()

# overwrite all ranks
tax_mutate(dietswap, across(everything(), .fns = toupper)) \%>\%
  tt_get() \%>\%
  head()

# add a new rank at the beginning
tax_mutate(dietswap, Root = "Bacteria", .before = 1) \%>\%
  tt_get() \%>\%
  head()

# this is an error as ranks can't be any other class than character
# tax_mutate(dietswap, Genus = 1:ntaxa(dietswap))
}
\seealso{
\code{\link[dplyr]{mutate}}

\code{\link{ps_mutate}}
}
