% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_name.R
\name{tax_name}
\alias{tax_name}
\title{Set unique taxa_names for phyloseq object}
\usage{
tax_name(
  ps,
  prefix = c("tax", "asv", "otu")[1],
  rank = NA,
  pad_number = TRUE,
  sep = "_"
)
}
\arguments{
\item{ps}{phyloseq object}

\item{prefix}{e.g. 'tax', 'asv', or 'otu' (or set your own)}

\item{rank}{name of taxonomic rank from which to use classifications in new names}

\item{pad_number}{should unique numbers have zeros added to the front (e.g. 001, 002) to be made the same number of characters?}

\item{sep}{character to separate the unique number and any taxonomic classification info (relevant if rank given)}
}
\value{
phyloseq object
}
\description{
Don't confuse this with the phyloseq function taxa_names().
If your current taxa_names aren't what you want (e.g. they are long DNA sequences), this function will help you set sensible names.
It combines:
\itemize{
\item a prefix like tax, asv, or otu (pick an appropriate prefix or set your own)
\item a unique (sequential) number
\item classification information from a chosen taxonomic rank (optional)
}
}
\examples{
library(phyloseq)
# get example data
data("enterotype")
ps <- enterotype
head(taxa_names(ps)) # these are most fine (except the -1), but imagine you wanted new names

# consider storing the original names for reference (e.g. if they are DNA sequences)
old_taxa_names <- taxa_names(ps)

ps <- tax_name(ps)
taxa_names(ps) \%>\% head()

# probably better to include the genus info to make these names more informative
ps <- tax_name(ps, rank = "Genus")
taxa_names(ps) \%>\% head()

# store new names with old names in dataframe for reference
names_df <- tibble::tibble(old = old_taxa_names, new = taxa_names(ps))

# alternative settings
tax_name(ps, pad_number = FALSE) \%>\%
  taxa_names() \%>\%
  head()
tax_name(ps, prefix = "whateveryoulike") \%>\%
  taxa_names() \%>\%
  head()
tax_name(ps, rank = "Genus", sep = "-") \%>\%
  taxa_names() \%>\%
  head()
}
