% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_names2rank.R
\name{tax_names2rank}
\alias{tax_names2rank}
\title{Add taxa_names as last column in phyloseq tax_table}
\usage{
tax_names2rank(data, colname = "unique")
}
\arguments{
\item{data}{phyloseq object, or ps_extra or tax_table (taxonomyTable)}

\item{colname}{name of new rank to add at right side of tax_table}
}
\value{
same class object as passed in to data
}
\description{
The taxa names in your phyloseq may specify a further unique classification
of your taxa, e.g. ASVs, that is not otherwise represented in the tax_table itself.
This function fixes that, and allows you to include this level in taxatree_plots for example.
}
