% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_palette.R
\name{tax_palette}
\alias{tax_palette}
\title{Make a fixed taxa colour palette e.g. for comp_barplot}
\usage{
tax_palette(
  data,
  rank,
  n,
  by = sum,
  pal = "brewerPlus",
  add = c(other = "lightgrey"),
  ...
)
}
\arguments{
\item{data}{phyloseq or ps_extra}

\item{rank}{taxonomic rank name or "unique"}

\item{n}{number of colours / taxa (not including "other")}

\item{by}{tax sorting method for tax_sort e.g. sum}

\item{pal}{palette name from distinct_palette function}

\item{add}{name = value pairs appended to end of output, or NA for none}

\item{...}{other args are passed to tax_sort}
}
\value{
named character vector
}
\description{
Makes a named palette vector from your phyloseq dataset
considering overall abundance to assign colours (or some other sorting)
}
\examples{
library(ggplot2)
data(dietswap, package = "microbiome")

myPal <- tax_palette(dietswap, rank = "Genus", pal = "brewerPlus", n = 40)
myPal \%>\% tax_palette_plot() # just to check the palette

# plot one subset of data
dietswap \%>\%
  ps_filter(nationality == "AFR", timepoint == 1, sex == "male") \%>\%
  comp_barplot(
    tax_level = "Genus", n_taxa = 15,
    bar_outline_colour = NA, bar_width = 0.7,
    palette = myPal, label = NULL
  )

# plot a different subset of data (top taxa differ but colours are the same)
dietswap \%>\%
  ps_filter(nationality != "AFR", timepoint == 1, sex == "male") \%>\%
  comp_barplot(
    tax_level = "Genus", n_taxa = 15,
    bar_outline_colour = NA, bar_width = 0.7,
    palette = myPal, label = NULL
  )

}
