% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_palette.R
\name{tax_palette_plot}
\alias{tax_palette_plot}
\title{tax_palette plotting helper function}
\usage{
tax_palette_plot(named_pal_vec, max_n = NA)
}
\arguments{
\item{named_pal_vec}{vector of colours named by taxa (e.g. tax_palette output)}

\item{max_n}{NA to display all colours, or limit this}
}
\value{
ggplot
}
\description{
Check the named palette colour vector you created with tax_palette()
}
\examples{
library(ggplot2)
data(dietswap, package = "microbiome")

myPal <- tax_palette(dietswap, rank = "Genus", pal = "brewerPlus", n = 40)
myPal \%>\% tax_palette_plot() # just to check the palette

}
