% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_sort.R
\name{tax_reorder}
\alias{tax_reorder}
\title{Reorder taxa in phyloseq object using vector of names}
\usage{
tax_reorder(ps, tax_order, tree_warn = TRUE)
}
\arguments{
\item{ps}{phyloseq object}

\item{tax_order}{names or current numerical indices of taxa
in desired order and same length as taxa_names(ps)}

\item{tree_warn}{If phylogenetic tree is present in phyloseq phy_tree slot, taxa cannot be reordered.
Default behaviour of tax_sort is to remove the phylogenetic tree and warn about this.
tree_warn = FALSE will suppress the warning message, but still remove the tree!}
}
\value{
phyloseq object (always without phy_tree)
}
\description{
Reorder taxa in phyloseq object using vector of names
}
\examples{
data("dietswap", package = "microbiome")
new_order <- c(
  "Fusobacteria", "Cyanobacteria", "Verrucomicrobia", "Spirochaetes",
  "Actinobacteria", "Firmicutes", "Proteobacteria", "Bacteroidetes"
)
tax_agg(dietswap, rank = "Phylum")[["ps"]] \%>\%
  phyloseq::taxa_names()
tax_agg(dietswap, rank = "Phylum")[["ps"]] \%>\%
  microViz:::tax_reorder(tax_order = new_order) \%>\%
  phyloseq::taxa_names()
}
