% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_scale.R
\name{tax_scale}
\alias{tax_scale}
\title{Mean-center and SD-scale taxa in phyloseq}
\usage{
tax_scale(data, center = TRUE, scale = TRUE, do = NA)
}
\arguments{
\item{data}{phyloseq or ps_extra or otu_table}

\item{center}{if TRUE: center each taxon by subtracting its mean}

\item{scale}{if TRUE, divide each centred taxon by its standard deviation (or divide by RMS if not centred!)}

\item{do}{alternative argument that overrides center and scale options! takes "both", "scale", "center" or "neither"}
}
\description{
Wrapper for applying base scale function to phyloseq otu_table
}
\examples{
library(phyloseq)
library(microbiome)
data("dietswap")
ps <- dietswap
ps \%>\%
  otu_get() \%>\%
  .[1:6, 1:6]
# standard use (mean center and SD scale)
tax_scale(ps) \%>\%
  otu_get() \%>\%
  .[1:6, 1:6]
# RMS scale only (directly on otu_table)
otu_table(ps) \%>\%
  tax_scale(center = FALSE) \%>\%
  .[1:6, 1:6] # Aerococcus is NaN as prevalence=0
# example using alternative `do` argument (to center only, no scaling)
tax_scale(ps, do = "center") \%>\%
  otu_get() \%>\%
  .[1:6, 1:6]
}
