% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_select.R
\name{tax_select}
\alias{tax_select}
\title{Subset phyloseq object by (partial) taxa names}
\usage{
tax_select(
  ps,
  tax_list,
  ranks_searched = "all",
  strict_matches = FALSE,
  n_typos = 1,
  deselect = FALSE
)
}
\arguments{
\item{ps}{phyloseq object}

\item{tax_list}{e.g. c('g__Bifidobacterium', 'g__Akkermansia', 'g__Bacteroides', 'g__Streptococcus')}

\item{ranks_searched}{'all' or a list of which taxonomic ranks should be searched for the names in tax_list?}

\item{strict_matches}{only perfect full name matches allowed if TRUE}

\item{n_typos}{how many typos to allow in each name? uses agrep approximate matching if > 0}

\item{deselect}{if TRUE, the matching taxa will be REMOVED instead!}
}
\value{
tax-filtered phyloseq object
}
\description{
Convenient name-based taxa selection/filtering of phyloseq object, including approximate name matching
}
\details{
Takes a phyloseq with tax table and a list/vector of taxonomic names (full or partial matches).
It will also search the otu names/rownames, BUT only for perfect matches.
}
\examples{
# Get example phyloseq object data
library(microbiome)
data("dietswap")
pSeq <- dietswap

# SELECTION EXAMPLES #
a <- pSeq \%>\% tax_select(tax_list = "Bif", n_typos = 0, ranks_searched = "Genus")
b <- pSeq \%>\% tax_select(tax_list = "Bifidobacterium", n_typos = 0)
c <- pSeq \%>\% tax_select(tax_list = "Bif", n_typos = 1)
identical(a, b) # TRUE
identical(a, c) # FALSE
pSeq \%>\% tax_select(tax_list = "Bifidobactrium") # default 1 typo allowed
one <- pSeq \%>\% tax_select(tax_list = "Akkarmensia", n_typos = 2)
two <- pSeq \%>\% tax_select(tax_list = "Akkermansia", n_typos = 0)
identical(one, two) # TRUE
# DESELECTION EXAMPLE # #
pSeq \%>\% tax_select(tax_list = "Bif", strict_matches = FALSE, deselect = TRUE)
# Incorrect example
# pSeq \%>\% tax_select(tax_list = "Bif", strict_matches = TRUE) # fails
}
