% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_sort.R
\name{tax_sort}
\alias{tax_sort}
\title{Sort taxa in phyloseq otu_table and tax_table}
\usage{
tax_sort(data, by = "name", ..., tree_warn = TRUE)
}
\arguments{
\item{data}{ps_extra or phyloseq}

\item{by}{how to sort, see description}

\item{...}{used if summary function given, or pass \code{undetected} arg for tax_transform("binary") if by = "prev" or "prevalence"}

\item{tree_warn}{If phylogenetic tree is present in phyloseq phy_tree slot, taxa cannot be reordered.
Default behaviour of tax_sort is to remove the phylogenetic tree and warn about this.
tree_warn = FALSE will suppress the warning message, but still remove the tree!}
}
\value{
sorted phyloseq or ps_extra
}
\description{
Multiple ways of sorting taxa are possible and determined by the \code{by} argument.
The \code{by} argument must be one of:
\itemize{
\item 'rev' to reverse the current order
\item 'name' or 'names' (sort alphabetically by taxa_names)
\item a taxonomic rank name (sort alphabetically by this rank)
\item a sample name (descending abundance sorting within that sample)
\item summary stat. function e.g. \code{sum} or \code{mean}
}
}
\details{
Don't forget to pass \code{na.rm = TRUE} to \code{...}
if using a summary stat function in \code{by}
}
\examples{
library(microbiome)
data("dietswap")
dietswap

# reverse current order
dietswap \%>\%
  tax_sort("rev") \%>\%
  tax_table() \%>\%
  head(30)

# sort alphabetically by a taxonomic rank (or "names" for taxa_names)
dietswap \%>\%
  tax_sort("Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# sequentially sorting by higher ranks
# sets tax_table in nested alphabetical order
dietswap \%>\%
  tax_sort("names") \%>\%
  tax_sort("Genus") \%>\%
  tax_sort("Family") \%>\%
  tax_sort("Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# order by descending abundance in a single named sample
dietswap \%>\%
  tax_sort("Sample-1") \%>\%
  otu_table() \%>\%
  .[1:5, 1:4]

# sort by function e.g. median abundance
dietswap \%>\%
  tax_sort(by = median) \%>\%
  taxa_names() \%>\%
  head(20)

# sum order should always equal mean order if non-negative abundances
# don't forget to add na.rm = TRUE if you expect NAs in otu_table somehow
dietswap \%>\%
  tax_sort(by = sum, na.rm = TRUE) \%>\%
  taxa_names() \%>\%
  head(20)

# if your phyloseq object has a phylogenetic tree,
# tax_sort will remove the tree, and warn you about this.
}
