% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_tree_funs.R
\name{tax_tree_funs}
\alias{tax_tree_funs}
\alias{tax_tree_nodes}
\alias{tax_tree_edges}
\title{Create node and edge dataframes for taxonomic tree graphs}
\usage{
tax_tree_nodes(ps, ranks = "all")

tax_tree_edges(nodes_df)
}
\arguments{
\item{ps}{phyloseq object}

\item{ranks}{selection of taxonomic ranks to make nodes for (all, names or numbers)}

\item{nodes_df}{the dataframe output of \code{tax_tree_nodes} (possibly with extra info/stats joined to it)}
}
\description{
\itemize{
\item \code{tax_tree_nodes} creates taxon nodes and calculates basic info about each taxon. From a phyloseq object.
\item \code{tax_tree_edges} uses the output of \code{tax_tree_nodes} to create a dataframe of edges.
}

\code{tax_tree_nodes} makes nodes for taxa at all ranks or for a list of consecutive (rooted) ranks.
You can also join at dataframe of additional taxon level information to the output of \code{tax_tree_nodes} before calling tax_tree_edges, e.g. the output of \code{corncob_models_to_var_stats}.
}
