% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_models.R
\name{taxatree_models}
\alias{taxatree_models}
\title{Statistical modelling for individual taxa across multiple ranks}
\usage{
taxatree_models(
  ps,
  ranks = NULL,
  type = "lm",
  variables = NULL,
  formula = NULL,
  use_future = FALSE,
  checkVars = TRUE,
  checkNA = "warning",
  verbose = "rank",
  trans = "identity",
  trans_args = list(),
  ...
)
}
\arguments{
\item{ps}{phyloseq object or psExtra}

\item{ranks}{vector of rank names at which to aggregate taxa for modelling}

\item{type}{name of regression modelling function, or the function itself}

\item{variables}{vector of variable names, to be used as model formula right hand side.
If variables is a list, not a vector, a model is fit for each entry in list.}

\item{formula}{Right hand side of a formula, as a formula object or character string. Or a list of these.
(alternative to variables argument, do not provide both)}

\item{use_future}{if TRUE parallel processing with future is possible, see details of ?tax_model.}

\item{checkVars}{check variance of variables?}

\item{checkNA}{check variables for NAs?}

\item{verbose}{message about progress: "rank" only notifies which rank is being processed;
TRUE notifies you about each taxon being processed; FALSE for no messages.}

\item{trans}{name of tax_transform transformation to apply to aggregated taxa before fitting statistical models}

\item{trans_args}{named list of any additional arguments to tax_transform e.g. list(zero_replace = "halfmin")}

\item{...}{extra arguments are passed directly to modelling function}
}
\description{
\code{taxatree_models} runs \code{tax_model} on every taxon at multiple taxonomic
ranks (you choose which ranks with the plural \code{ranks} argument).
It returns the results as a named nested list of models
attached to a psExtra object.
One list per rank, one model per taxon at each rank.

The result can then be used with \code{taxatree_models2stats} to extract a
dataframe of statistics for use with \code{taxatree_plots}.
}
\seealso{
\code{\link{tax_model}} for more details and examples

\code{\link{taxatree_plots}} for how to plot the output of \code{taxatree_models}
}
