% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_plot.R
\name{taxatree_plot}
\alias{taxatree_plot}
\title{Plot taxonomic stats on a ggraph tree}
\usage{
taxatree_plot(
  graph,
  preset_style = NA,
  size_stat = "taxon_mean",
  colour_stat = "b_mu",
  max_node_size = 8,
  max_edge_width = max_node_size - 2,
  sig_var = "p_mu",
  sig_threshold = 0.05,
  sig_stroke_width = 1.5,
  colour_palette = "Green-Brown",
  reverse_colours = FALSE,
  luminance_l2 = 80,
  colour_lims = c(-3, 3),
  layout = "tree"
)
}
\arguments{
\item{graph}{tidygraph made with taxatree_nodes and taxatree_edges}

\item{size_stat}{name of variable to scale size of nodes and edges (natural log scale)}

\item{colour_stat}{name of variable to scale colour/fill of nodes and edges (natural log scale)}

\item{max_node_size}{set this to avoid overlapping nodes at different scales}

\item{max_edge_width}{max edge width (optionally relative to max node size)}

\item{sig_var}{name of variable used to define statistical significance (outlines significant nodes) - NULL for none}

\item{sig_threshold}{value of sig_var below which statistical significance outlines are drawn}

\item{sig_stroke_width}{fixed width of statistical significance node outlines}

\item{colour_palette}{any diverging hcl colour palette name see: \code{colorspace::hcl_palettes("diverging", n = 7, plot = TRUE)}}

\item{reverse_colours}{reverse direction of colour scale?}

\item{luminance_l2}{luminance of colourscale midpoint (higher is lighter)}

\item{colour_lims}{limits of colour and fill scale, used to harmonise across several plots (values squished to fit this range!)}

\item{layout}{name of ggraph layout or manually specified layout with data.frame of x and y coordinates}
}
\value{
ggraph ggplot
}
\description{
Uses ggraph (see help for main underlying graphing function with ?ggraph::ggraph)
See tree-development vignette 2021.
}
