% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_plotkey.R
\name{taxatreePruneLabelArgs}
\alias{taxatreePruneLabelArgs}
\title{Helper, removes unwanted arguments from geom_* call args list}
\usage{
taxatreePruneLabelArgs(fun, args)
}
\arguments{
\item{fun}{function name or function itself e.g. geom_label_repel}

\item{args}{named list of args}
}
\value{
args list possibly with elements removed
}
\description{
Helper, removes unwanted arguments from geom_* call args list
}
\examples{
microViz:::taxatreePruneLabelArgs(fun = "geom_label", args = list(size = 2, seed = 2))
microViz:::taxatreePruneLabelArgs(fun = library, args = list(size = 2, seed = 2))
}
\keyword{internal}
