% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxAnnotation.R
\name{taxCalcAbund}
\alias{taxCalcAbund}
\title{Internal helper for anno_tax_box and anno_tax_density}
\usage{
taxCalcAbund(data, taxa, trans, zero_replace, undetected, only_detected)
}
\arguments{
\item{data}{ps_extra or phyloseq}

\item{taxa}{vector of taxa to return abundances for}

\item{trans}{transformation for tax_transform}

\item{zero_replace}{zero_replace for tax_transform}

\item{undetected}{the value above which taxa are classed as detected/present in a sample}

\item{only_detected}{replace values with NaN for samples where taxon abundance is <= undetected}
}
\value{
numeric matrix, taxa as columns
}
\description{
Get taxon abundance matrix (with undetected replaced with NaN, optionally)
}
