% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_anova}
\alias{test_anova}
\title{Test gene expression differences between multiple groups with one-way ANOVA.}
\usage{
test_anova(
  data,
  split_fct,
  variables,
  confounder = NULL,
  adj_method = "BH",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{confounder}{name of the confounding variable. Defaults to NULL, which
means that no confounder is included in the model.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[stats]{aov}}.}
}
\value{
\code{NULL} with a warning.
}
\description{
The testing functions are not active anymore, because we offer a faster and
more versatile tools for high-throughput statistical hypothesis testing.
Please use functions from \code{fastTest} package instead
(https://github.com/PiotrTymoszuk/fastTest/)
}
\details{
These functions are not active anymore and return \code{NULL} with a warning
}
