% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_nb}
\alias{test_nb}
\title{Test gene expression differences with negative binomial modeling.}
\usage{
test_nb(
  data,
  split_fct,
  variables,
  confounder = NULL,
  adj_method = "BH",
  dev_test = "Chisq",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{confounder}{name of the confounding variable. Defaults to NULL, which
means that no confounder is include in the model.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{dev_test}{name of the test used for analysis of deviance. One of:
'Rao', 'LRT', 'Chisq' (default), 'F', 'Cp'.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[MASS]{glm.nb}}.}
}
\value{
\code{NULL} with a warning.
}
\description{
The testing functions are not active anymore, because we offer a faster and
more versatile tools for high-throughput statistical hypothesis testing.
Please use functions from \code{fastTest} package instead
(https://github.com/PiotrTymoszuk/fastTest/).
When looking specifically for negative binomial modeling of gene expression
data, please use EdgeR (https://bioconductor.org/packages/release/bioc/html/edgeR.html)
or DESeq2 (https://bioconductor.org/packages/release/bioc/html/DESeq2.html)
instead.
}
\details{
These functions are not active anymore and return \code{NULL} with a warning
}
