% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_two_groups}
\alias{test_two_groups}
\title{Test gene expression differences between two groups.}
\usage{
test_two_groups(
  data,
  split_fct,
  variables,
  type = c("t", "wilcox"),
  adj_method = "BH",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{type}{test type, currently 't' for T test (default) or 'wilcox' for
the Wilcoxon/Mann-Whitney test.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[stats]{t.test}} or
\code{\link[stats]{wilcox.test}}.}
}
\value{
a data frame with the regulation statistic, confidence intervals,
t and p values (raw and corrected) as well as effect size measures.
}
\description{
Compares gene expression differences between two groups with
T tests or Wilcoxon tests.
}
\details{
The function returns Cohen's d as an effect size metric for T tests and
r as an effect size statistic for Wilcoxon tests. The later is computed
with \code{\link[rcompanion]{wilcoxonR}}.
}
