% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_micro}
\alias{theme_micro}
\title{Deafult graphics theme for microViz tools.}
\usage{
theme_micro()
}
\description{
The default ggplot graphics theme for graphical output of the
package functions.

A ggplot theme object.
}
