% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_extra.R
\name{upgrade_ps_extra_to_psExtra}
\alias{upgrade_ps_extra_to_psExtra}
\title{Convert old format "ps_extra" objects to new "psExtra" objects}
\usage{
upgrade_ps_extra_to_psExtra(ps_extra)
}
\arguments{
\item{ps_extra}{an old format "ps_extra" object,
as generated by old microViz versions (< 0.10.0)}
}
\value{
new format "psExtra" S4 object
}
\description{
This will only be necessary if you have saved old format "ps_extra" objects
generated by an old microViz version (< 0.10.0), and you cannot or do not
want to regenerate these old format objects from your original phyloseq object.
}
\examples{
# read your old saved 'ps_extra' object that you want to keep using
# oldObject <- readRDS("old-object-path.rds")
# newObject <- upgrade_ps_extra_to_psExtra(oldObject)
# continue with your next analysis or plotting steps...
}
