% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varAnnotation.R
\name{varAnnotation}
\alias{varAnnotation}
\title{Helper to specify a HeatmapAnnotation for variables in cor_heatmap}
\usage{
varAnnotation(
  ...,
  name,
  annotation_legend_param = list(),
  show_legend = TRUE,
  gp = grid::gpar(col = NA),
  border = FALSE,
  gap = grid::unit(2, "mm"),
  show_annotation_name = TRUE,
  annotation_label = NULL,
  annotation_name_gp = grid::gpar(),
  annotation_name_offset = NULL,
  annotation_name_rot = NULL,
  annotation_name_align = FALSE,
  annotation_name_side = "auto",
  .data = NULL,
  .vars = NULL,
  .side = NULL
)
}
\arguments{
\item{...}{Name-value pairs where the names correspond to annotation names and values
are the output of variable annotation functions
such as anno_var_box(), or manually specified AnnotationFunction objects}

\item{name}{Name of the heatmap annotation, optional.}

\item{annotation_legend_param}{A list which contains parameters for annotation legends. See \code{\link[ComplexHeatmap]{color_mapping_legend,ColorMapping-method}} for all possible options.}

\item{show_legend}{Whether show annotation legends. The value can be one single value or a vector.}

\item{gp}{Graphic parameters for simple annotations (with \code{fill} parameter ignored).}

\item{border}{border of single annotations.}

\item{gap}{Gap between annotations. It can be a single value or a vector of \code{\link[grid]{unit}} objects.}

\item{show_annotation_name}{Whether show annotation names? For column annotation, annotation names are drawn either on the left or the right, and for row annotations, names are draw either on top or at the bottom. The value can be a vector.}

\item{annotation_label}{Labels for the annotations. By default it is the same as individual annotation names.}

\item{annotation_name_gp}{Graphic parameters for anntation names. Graphic paramters can be vectors.}

\item{annotation_name_offset}{Offset to the annotation names, a \code{\link[grid]{unit}} object. The value can be a vector.}

\item{annotation_name_rot}{Rotation of the annotation names. The value can be a vector.}

\item{annotation_name_align}{Whether to align the annotation names.}

\item{annotation_name_side}{Side of the annotation names.}

\item{.data}{OPTIONAL phyloseq or ps_extra,
only set this to override use of same data as in heatmap}

\item{.vars}{OPTIONAL selection vector of variables (names, numbers or logical),
only set this if providing .data argument to override default}

\item{.side}{OPTIONAL string, indicating the side for the variable annotations:
only set this to override default}
}
\value{
HeatmapAnnotation object
}
\description{
Helper to specify a HeatmapAnnotation for variables in cor_heatmap
}
