% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_similarity.R
\name{vector_similarity}
\alias{vector_similarity}
\alias{set_similarity}
\title{Similarity coefficients for pairs and ists of vectors.}
\usage{
vector_similarity(
  x,
  y,
  method = c("jaccard", "overlap", "dice", "tversky"),
  a = 1,
  b = 1
)

set_similarity(
  x,
  method = c("jaccard", "overlap", "dice", "tversky"),
  a = 1,
  b = 1
)
}
\arguments{
\item{x}{a vector coercible to a character vector (\verb{vector similarity}) or
a list of such vectors (\code{set_similarity}).}

\item{y}{a vector coercible to a character vector.}

\item{method}{type of the coefficient.}

\item{a}{a single numeric, the alpha constant used for calculation of
the Tversky similarity coefficient.}

\item{b}{a single numeric, the beta constant used for calculation of
the Tversky similarity coefficient.}
}
\value{
a numeric (\code{vector_similarity()}) or a numeric matrix of similarity
coefficients (\code{set_similarity()}).
}
\description{
The functions \code{vector_similarity()} and \code{set_similarity()} calculate the most
popular similarity coefficients for sets: Jaccard, overlap (also called
Szymkiewicz–Simpson index), Dice (Dice-Sørensen), and Tversky coefficients.
While \code{vector_similarity()} takes a pair of vectors as arguments and returns
a single numeric, \code{set_similarity()} operates on lists of vectors and returns
a matrix of similarity coefficients.
}
\details{
\code{set_similarity()} returns a matrix of class \code{set_sim} which can be handled
as a usual matrix and, in addition, allows for dimensionality reduction with
the \code{\link{components.set_sim}} method, and visualization of distances between
the list elements via \code{\link{plot.set_sim}}.
Those methods are useful at exploring similarity of e.g. sets of
differentially regulated genes identified in multiple data sets.
}
