% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{wilcox_tester}
\alias{wilcox_tester}
\title{Wilcoxon test.}
\usage{
wilcox_tester(data, split_fct, variable, conf.int = FALSE, ...)
}
\arguments{
\item{data}{a data frame.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variable}{dependent variable name.}

\item{conf.int}{logical, should confidence intervals be calculated?}

\item{...}{additional arguments passed to \code{\link[stats]{wilcox.test}}.}
}
\value{
a data frame with the regulation statistic, confidence intervals,
t and p values
}
\description{
Test for the difference with a T test.
}
