% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_vector}
\alias{wrap_vector}
\title{Wrap text with a defined word count per line.}
\usage{
wrap_vector(x, len = 5, word_sep = ", ", line_sep = ",\\n")
}
\arguments{
\item{x}{a vector.}

\item{len}{line length in words.}

\item{word_sep}{word separator, comma by default.}

\item{line_sep}{line separator, comma + newline by default.}
}
\value{
a text value.
}
\description{
Converts a character vector into a text with the defined line
length (word count). If a non-character vector is provided, it will be
coerced to a character one.
}
