% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{zScores}
\alias{zScores}
\alias{zScores.default}
\alias{zScores.matrix}
\alias{zScores.data.frame}
\title{Normalization, standardization and Z-scores.}
\usage{
zScores(x, ...)

\method{zScores}{default}(
  x,
  center = c("mean", "median", "geo_mean", "harm_mean"),
  dispersion = c("sd", "sem"),
  ...
)

\method{zScores}{matrix}(
  x,
  feature_type = c("columns", "rows"),
  center = c("mean", "median", "geo_mean", "harm_mean"),
  dispersion = c("sd", "sem"),
  ...
)

\method{zScores}{data.frame}(
  x,
  variables = NULL,
  center = c("mean", "median", "geo_mean", "harm_mean"),
  dispersion = c("sd", "sem"),
  ...
)
}
\arguments{
\item{x}{a numeric vector, numeric matrix or a data frame.}

\item{...}{extra arguments passed to methods.}

\item{center}{a string that specifies the centrality statistic: mean, median,
geometric mean or harmonic mean.}

\item{dispersion}{a string that specified the dispersion statistic:
SD or SEM.}

\item{feature_type}{specifies if columns (default) or rows are subjected to
scaling.}

\item{variables}{a vector of names of numeric variables to be normalized.
If \code{NULL}, all variables of a data frame are scaled. Please note, that
variable selection makes the computation slower.}
}
\value{
a numeric vector, numeric matrix or a data frame.
}
\description{
Calculates classical Z-scores or their variants employing median, geometric
mean, harmonic mean as centrality statistics, and SD or SEM as dispersion
statistics.
}
\details{
The Z-scores are calculated for a numeric vector \code{x} with a general formula:

\deqn{Z(x_i) = \frac{x_i - centFunc(x)}{dispFunc(x)}}

where \eqn{centFunc} stands for a function that returns a centrality
statistic such as mean or median, and \eqn{dispFunc(x)} represents a
function that returns a dispersion statistic such as standard deviation (SD)
or standard error of the mean (SEM).
}
