% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAnnotation.R
\name{anno_cat}
\alias{anno_cat}
\title{Create colored rectangle annotations for categorical data}
\usage{
anno_cat(
  x,
  which,
  renamer = identity,
  col = distinct_palette(),
  width = NULL,
  height = NULL,
  gp = grid::gpar(col = "white", lwd = 1),
  border = FALSE,
  border_gp = grid::gpar(col = "black", fill = NA),
  legend = TRUE,
  legend_title = ""
)
}
\arguments{
\item{x}{data vector, treated as categorical}

\item{which}{Whether it is a column annotation or a row annotation?}

\item{col}{colors vector, at least as long as unique(x), optionally named by x levels}

\item{width}{Width of the annotation. The value should be an absolute unit. Width is not allowed to be set for column annotation.}

\item{height}{Height of the annotation. The value should be an absolute unit. Height is not allowed to be set for row annotation.}

\item{gp}{gpar object: don't set the fill!}

\item{border}{draw a border around the whole annotation (or each split)}

\item{border_gp}{gpar for border, if drawn}
}
\value{
AnnotationFunction
}
\description{
Similar to anno_simple but with individual boxes!
}
\examples{
library(ComplexHeatmap)
# draw the annotation without a heatmap, you will never normally do this!
vp <- viewport(width = 0.75, height = 0.75)

grid::grid.newpage()
pushViewport(vp)
cats <- letters[1:4]
draw(anno_cat(cats, which = "row"))

grid::grid.newpage()
pushViewport(vp)
draw(
  anno_cat(
    x = cats, col = structure(names = cats, 1:4), which = "column",
    gp = grid::gpar(col = "black", lwd = 5)
  )
)

# developer note #
# list of annotations can be split and ordered (adding NULL makes a list)
# https://jokergoo.github.io/ComplexHeatmap-reference/book/a-list-of-heatmaps.html#concatenate-only-the-annotations
grid::grid.newpage()
pushViewport(vp)
(rowAnnotation(hi = anno_cat(cats, which = "row", border = T)) + NULL) \%>\%
  draw(row_split = c(1, 1:3), row_order = 4:1)
pushViewport(viewport(x = 0.6))
draw(anno_cat(cats, "row", legend_title = "abcd") \%>\% attr("Legend"))
}
