% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAnnotation.R
\name{anno_sample}
\alias{anno_sample}
\title{Helper to specify simple comp_heatmap annotation for other sample data}
\usage{
anno_sample(var, fun = identity, data = NULL, samples = NULL)
}
\arguments{
\item{var}{name of variable to use for annotation data}

\item{fun}{function to transform variable \code{var}}

\item{data}{OPTIONAL phyloseq or ps_extra,
only set this to override use of same data as in heatmap}

\item{samples}{OPTIONAL selection vector of sample names,
only set this if providing data argument to override default}

\item{...}{
  Arguments passed on to \code{\link[ComplexHeatmap:anno_simple]{ComplexHeatmap::anno_simple}}
  \describe{
    \item{\code{col}}{Color that maps to \code{x}. If \code{x} is numeric and needs a continuous mapping, \code{col}  should be a color mapping function which accepts a vector of values and returns a vector of colors. Normally it is generated by \code{\link[circlize]{colorRamp2}}. If \code{x} is discrete (numeric or character) and needs a discrete color mapping, \code{col} should be a vector of  colors with levels in \code{x} as vector names. If \code{col} is not specified, the color mapping is randomly generated by \code{ComplexHeatmap:::default_col}.}
    \item{\code{na_col}}{Color for NA value.}
  }}
}
\value{
vector of values
}
\description{
Use this as an argument to sampleAnnotation(),
which itself is used by comp_heatmap() as sample_anno argument.

This creates a vector, which sampleAnnotation() interprets as a
simple annotation, so then you set colours and legend parameters
for each simple annotation as further arguments in sampleAnnotation.
}
\examples{
# see `?sampleAnnotation()`
}
\seealso{
\code{\link[=sampleAnnotation]{sampleAnnotation()}}
}
