% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps-sampleAnnotation.R
\name{anno_sample}
\alias{anno_sample}
\title{Helper to specify simple comp_heatmap annotation for other sample data}
\usage{
anno_sample(var, fun = identity, data = NULL, samples = NULL)
}
\arguments{
\item{var}{name of variable to use for annotation data}

\item{fun}{function to transform variable \code{var}}

\item{data}{OPTIONAL phyloseq or psExtra,
only set this to override use of same data as in heatmap}

\item{samples}{OPTIONAL selection vector of sample names,
only set this if providing data argument to override default}
}
\value{
vector of values
}
\description{
Use this as an argument to sampleAnnotation(),
which itself is used by comp_heatmap() as sample_anno argument.

This creates a vector, which sampleAnnotation() interprets as a
simple annotation, so then you set colours and legend parameters
for each simple annotation as further arguments in sampleAnnotation.
}
\examples{
# see `?sampleAnnotation()`
}
\seealso{
\code{\link[=sampleAnnotation]{sampleAnnotation()}}
}
