% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxAnnotation.R
\name{anno_tax_box}
\alias{anno_tax_box}
\title{Helper to specify heatmap annotation for showing taxa abundance on boxplot}
\usage{
anno_tax_box(
  undetected = 0,
  only_detected = TRUE,
  trans = "compositional",
  zero_replace = 0,
  use_counts = TRUE,
  size = grid::unit(30, "mm"),
  border = TRUE,
  gp = grid::gpar(fill = "#CCCCCC"),
  ylim = NULL,
  extend = 0.05,
  outline = TRUE,
  box_width = 0.6,
  pch = 1,
  pointsize = grid::unit(0.5, "mm"),
  axis = TRUE,
  ...,
  data = NULL,
  taxa = NULL,
  which = NULL
)
}
\arguments{
\item{undetected}{the value above which taxa are classed as detected/present in a sample}

\item{only_detected}{only plot values for samples where the taxon abundance is > undetected}

\item{trans}{transformation suitable for tax_transform}

\item{zero_replace}{zero_replace value for for tax_transform}

\item{use_counts}{try to retrieve counts from data object?}

\item{size}{width or height as a grid unit object}

\item{border}{Wether draw borders of the annotation region?}

\item{gp}{Graphic parameters for points. The length of each graphic parameter can be 1, length of \code{x} if \code{x} is a vector, or number of columns of \code{x} is \code{x} is a matrix.}

\item{ylim}{Data ranges. By default it is \code{range(x)} if \code{x} is a vector, or \code{range(rowSums(x))} if \code{x} is a matrix.}

\item{extend}{The extension to both side of \code{ylim}. The value is a percent value corresponding to \code{ylim[2] - ylim[1]}.}

\item{outline}{Whether draw outline of boxplots?}

\item{box_width}{Relative width of boxes. The value should be smaller than one.}

\item{pch}{Point style.}

\item{pointsize}{size of outlier points, as grid::unit() object}

\item{axis}{Whether to add axis?}

\item{...}{Other arguments.}

\item{data}{OPTIONAL phyloseq or ps_extra,
only set this to override use of same data as in heatmap}

\item{taxa}{OPTIONAL selection vector of taxa (names, numbers or logical),
only set this if providing data argument to override default}

\item{which}{OPTIONAL indicating if it is a 'column' or a 'row' annotation,
only set this if providing data argument to override default}
}
\value{
function or ComplexHeatmap AnnotationFunction object
}
\description{
Use this as an argument to taxAnnotation(),
which itself is used by cor_heatmap and comp_heatmap as tax_anno argument.
}
\examples{
library("ComplexHeatmap")
data("ibd_phylo", package = "corncob")
psq <- tax_filter(ibd_phylo, min_prevalence = 5)
psq <- tax_mutate(psq, Species = NULL)
psq <- tax_fix(psq)
psq <- tax_agg(psq, rank = "Family")
taxa <- tax_top(psq, n = 15, rank = "Family")
# makes a function that takes data, taxa and which (at minimum)
fun <- anno_tax_box()
# manually specify the prevalence barplot function by giving it data etc.
heatmapAnnoFunction <- fun(data = psq, which = "column", taxa = taxa)
# draw the barplot without a heatmap, you will never normally do this!
grid::grid.newpage()
draw(heatmapAnnoFunction)
# let's change some style options and specify the data up front
grid::grid.newpage()
draw(anno_tax_box(
 data = psq, taxa = taxa, which = "row", pointsize = grid::unit(1, "mm"),
 gp = grid::gpar(fill = "red"), border = FALSE, box_width = 0.2
))
# clear drawings
grid::grid.newpage()
}
