% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{anova_tester}
\alias{anova_tester}
\title{One-way ANOVA.}
\usage{
anova_tester(data, split_fct, variable, ...)
}
\arguments{
\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variable}{dependent variable name.}

\item{...}{additional arguments passed to \code{\link[stats]{aov}}.}
}
\value{
a data frame with the F statistic, degrees of freedom and p value.
}
\description{
A simplified version of one-way ANOVA suitable for whole-genome
testing.
}
