% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distances between pairs of samples in phyloseq object}
\usage{
calc_dist(
  data,
  dist = c("bray", "gunifrac", "unifrac", "wunifrac", "va-wunifrac", "aitchison",
    "euclidean")[1],
  return = "all",
  gunifrac_alpha = 0.5,
  ...
)
}
\arguments{
\item{data}{phyloseq object or list object output from tax_agg}

\item{dist}{name of distance to calculate between pairs of samples}

\item{return}{choose which parts of list object to return}

\item{gunifrac_alpha}{setting alpha value only relevant if gunifrac distance used}

\item{...}{optional distance-specific named args passed to phyloseq::distance()}
}
\value{
list with distance matrix, phyloseq object, and name of distance used
}
\description{
For use directly with phyloseq object or with list output of tax_agg.
Computes various unifrac distances, aitchison distance or any distance from vegdist.
Returns phyloseq and name of distance in addition to the distance matrix itself!
Result intended to be piped into ordin8 or permanova functions.
}
\details{
Aitchison distance note: You should EITHER
\enumerate{
\item skip the calc_dist function and call ordin8(method = "RDA") directly on an object with taxa transformed with tax_transform(transformation = "clr")
\item pass an object with untransformed (or 'identity' transformed) taxa to the data argument of calc_dist and specify dist = "aitchison".
If ordination plots with taxon loading distances are desired, users should prefer option 1.
If the distance matrix is required for permanova, users should prefer option 2.
}
}
\examples{
# bray curtis distance on genera-level features
data("dietswap", package = "microbiome")
bc <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  calc_dist("bray")
class(bc)

# gunifrac distance using phyloseq input
data("esophagus", package = "phyloseq")
gunifrac <- esophagus \%>\%
  calc_dist("gunifrac", return = "distMat")
class(gunifrac)
}
