% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_df}
\alias{check_df}
\alias{check_mtx}
\title{Check for a numeric data frame.}
\usage{
check_df(x)

check_mtx(x)
}
\arguments{
\item{x}{a data frame or a matrix.}
}
\value{
a logical value.
}
\description{
The functions checks if all columns of a data frame are numeric or
if a numeric matrix is provided.
}
