% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify}
\alias{classify}
\title{Classify variables by their specificity for data set subsets.}
\usage{
classify(data, variables, split_fct, direction = c("<", ">", "auto"))
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector of names of the variables of interest.}

\item{split_fct}{splitting factor defining the data set subsets.}

\item{direction}{direction of the comparison in the ROC analysis.
See: \code{\link[pROC]{roc}} for details.}
}
\value{
A list with two elements, \code{auc} and \code{classification}.
The \code{auc} element is a data frame that stores the complete analysis results.
\code{classification} lists variables along with their most specific subsets of
the data set. Both elements are data frames with the following columns:
\itemize{
\item \code{variable}: variable name
\item \code{split_fct}: a column named after the splitting factor
\item \code{auc}: are under the ROC curve for detection of the particular data subset
\item \code{delta_auc}: difference in AUC ROC for the particular subset and mean AUC
ROC computed across all subsets.
}
}
\description{
Classifies selected features, e.g. differentially expressed genes, by their
specificity for the data subsets defined by a splitting factor.
}
\details{
Specificity for a given data subset is determined by calculation of area
under the receiver-operating characteristic (ROC) in a comparison of the
subset with the pooled remaining subsets (i.e. the rest of the data set).
Of note, these remaining subsets are set as the baseline or control in the
ROC analysis. AUC ROC is computed with \code{\link[pROC]{roc}}.
}
