% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{comp_heatmap}
\alias{comp_heatmap}
\title{Heatmap of taxonomic composition across samples}
\usage{
comp_heatmap(
  data,
  taxa = NA,
  samples = NA,
  anno_tax = NULL,
  anno_samples = NULL,
  colors = heat_palette(),
  tax_transform_colors = "clr",
  tax_scale_colors = "neither",
  tax_select = NA,
  sample_select = NA,
  taxa_which = "row",
  seriation_method = "OLO_ward",
  seriation_dist = "euclidean",
  seriation_method_col = seriation_method,
  seriation_dist_col = seriation_dist,
  numbers = NULL,
  tax_transform_numbers = tax_transform_colors,
  tax_scale_numbers = tax_scale_colors,
  ...
)
}
\arguments{
\item{tax_scale_numbers}{}
}
\description{
Heatmap of taxonomic composition across samples
}
