% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{comp_heatmap}
\alias{comp_heatmap}
\title{Heatmap of taxonomic composition across samples}
\usage{
comp_heatmap(
  data,
  taxa = NA,
  taxa_side = "right",
  tax_anno = NULL,
  taxon_renamer = identity,
  samples = NA,
  sample_side = adjacent_side(taxa_side),
  sample_anno = NULL,
  sample_names_show = FALSE,
  colors = heat_palette(palette = "Rocket", rev = TRUE),
  numbers = NULL,
  sample_seriation = "OLO_ward",
  sample_ser_dist = "euclidean",
  sample_ser_counts = !sample_ser_dist \%in\% c("euclidean", "maximum", "manhattan",
    "canberra", "binary"),
  sample_ser_trans = NULL,
  tax_seriation = "OLO_ward",
  tax_ser_dist = "euclidean",
  tax_ser_counts = FALSE,
  tax_ser_trans = NULL,
  numbers_trans = NULL,
  numbers_zero_replace = 0,
  numbers_use_counts = TRUE,
  grid_col = "white",
  grid_lwd = 0.1,
  name = "Abd.",
  anno_tax = NULL,
  ...
)
}
\arguments{
\item{data}{phyloseq extra output of tax_transform or tax_agg}

\item{taxa}{list of taxa to include (selection occurs AFTER any tax_transform and scaling)}

\item{taxa_side}{controls heatmap orientation and where any anno_tax annotations are placed (top/bottom/left/right)}

\item{samples}{list of taxa to include (selection occurs AFTER any tax_transform and scaling)}

\item{colors}{output of heat_palette() to set heatmap fill color scheme}

\item{numbers}{NULL or output of heat_numbers() to draw numbers on heatmap cells}

\item{name}{used as legend title (colourbar)}

\item{anno_tax}{NULL or tax_anno() list output}

\item{...}{
  Arguments passed on to \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap}}
  \describe{
    \item{\code{row_dend_side}}{Should the row dendrogram be put on the left or right of the heatmap?}
    \item{\code{row_dend_width}}{Width of the row dendrogram, should be a \code{\link[grid]{unit}} object.}
    \item{\code{show_row_dend}}{Whether show row dendrogram?}
    \item{\code{row_dend_gp}}{Graphic parameters for the dendrogram segments. If users already provide a \code{\link[stats]{dendrogram}} object with edges rendered, this argument will be ignored.}
    \item{\code{show_row_names}}{Whether show row names.}
    \item{\code{row_names_gp}}{Graphic parameters for row names.}
    \item{\code{row_names_rot}}{Rotation of row names.}
    \item{\code{row_names_centered}}{Should row names put centered?}
  }}

\item{anno_samples}{NULL only support so far TODO}

\item{seriation_method}{method to order the rows (in seriation::seriate)}

\item{seriation_dist}{distance to use in seriation_method (if needed)}

\item{seriation_method_col}{method to order the columns (in seriation::seriate)}

\item{seriation_dist_col}{distance to use in seriation_method_col (if needed)}

\item{tax_transform_numbers}{transformation applied to otu_table used only for any numbers printed}

\item{tax_scale_numbers}{scaling applied to numbers otu_table after transformation}

\item{gridlines}{list output of heat_grid() for setting gridline style}
}
\description{
Heatmap made with ComplexHeatmap, with optional (but default) annotation of taxa prevalence and abundance.

Transform your data with tax_transform prior to plotting (and/or scale with tax_scale).
Plotting "compositional" data can give an idea of the dominant taxa in each sample.
Plotting some form of log or clr transformed (or scaled) microbial features can highlight
other patterns.
The transformed data will be ordered via your selected seriation methods and distances.
Any cell numbers printed can be set to different values, and do not affect ordering.
}
\examples{
library(dplyr)
data("dietswap", package = "microbiome")
# create a couple of numerical variables to use
psq <- dietswap \%>\%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0),
    african = if_else(nationality == "AFR", true = 1, false = 0)
  )
psq <- tax_filter(psq, min_prevalence = 1 / 10, min_sample_abundance = 1 / 10)
psq <- tax_agg(psq, "Genus")

# randomly select 20 taxa from the 40 top taxa, and 40 random samples

set.seed(123)
taxa <- sample(tax_top(psq, n = 40), size = 20)
samples <- sample(1:122, size = 40)

comp_heatmap(data = psq, taxa = taxa, samples = samples)

# transforming taxon abundances #

# NOTE: if you plan on transforming taxa (e.g. to compositional data or clr)
# but only want to plot a subset of the taxa (e.g. most abundant)
# you should NOT subset the original phyloseq before transformation!
# Instead, choose the subset of taxa plotted with:

# Note 2, choose a symmetrical palette for clr-transformed data
psq \%>\%
  tax_transform("clr", zero_replace = "halfmin") \%>\%
  comp_heatmap(
    taxa = taxa, samples = samples, colors = heat_palette(sym = TRUE)
  )

# Almost all the taxa have high values (>> 0) because they are a highly
# abundant subset taken after clr transformation was calculated on all taxa

# See how just taking the first 30 taxa from the dataset gives more balance
psq \%>\%
  tax_transform("clr", zero_replace = "halfmin") \%>\%
  comp_heatmap(
    taxa = 1:30, samples = samples, colors = heat_palette(sym = TRUE)
  )

# annotating taxa #

# Notes:
# - Unlike cor_heatmap, taxa are not annotated by default
# - Detection threshold set to 50 as HITchip example data seems to have background noise

comp_heatmap(
  data = psq, taxa = taxa, samples = samples,
  tax_anno = taxAnnotation(Prev = anno_tax_prev(undetected = 50))
)

# annotating samples #

htmp <- psq \%>\%
  tax_transform("clr", zero_replace = "halfmin") \%>\%
  comp_heatmap(
    taxa = taxa, samples = samples, colors = heat_palette(sym = TRUE),
    sample_anno = sampleAnnotation(
      Nation. = anno_sample_cat("nationality", legend_title = "Nation.")
    )
  )
htmp

# legends from `anno_sample_cat()` are stored as an attribute of the Heatmap
ComplexHeatmap::draw(
  object = htmp,
  annotation_legend_list = attr(htmp, "AnnoLegends"), merge_legends = TRUE
)
}
