% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Microbe-to-sample-data correlation heatmap}
\usage{
cor_heatmap(
  data,
  taxa = phyloseq::taxa_names(ps_get(data)),
  anno_tax = tax_anno(),
  vars = NA,
  cor = c("pearson", "kendall", "spearman")[1],
  cor_use = "everything",
  colors = heat_palette(palette = "Green-Orange", sym = TRUE),
  numbers = heat_numbers(),
  taxa_which = "row",
  seriation_method = "OLO_ward",
  seriation_dist = "euclidean",
  seriation_method_col = seriation_method,
  seriation_dist_col = seriation_dist,
  tax_transform = "identity",
  gridlines = heat_grid(lwd = 0.5),
  ...
)
}
\arguments{
\item{data}{phyloseq or phyloseq extra}

\item{taxa}{list of taxa to include, or NA for all}

\item{anno_tax}{taxAnnotation output describing taxa distribution or NULL}

\item{vars}{selection of variable names from sample_data}

\item{cor}{name of correlation method}

\item{cor_use}{passed to cor(use = cor_use)}

\item{colors}{output of heat_palette() to set heatmap fill color scheme}

\item{numbers}{output of heat_numbers() to draw numbers on heatmap cells}

\item{taxa_which}{row or column (which represents a taxon)}

\item{seriation_method}{method to order the rows (in seriation::seriate)}

\item{seriation_dist}{distance to use in seriation_method (if needed)}

\item{seriation_method_col}{method to order the columns (in seriation::seriate)}

\item{seriation_dist_col}{distance to use in seriation_method_col (if needed)}

\item{tax_transform}{transformation applied to otu_table before correlating (and BEFORE selection of taxa)}

\item{gridlines}{list output of heat_grid() for setting gridline style}

\item{...}{extra args, for cor_heatmap passed to viz_heatmap (for heat_numbers() dots are passed to grid::gpar for grid::grid.text)}
}
\description{
Correlate microbial abundances to numeric sample_data variables
}
\examples{
library(dplyr)
data("dietswap", package = "microbiome")
# create a couple of numerical variables to use
psq <- dietswap \%>\%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0),
    african = if_else(nationality == "AFR", true = 1, false = 0)
  )
psq <- tax_filter(ps = psq, min_prevalence = 1 / 10, min_sample_abundance = 1 / 10)
psq <- tax_agg(psq, "Genus")

# randomly select 30 taxa from the 50 most abundant taxa
set.seed(123)
taxa <- sample(microbiome::top_taxa(ps_get(psq))[1:50], size = 30)

# make simple correlation heatmap with all numeric-like variables
p <- cor_heatmap(psq, taxa, anno_tax = tax_anno(undetected = 50))
# NOTE: detection threshold set to 50 as HITchip example data seems to have background noise
p

# optionally you can set the absolute width and height of the heatmap body manually
p1 <- cor_heatmap(psq, taxa,
  anno_tax = tax_anno(undetected = 50),
  width = grid::unit(5, "cm"), height = grid::unit(15, "cm")
)
p1

# you can also change the annotation sizes
p2 <- cor_heatmap(psq, taxa,
  anno_tax = tax_anno(undetected = 50, size = 45, rel_sizes = c(1, 2)),
  width = grid::unit(5, "cm"), height = grid::unit(15, "cm")
)
p2

# make the same correlation heatmap, but rotated
p3 <- cor_heatmap(
  psq, taxa,
  taxa_which = "column",
  anno_tax = tax_anno(undetected = 50, which = "column")
)
p3
}
