% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_palette.R
\name{distinct_palette}
\alias{distinct_palette}
\title{Arbitrary expansion of the "Paired" colorbrewer palette}
\usage{
distinct_palette(n = NA)
}
\arguments{
\item{n}{number of colours to return}
}
\value{
vector of colours
}
\description{
Good for generating palettes for large sets of categories,
e.g. in the comp_barplot function
}
\examples{
pal <- distinct_palette()
scales::show_col(c(pal, "lightgrey"))
}
