% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heat_map}
\alias{heat_map}
\title{Heat map of levels of variables of interest.}
\usage{
heat_map(
  data,
  variables,
  split_fct,
  normalize = TRUE,
  norm_center = c("mean", "median"),
  variable_classification = NULL,
  direction = "<",
  facet = TRUE,
  plot_title = NULL,
  plot_subtitle = NULL,
  x_lab = "observation",
  y_lab = "variable",
  fill_lab = "Z-score",
  hide_x_axis_text = FALSE,
  cust_theme = theme_micro(),
  color_scale = c("steelblue", "black", "firebrick"),
  midpoint = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{variables}{a vector with names of the variables of interest.}

\item{split_fct}{name of the splitting factor.}

\item{normalize}{logical, should the data frame variables be normalized prior
to plotting?}

\item{norm_center}{defines centering of the variable during scaling: mean
(default) or median. Ignored if \code{normalize = FALSE}.}

\item{variable_classification}{a optional two column data frame, which
defines the classification scheme of the variables, e.g. cluster assignment.
The first variable stores the variable name and the second indicated the
variable subset. If not provided, the variables are classified with by their
specificity for the data set subsets defined by \code{split_fct} with
\code{\link{classify}}.}

\item{direction}{determines direction of the comparison between the data set
subsets, see: \code{\link{classify}}. Ignored if \code{variable_classification}
is provided.}

\item{facet}{logical, should the variable classification scheme be presented
by horizontal facets of the plot?}

\item{plot_title}{plot title.}

\item{plot_subtitle}{plot_subtitle. If not provided, numbers of observations
in the data set subsets will be displayed here.}

\item{x_lab}{X axis label.}

\item{y_lab}{Y axis label.}

\item{fill_lab}{title of the fill scale.}

\item{hide_x_axis_text}{logical, hide the X axis text?}

\item{cust_theme}{a custom \code{ggplot} theme.}

\item{color_scale}{a character vector of length 3, which defines the lower,
middle and upper point of the color scale.}

\item{midpoint}{optional, the middle point of the fill scale.}

\item{...}{additional arguments passed to
\code{\link[ggplot2]{scale_fill_gradient2}}.}
}
\value{
a \code{ggplot} graphics.
}
\description{
The function draws a heat map of levels of the variables of interest in
data set subsets defined by a splitting factor.
}
