% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_numbers}
\alias{heat_numbers}
\title{Aesthetic settings for drawing numbers on heatmap tiles}
\usage{
heat_numbers(
  decimals = 0,
  fontsize = 7,
  col = "darkgrey",
  fontface = "bold",
  fmt = NULL,
  ...
)
}
\arguments{
\item{decimals}{number of decimal places to print}

\item{fontsize}{fontsize specification,}

\item{col}{colour of font}

\item{fontface}{plain, bold, italic}

\item{fmt}{NULL or number print format, see ?sprintf, overrides decimals arg if set}

\item{...}{passed to grid::gpar() for grid.text}
}
\value{
list
}
\description{
Works with comp_heatmap() and cor_heatmap().
See the help for those functions.
}
