% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_numbers}
\alias{heat_numbers}
\title{helps for drawing numbers on heatmap tiles}
\usage{
heat_numbers(what = "values", fmt = "\%.1f", fontsize = 7, ...)
}
\arguments{
\item{what}{what numbers to print (currently only "values" is supported)}

\item{fmt}{number print format}

\item{fontsize}{fontsize specification,}
}
\description{
helps for drawing numbers on heatmap tiles
}
