% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_palette}
\alias{heat_palette}
\title{Easy palettes for ComplexHeatmap}
\usage{
heat_palette(
  palette = "Greens",
  breaks = 5,
  range = NA,
  rev_palette = FALSE,
  sym = FALSE
)
}
\arguments{
\item{palette}{named palette from colorspace::hcl_palettes(type = "diverging") or a vector of colour names (same length as indicated by breaks)}

\item{breaks}{integer number of breaks, or numeric vector of values for colour scale breaks (including ends)}

\item{range}{NA to return palette generating function that takes a range, or numeric vector indicating the range, to return a palette}

\item{rev_palette}{reverse the colorspace palette}
}
\value{
circlize::colorRamp2 palette if !is.na(range), or function returning a palette when given a range
}
\description{
Pass a named diverging colorspace hcl palette to circlize::colorRamp2
}
