% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heat_palette}
\alias{heat_palette}
\title{Easy palettes for ComplexHeatmap}
\usage{
heat_palette(
  palette = ifelse(sym, "Blue-Red 3", "Rocket"),
  breaks = "auto",
  range = NA,
  sym = FALSE,
  rev = FALSE
)
}
\arguments{
\item{palette}{named palette from colorspace::hcl_palettes() diverging/sequential
or a vector of colour names/hexcodes}

\item{breaks}{number of breaks, "auto" is 11 for a named palette, or uses palette length}

\item{range}{NA to return palette generating function that takes range
or numeric vector indicating the range, to return a palette}

\item{sym}{makes palette range symmetrical around 0 if TRUE}

\item{rev}{reverse the palette?}
}
\value{
circlize::colorRamp2 palette if range = NA,
or function returning a palette when given a range
}
\description{
Pass a named colorspace hcl palette to circlize::colorRamp2.
\itemize{
\item If you do not specify a range this function returns a function and
the heatmap color palette will use the range of the data automatically
\item If you do specify a range, this returns a colour palette with that range
}
}
