% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{taxAnnotation}
\alias{taxAnnotation}
\alias{anno_prev}
\alias{anno_abund}
\title{taxAnnotation for ComplexHeatmap}
\usage{
taxAnnotation(
  data,
  taxa,
  undetected = 0,
  which = "row",
  prev = TRUE,
  abun = TRUE,
  ...
)

anno_prev(data, taxa, undetected = 0, which = "row", ...)

anno_abund(data, taxa, undetected = 0, which = "row", ...)
}
\arguments{
\item{data}{phyloseq or ps-extra}

\item{taxa}{names of taxa}

\item{undetected}{value above which taxa are considered present/detected in a sample}

\item{which}{'row' or 'column' annotation}

\item{prev}{if TRUE, prevalence annotation plot drawn with anno_prev}

\item{abun}{if TRUE, abundance annotation plot drawn with anno_abund}

\item{...}{args passed to anno_barplot}
}
\value{
a ComplexHeatmap anno_barplot
}
\description{
taxAnnotation is used for cor_heatmap, comp_heatmap & tax_model_heatmap (will be)
}
