% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{identify_significant}
\alias{identify_significant}
\title{Identify significant effects.}
\usage{
identify_significant(
  data,
  label_variable,
  p_variable,
  regulation_variable = NULL,
  signif_level = 0.05,
  regulation_level = 0,
  return_data = FALSE
)
}
\arguments{
\item{data}{a data frame with the testing results.}

\item{label_variable}{name of the variable storing names of the tested
variables.}

\item{p_variable}{name of the variable storing p values.}

\item{regulation_variable}{name of the variable storing regulation or
effect size figures. If \code{NULL}, significance is determined solely by the
p value threshold.}

\item{signif_level}{p value threshold.}

\item{regulation_level}{regulation or effect size threshold.}

\item{return_data}{if \code{TRUE}, the input data frame is returned with additional
columns indicating significance and regulation sign.}
}
\value{
If \code{return_data = FALSE}: a character vector or a list of character vectors
with names of significant variables derived from \code{label_variable}.
Otherwise, the function returns the input data frame with the following
addtional columns:
\itemize{
\item \code{.significant}: indicates if the effect reaches statistical significance
\item \code{.regulation}: indicates the sign of significant regulation
}
}
\description{
Identifies significant effects with a given significance and regulation
or effect size threshold.
}
