% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_explore.R
\name{ord_explore}
\alias{ord_explore}
\title{Interactively explore microbial compositions of ordinated samples}
\usage{
ord_explore(
  data,
  sample_id = NULL,
  seriate_method = "OLO_ward",
  tax_transform_for_ordering = "identity",
  app_options = list(launch.browser = TRUE),
  plot_widths = c(7, 9),
  ...
)
}
\arguments{
\item{data}{ps_extra list output of ord_calc, or phyloseq}

\item{sample_id}{name of sample ID variable to use as default for selecting samples}

\item{seriate_method}{seriation method to order phyloseq samples by similarity}

\item{tax_transform_for_ordering}{transform taxa before ordering with ps_seriate}

\item{app_options}{passed to shinyApp() options argument}

\item{plot_widths}{widths of plots in inches, including any legends
(first number is ordination, second is composition barplot)}

\item{...}{additional arguments passed to ord_plot}
}
\value{
nothing, opens default browser
}
\description{
A Shiny app used like an interactive version of \code{ord_plot()}.
You can select samples on an ordination plot and view their composition with stacked barplots.

You can give the \code{ord_explore()} data argument either of the following:
\itemize{
\item the output of \code{ord_calc()} (i.e. a ps_extra with an ordination)
\item a plain phyloseq object: \code{ord_explore()} will help you build an ordination
}

Once running:
\enumerate{
\item edit the ordination if required
\item style the ordination plot (e.g. choose dimensions; set colour and size; ...)
\item click or use the lasso tool to select 2 or more samples to view their compositions
\item style the taxonomic compositions barplot
}
}
\details{
If you get an interactive error like the one below:

"ids don't have the same length than str (most often, it occurs because of clipping)"
\enumerate{
\item make your points smaller
\item set point shape to fixed, or to a variable with fewer categories
}

Limitation: When a selection grouping variable is NA for some samples,
that grouping variable cannot be used to select those samples
}
\examples{
library(phyloseq)
library(dplyr)

# example of quickstart approach with interactive ordination calculation #
if (interactive()) {
  corncob::ibd_phylo \%>\%
    tax_fix() \%>\%
    ord_explore()
}

# simple example with precalculated ordination #
if (interactive()) {
  data("enterotype")
  taxa_names(enterotype)[1] <- "unclassified" # replaces the "-1" taxon name
  ps <- tax_fix(enterotype) # remove NA taxa
  ord1 <- ps \%>\%
    tax_transform("identity", rank = "Genus") \%>\%
    dist_calc("bray") \%>\%
    ord_calc(method = "PCoA")

  ord_explore(data = ord1, auto_caption = 6)
}

# constrained ordination example #
if (interactive()) {
  data("dietswap", package = "microbiome")

  # create a couple of numerical variables to use as constraints
  dietswap <- dietswap \%>\%
    ps_mutate(
      weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
      female = if_else(sex == "female", true = 1, false = 0)
    ) \%>\%
    tax_agg("Genus")

  constrained_aitchison_rda <- dietswap \%>\%
    tax_transform("clr") \%>\%
    ord_calc(constraints = c("weight", "female"))

  constrained_aitchison_rda \%>\%
    ord_explore(
      plot_taxa = 1:5, tax_lab_style = list(size = 3),
      constraint_lab_style = list(size = 4), auto_caption = 7
    )
}
# try changing the point colour to bmi_group or similar
# (style interactively! e.g. colour doesn't work as argument to ord_explore)

# another dataset, where "size" variable drives gradient on PC1
# (try setting size and/or alpha to correspond to "size"!)
# then edit the ordination to use "size" as a condition, see what happens
if (interactive()) {
  microbiomeutilities::hmp2 \%>\%
    tax_fix() \%>\%
    tax_transform(rank = "Genus", "identity") \%>\%
    dist_calc("aitchison") \%>\%
    ord_calc() \%>\%
    ord_explore()
}

# another dataset
if (interactive()) {
  data("soilrep", package = "phyloseq")
  # test auto creation of SAMPLE var
  ps <- soilrep \%>\% ps_select(-Sample)
  # The barplot is actually quite useless with the 16000+ anonymous OTUs
  # in this dataset, but the 1000s of unmerged "other" categories do render
  phyloseq_validate(ps) \%>\%
    tax_fix() \%>\%
    dist_calc("aitchison") \%>\%
    ord_calc() \%>\%
    ord_explore()
}
}
