% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_explore.R
\name{ord_explore}
\alias{ord_explore}
\title{Interactively explore compositions of ordinated samples}
\usage{
ord_explore(
  data,
  sample_id = NULL,
  seriate_method = "OLO_ward",
  tax_transform_for_ordering = "identity",
  app_options = list(launch.browser = TRUE),
  ...
)
}
\arguments{
\item{data}{ps_extra list output of ord_calc}

\item{sample_id}{name of id variable for ordering}

\item{seriate_method}{seriation method to order phyloseq samples by similarity}

\item{tax_transform_for_ordering}{transform tax before ordering with ps_seriate}

\item{app_options}{passed to shinyApp() options argument}

\item{...}{additional arguments passed to ord_plot}
}
\value{
nothing, opens default browser
}
\description{
Once running: click and drag to draw box over 2 or more samples to view their compositions.
You can style the ordination plot points using the options on the left panel.
}
\details{
If you get an error like the one below:

"ids don't have the same length than str (most often, it occurs because of clipping)"
\enumerate{
\item make your points smaller
\item set point shape to fixed, or to a variable with fewer categories
}

Limitation: When a selection grouping variable is NA for some samples,
that grouping variable cannot be used to select those samples
}
\examples{
library(phyloseq)
library(dplyr)
data("enterotype")

# simple example #
taxa_names(enterotype)[1] <- "unclassified"
ps <- tax_fill_unknowns(enterotype) # remove NA taxa
ord1 <- ps \%>\%
  dist_calc("bray") \%>\%
  ord_calc(method = "PCoA")

# ord_explore(
#   data = ord1, auto_caption = NA,
#   sample_id = "Sample_ID", ord_tooltip = "Sample_ID"
# )

# constrained biplot example #
data("dietswap", package = "microbiome")

# create a couple of numerical variables to use as constraints
dietswap <- dietswap \%>\%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0)
  ) \%>\%
  tax_agg("Genus")

constrained_aitchison_rda <- dietswap \%>\%
  tax_transform("clr") \%>\%
  ord_calc(constraints = c("weight", "female"))

# note if you want to visualise an ordination that required transformation
# you must take care to provide an untransformed phyloseq to ps,
# otherwise the compositions plotted will be also transformed!
# constrained_aitchison_rda \%>\%
#   ord_explore(
#     plot_taxa = 1:5, tax_lab_style = list(size = 3),
#     constraint_lab_style = list(size = 4), auto_caption = 7
#   )
# try changing the point colour to bmi_group or similar
# (style interactively! e.g. colour doesn't work as argument to ord_explore)

# another dataset
# microbiomeutilities::hmp2 \%>\%
#   tax_fill_unknowns() \%>\%
#   dist_calc("aitchison") \%>\%
#   ord_calc() \%>\%
#   ord_explore()

data("soilrep", package = "phyloseq")
# test auto creation of SAMPLE var
ps <- soilrep \%>\% ps_select(-Sample)
# The barplot is actually quite useless with the 16000+ anonymous OTUs
# in this dataset, but the 1000s of unmerged "other" categories do render
# phyloseq_validate(ps) \%>\%
#   tax_fill_unknowns() \%>\%
#   dist_calc("aitchison") \%>\%
#   ord_calc() \%>\%
#   ord_explore()
}
