% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_plot.R
\name{ord_plot}
\alias{ord_plot}
\title{Customisable ggplot of ord_calc ordination result}
\usage{
ord_plot(
  data,
  axes = 1:2,
  constraint_vec_length = 1,
  constraint_vec_style = list(),
  constraint_lab_length = constraint_vec_length * 1.1,
  constraint_lab_style = list(),
  var_renamer = function(x) identity(x),
  plot_taxa = FALSE,
  tax_vec_length = 1,
  tax_vec_style_all = list(),
  tax_vec_style_sel = list(),
  tax_lab_length = tax_vec_length * 1.1,
  tax_lab_style = list(),
  taxon_renamer = function(x) identity(x),
  plot_samples = TRUE,
  auto_caption = TRUE,
  ...
)
}
\arguments{
\item{data}{list object output from ord_calc}

\item{axes}{which axes to plot: numerical vector of length 2}

\item{constraint_vec_length}{relative length of line segment drawn for any constraints (relative to default length which is defined by correlation with each drawn axis)}

\item{constraint_vec_style}{list of aesthetics/arguments (colour, alpha etc) for the constraint vectors}

\item{constraint_lab_length}{relative length of label drawn for any constraints (relative to default position which is defined by correlation with each drawn axis)}

\item{constraint_lab_style}{list of aesthetics/arguments (colour, size etc) for the constraint labels}

\item{var_renamer}{function to rename constraining variables for plotting their labels}

\item{plot_taxa}{if ord_calc method was "PCA/RDA" draw the taxa loading vectors (see details)}

\item{tax_vec_length}{length multiplier of line segment drawn for any taxa}

\item{tax_vec_style_all}{list of named aesthetic attributes for all (background) taxon vectors}

\item{tax_vec_style_sel}{list of named aesthetic attributes for taxon vectors for the taxa selected by plot_taxa}

\item{tax_lab_length}{multiplier for label distance/position for any selected taxa}

\item{tax_lab_style}{list of fixed aesthetics (colour, size etc) for the taxon labels}

\item{taxon_renamer}{function that takes any plotted taxon names and returns modified names for labels}

\item{plot_samples}{if TRUE, plot sample points with geom_point}

\item{auto_caption}{if TRUE, add a small font caption with info about the ordination}

\item{...}{pass aesthetics arguments for sample points, drawn with geom_point using aes_string}
}
\value{
ggplot
}
\description{
Ordination visualisation. Utilises results of \code{ord_calc}.
\itemize{
\item For interpretation see the the relevant pages on PCA, PCoA, RDA, or CCA on the "gusta me" website: \url{https://sites.google.com/site/mb3gustame/}
\item Some other inspiration was from here: \url{https://ourcodingclub.github.io/2018/05/04/ordination.html}
}
}
\details{
How to specify the plot_taxa argument:
\itemize{
\item FALSE --> plot no taxa vectors or labels
\item integer vector e.g. 1:3 --> plot labels for top 3 taxa (by longest line length)
\item single numeric value e.g. 0.75 --> plot labels for taxa with line length > 0.75
\item character vector e.g. c('g__Bacteroides', 'g__Veillonella') --> plot labels for the exactly named taxa
}
}
\examples{
library(phyloseq)
library(dplyr)
library(vegan)
library(microbiome)
data("dietswap", package = "microbiome")

# create a couple of numerical variables to use as constraints or conditions
dietswap <- dietswap \%>\%
  ps_mutate(
    weight = recode(bmi_group, obese = 3, overweight = 2, lean = 1),
    female = if_else(sex == "female", true = 1, false = 0)
  )

# compute and plot ordinations for demonstration of conditioning
unconstrained_aitchison_pca <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ord_calc(method = "RDA")

unconstrained_aitchison_pca \%>\%
  ord_plot(
    colour = "bmi_group",
    plot_taxa = 1:5, tax_vec_length = 0.75, tax_lab_length = 0.85,
  ) +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ord_calc(method = "RDA", conditions = "weight") \%>\%
  ord_plot(colour = "bmi_group") +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

constrained_aitchison_rda <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  tax_transform("clr") \%>\%
  ord_calc(method = "RDA", constraints = c("weight", "female"))

constrained_aitchison_rda \%>\%
  ord_plot(colour = "bmi_group", constraint_vec_length = 2) +
  stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))

# ggplot allows additional customisation of the resulting plot
p <- constrained_aitchison_rda \%>\%
  ord_plot(
    colour = "bmi_group", constraint_vec_length = 2,
    plot_taxa = 1:3, tax_vec_length = 5
  ) +
  lims(x = c(-5, 6), y = c(-5, 5)) +
  scale_colour_brewer(palette = "Set1")

p + stat_ellipse(aes(linetype = bmi_group, colour = bmi_group))
p + stat_density2d(aes(colour = bmi_group))

# you can rename the taxa on the labels with any function that takes and modifies a character vector
constrained_aitchison_rda \%>\%
  ord_plot(
    colour = "bmi_group", constraint_vec_length = 2,
    plot_taxa = 1:3, tax_vec_length = 5,
    taxon_renamer = function(x) stringr::str_extract(x, "^.")
  ) +
  lims(x = c(-5, 6), y = c(-5, 5)) +
  scale_colour_brewer(palette = "Set1")

# it is possible to facet only unconstrained ordination plots (with plot_taxa = FALSE)
unconstrained_aitchison_pca \%>\%
  ord_plot(color = "sex") +
  facet_wrap("sex") +
  stat_density2d(aes(colour = sex)) +
  guides(colour = FALSE)

unconstrained_aitchison_pca \%>\%
  ord_plot(color = "bmi_group", plot_samples = FALSE) +
  facet_wrap("sex") +
  stat_density2d_filled(show.legend = FALSE)
}
