% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova}
\alias{permanova}
\title{Calculate PERMANOVA after calc_dist}
\usage{
permanova(
  data,
  variables,
  interactions = NULL,
  complete_cases = TRUE,
  n_processes = 1,
  n_perms = 999,
  seed = 1,
  adonis2 = FALSE,
  verbose = TRUE,
  return = "all"
)
}
\arguments{
\item{data}{list output from calc_dist}

\item{variables}{character vector of variables to include in model}

\item{interactions}{interactions between variables, written in the style of e.g. "var_a * var_b"}

\item{complete_cases}{if TRUE, drops observations if they contain missing values}

\item{n_processes}{how many parallel processes to use? (uses parallel::makePSOCKcluster)}

\item{n_perms}{how many permutations?}

\item{seed}{sets a random number generator seed to ensure you get the same results each run}

\item{adonis2}{use adonis2 (with by = "margin") instead of adonis (better if you have interactions)}

\item{verbose}{sends messages about progress if true}

\item{return}{what parts of return list to return, defaults to all parts}
}
\value{
list containing permanova results and input objects
}
\description{
Wrapper for vegan adonis/adonis2, takes the output of calc_dist (a list containing a phyloseq object and calculated distance matrix)
}
\details{
Drops observations with missing values if complete_cases is TRUE, otherwises throws an error.
}
\examples{
library(phyloseq)
library(vegan)
data("dietswap", package = "microbiome")

# add some missings to demonstrate automated removal
sample_data(dietswap)$sex[3:6] <- NA

# compute distance
testDist <- dietswap \%>\%
  tax_agg("Genus") \%>\%
  calc_dist("bray")

PERM <- testDist \%>\%
  permanova(
    variables = c("sex", "bmi_group"),
    n_processes = 1, n_perms = 99
  )
str(PERM, max.level = 1)

PERM2 <- testDist \%>\%
  permanova(
    variables = c("sex", "bmi_group"),
    interactions = "sex * bmi_group",
    n_processes = 1, n_perms = 99,
    adonis2 = TRUE
  )
PERM2$permanova

# take the same distance matrix used for the permanova and plot an ordination
PERM2 \%>\%
  ordin8("PCoA") \%>\%
  plot_ordin8(color = "bmi_group")
# this ensures any samples dropped from the permanova for having missing values
# in the covariates are NOT included in the corresponding ordination plot
}
