% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_validate.R
\name{phyloseq_validate}
\alias{phyloseq_validate}
\title{Check for (and fix) common problems with phyloseq objects}
\usage{
phyloseq_validate(ps, min_tax_length = 4, verbose = TRUE)
}
\arguments{
\item{ps}{phyloseq object}

\item{min_tax_length}{minimum number of characters to not consider a tax_table entry suspiciously short}

\item{verbose}{print informative messages if true}
}
\value{
possibly modified phyloseq object
}
\description{
\itemize{
\item Ensures the storage mode of a phyloseq object's otu table is "double" instead of e.g. "integer".
\item Checks for, and messages about, common uninformative entries in the tax_table, which often cause unwanted results
\item Replaces missing sample_data with a dataframe including only sample_names (as "SAMPLE" variable)
\item Removes taxa where phyloseq::taxa_sums() is equal to zero, with a warning
}
}
\examples{
data(dietswap, package = "microbiome")

tax_filter(dietswap, )
phyloseq_validate(dietswap, verbose = TRUE) # expect warning about taxa summing to zero

# verbose = FALSE will suppress messages and warnings but still:
# fix storage.mode, replace NULL sample_data, and remove taxa that sum to 0 across all samples
phyloseq_validate(dietswap, verbose = FALSE)

# no messages or warnings means no problems detected
}
