% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_dedupe.R
\name{ps_dedupe}
\alias{ps_dedupe}
\title{De-duplicate phyloseq samples}
\usage{
ps_dedupe(ps, vars, method = "readcount", verbose = TRUE)
}
\arguments{
\item{ps}{phyloseq with sample data}

\item{vars}{names of variables, whose (combined) levels identify groups from which only 1 sample is desired}

\item{method}{keep 1 sample with max "readcount" or the "first" or "last" samples encountered in given sample_data order for each dupe group}

\item{verbose}{message names of samples removed if TRUE (or "debug" for more info, when method = "readcount")}
}
\value{
phyloseq object
}
\description{
Use 1 or more variables in the sample_data to identify and remove duplicate samples (leaving 1 per category).

\strong{methods:}
\itemize{
\item method = "readcount" keeps the one sample in each duplicate group with the highest total number of reads according to sample_sums
\item method = "first" keeps the first sample in each duplicate group encountered in the row order of the sample_data
\item method = "last" keeps the last sample in each duplicate group encountered in the row order of the sample_data
}
}
\details{
What happens when duplicated samples have exactly equal readcounts in method = "readcount"?
The first encountered maximum is kept (in sample_data row order, like method = "first")
}
\examples{
data("dietswap", package = "microbiome")

dietswap
# let's pretend the dietswap data contains technical replicates from each subject
# we want to keep only one of them
ps_dedupe(dietswap, vars = "subject", method = "readcount", verbose = TRUE)

# contrived example to show identifying "duplicates" via the interaction of multiple columns
ps1 <- ps_dedupe(
  ps = dietswap, method = "readcount", verbose = TRUE,
  vars = c("timepoint", "group", "bmi_group")
)
phyloseq::sample_data(ps1)

ps2 <- ps_dedupe(
  ps = dietswap, method = "first", verbose = TRUE,
  vars = c("timepoint", "group", "bmi_group")
)
phyloseq::sample_data(ps2)
}
\seealso{
\code{\link{ps_filter}} for filtering samples by sample_data variables
}
