% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_drop_incomplete.R
\name{ps_drop_incomplete}
\alias{ps_drop_incomplete}
\title{Deselect phyloseq samples with sample_data missings}
\usage{
ps_drop_incomplete(ps, vars = NA, verbose = FALSE)
}
\arguments{
\item{ps}{phyloseq with sample_data}

\item{vars}{vector of variable names to check for missings (or NA, which uses all variables in sample data)}

\item{verbose}{message about number of samples dropped if verbose not FALSE, (and only if > 0 samples dropped)
and message about number of missing per variable in vars if verbose = "max" (and message even if 0 samples dropped)}
}
\value{
phyloseq
}
\description{
Check phyloseq object sample data for missing values (NAs)
\itemize{
\item specify which variables to check with vars argument, or check all)
\item drop samples with any missings
}
}
\details{
This is a wrapper for \code{\link{stats::complete.cases()}} function.
}
\examples{
library(phyloseq)
data("enterotype", package = "phyloseq")

enterotype
ps_drop_incomplete(enterotype)
ps_drop_incomplete(enterotype, vars = "Enterotype", verbose = TRUE)
ps_drop_incomplete(enterotype, vars = "Sample_ID", verbose = TRUE)
ps_drop_incomplete(enterotype, vars = c("Enterotype", "Sample_ID"))
ps_drop_incomplete(enterotype, verbose = "max")
}
