% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_extra-accessors.R
\name{ps_extra-accessors}
\alias{ps_extra-accessors}
\alias{ps_get}
\alias{dist_get}
\alias{ord_get}
\alias{info_get}
\alias{perm_get}
\alias{bdisp_get}
\alias{otu_get}
\title{Extract elements from ps_extra class}
\usage{
ps_get(ps_extra)

dist_get(ps_extra)

ord_get(ps_extra)

info_get(ps_extra)

perm_get(ps_extra)

bdisp_get(ps_extra)

otu_get(data)
}
\arguments{
\item{ps_extra}{ps_extra class object}

\item{data}{phyloseq or ps_extra}
}
\value{
element of ps_extra class object (or NULL)
}
\description{
\itemize{
\item \code{ps_get}    returns phyloseq
\item \code{info_get}  returns ps_extra_info object
\item \code{dist_get}  returns distance matrix (or NULL)
\item \code{ord_get}   returns ordination object (or NULL)
\item \code{perm_get}  returns adonis2() permanova model (or NULL)
\item \code{bdisp_get} returns results of betadisper() (or NULL)
}
}
\examples{
library(phyloseq)
data("esophagus")
}
