% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_extra-accessors.R
\name{ps_extra-accessors}
\alias{ps_extra-accessors}
\alias{ps_get}
\alias{dist_get}
\alias{ord_get}
\alias{info_get}
\alias{perm_get}
\alias{bdisp_get}
\alias{otu_get}
\alias{tt_get}
\alias{samdat_tbl}
\title{Extract elements from ps_extra class}
\usage{
ps_get(ps_extra)

dist_get(ps_extra)

ord_get(ps_extra)

info_get(ps_extra)

perm_get(ps_extra)

bdisp_get(ps_extra)

otu_get(data)

tt_get(data)

samdat_tbl(data, sample_names_col = ".sample_name")
}
\arguments{
\item{ps_extra}{ps_extra class object}

\item{data}{phyloseq or ps_extra}

\item{sample_names_col}{name of column where sample_names are put.
if NA, return data.frame with rownames (sample_names)}
}
\value{
element of ps_extra class object (or NULL)
}
\description{
\itemize{
\item \code{ps_get}     returns phyloseq
\item \code{info_get}   returns ps_extra_info object
\item \code{dist_get}   returns distance matrix (or NULL)
\item \code{ord_get}    returns ordination object (or NULL)
\item \code{perm_get}   returns adonis2() permanova model (or NULL)
\item \code{bdisp_get}  returns results of betadisper() (or NULL)
\item \code{otu_get}    returns phyloseq otu_table matrix with taxa as columns
\item \code{tt_get}     returns phyloseq tax_table
\item \code{samdat_tbl} returns phyloseq sample_data as a tibble, with sample_names as new first column called .sample_name
}
}
\examples{
data("dietswap", package = "microbiome")
psx <- tax_transform(dietswap, "identity", rank = "Genus")
psx

ps_get(psx)
info_get(psx)

dist_get(psx) # this ps_extra has no dist_calc result
ord_get(psx) # this ps_extra has no ord_calc result
perm_get(psx) # this ps_extra has no dist_permanova result
bdisp_get(psx) # this ps_extra has no dist_bdisp result

# these can be returned from phyloseq objects too
otu_get(psx)[1:6, 1:4]
tt_get(psx) \%>\% head()
samdat_tbl(psx) \%>\% head()
}
