% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_filter.R
\name{ps_filter}
\alias{ps_filter}
\title{Filter phyloseq samples by sample_data variables}
\usage{
ps_filter(ps, ..., .target = "sample_data", .keep_all_taxa = FALSE)
}
\arguments{
\item{ps}{phyloseq object}

\item{...}{passed directly to dplyr::filter (see examples and ?dplyr::filter)}

\item{.target}{which slot of phyloseq to use for filtering by,
currently only "sample_data" supported}

\item{.keep_all_taxa}{if FALSE (the default),
remove taxa which are no longer present in the dataset after filtering}
}
\value{
phyloseq object (with filtered sample_data)
}
\description{
Keep only samples with sample_data matching one or more conditions.
Use this function as you would use use dplyr::filter(), but with a phyloseq object!
}
\examples{
library(phyloseq)
library(dplyr)

data("enterotype", package = "phyloseq")
enterotype
sample_data(enterotype)[1:10, 1:5]

# keep only samples with seqtech not equal to sanger
ps1 <- ps_filter(enterotype, SeqTech != "Sanger")
ps1
sample_data(ps1)[1:10, 1:5]

# keep only samples with no NAs in any variables
ps2 <- enterotype \%>\% ps_filter(!if_any(everything(), is.na))
ps2
sample_data(ps2)[1:8, 1:8]

# ps2 is equivalent to dropping samples with incomplete sample_variables and tax_filtering 0s
ps3 <- enterotype \%>\%
  ps_drop_incomplete() \%>\%
  tax_filter(undetected = 0, use_counts = FALSE)
# we needed to set a low detection threshold because this example data is proportions
identical(ps2, ps3) # TRUE

# function will give warning if some of the otu_values are negative
# (which may happen when filtering data that has e.g. clr-transformed taxa abundances)
# as it attempts to discard any taxa that become always absent/0 after filtering (by default)
# set .keep_all_taxa = TRUE to avoid this filtering behaviour, which is unwanted in this case
enterotype \%>\%
  tax_transform("clr") \%>\%
  ps_get() \%>\%
  ps_filter(SeqTech == "Sanger", .keep_all_taxa = TRUE)
}
\seealso{
\code{\link[dplyr]{filter}} explains better how to give arguments to this function

\code{\link{tax_filter}} for filtering taxa (not samples)
}
