% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_mutate.R
\name{ps_mutate}
\alias{ps_mutate}
\title{Modify or compute new sample_data in phyloseq object}
\usage{
ps_mutate(ps, ...)
}
\arguments{
\item{ps}{phyloseq object with sample data}

\item{...}{passed straight to dplyr::mutate (see examples and dplyr::mutate help)}
}
\value{
phyloseq object with modified sample_data
}
\description{
Add or compute new phyloseq sample_data variables.
Uses \code{dplyr::mutate()} syntax.
}
\examples{
library(phyloseq)
library(dplyr)
data("enterotype")

sample_data(enterotype)[1:10, ]

months_in_year <- 12
ps <- enterotype \%>\%
  ps_mutate(
    age_months = Age * months_in_year,
    IDs_match = as.character(Sample_ID) == as.character(SampleID),
    placeholder = "Word"
  )

sample_data(ps)[1:10, ]

# Using the dplyr::across functionality is also possible
ps <- ps \%>\%
  ps_mutate(
    dplyr::across(where(is.factor), toupper),
    another_var = TRUE,
    SeqTech = NULL # deletes SeqTech variable
  )

head(sample_data(ps))
}
\seealso{
\code{\link{tax_mutate}} for manipulation of tax_table variables

\code{\link[dplyr]{mutate}}
}
