% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_mutate.R
\name{ps_mutate}
\alias{ps_mutate}
\title{Mutate sample data in phyloseq object}
\usage{
ps_mutate(ps, .target = "sample_data", .across = NULL, ...)
}
\arguments{
\item{ps}{phyloseq object with sample data}

\item{.target}{which slot to mutate, currently only "sample_data" supported}

\item{.across}{placeholder for future implementation of working dplyr::across functionality}

\item{...}{passed straight to dplyr::mutate (see examples and dplyr::mutate help)}
}
\value{
phyloseq object with modified sample_data
}
\description{
Wrapper for dplyr::mutate. Use mutate syntax with all the non-standard evaluation.
}
\examples{
library(phyloseq)
library(dplyr)
data("enterotype")

sample_data(enterotype)[1:10, ]

months_in_year <- 12
ps <- enterotype \%>\%
  ps_mutate(
    age_months = Age * months_in_year,
    IDs_match = as.character(Sample_ID) == as.character(SampleID),
    placeholder = "Word"
  )

sample_data(ps)[1:10, ]

# Using the dplyr::across functionality is a little different,
# but is possible using the dedicated .across argument.
# Further expressions can be passed to dots at the same time as using .across.
  ps <- ps \%>\%
  ps_mutate(
    .across = dplyr::across(where(is.factor), toupper),
    dots_still_work = TRUE
  )

sample_data(ps)[1:10, ]

}
