% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_reorder.R
\name{ps_reorder}
\alias{ps_reorder}
\title{Set order of samples in phyloseq object}
\usage{
ps_reorder(ps, sample_order)
}
\arguments{
\item{ps}{phyloseq}

\item{sample_order}{names or current numerical indices of samples in desired order}
}
\value{
phyloseq
}
\description{
Ordering of samples is a phyloseq is controlled from the otu_table slot!
}
\examples{
library(phyloseq)
library(microbiome)
data("dietswap")

dietswap \%>\%
  sample_data() \%>\%
  head(8)

new_order <- rev(sample_names(dietswap))
dietswap \%>\%
  ps_reorder(new_order) \%>\%
  sample_data() \%>\%
  head(8)

# random ordering with numbers
set.seed(1000)
random_order <- sample(1:nsamples(dietswap))
dietswap \%>\%
  ps_reorder(random_order) \%>\%
  sample_data() \%>\%
  head(8)

}
