% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_agg.R
\name{tax_agg}
\alias{tax_agg}
\title{Aggregate taxa if requested}
\usage{
tax_agg(ps, agg_level)
}
\arguments{
\item{ps}{phyloseq object}

\item{agg_level}{"none", "asv", "given" or name of valid taxonomic rank (try phyloseq::rank_names(ps))}
}
\value{
list including phyloseq object and (tax_)level argument value
}
\description{
This function is a simple wrapper around microbiome::aggregate_taxa.
tax_agg allows you to pass 'none', 'asv', or 'given', to the level argument which will NOT aggregate the taxa.
If you pass any valid rank name it will aggregate taxa with microbiome::aggregate_taxa.
}
\examples{
library(microbiome)
data("dietswap", package = "microbiome")
tax_agg(ps = dietswap, agg_level = "none")
tax_agg(ps = dietswap, agg_level = "Phylum")
}
