% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_fill_unknowns.R
\name{tax_fill_unknowns}
\alias{tax_fill_unknowns}
\title{Replace unknown, NA, or short tax_table values}
\usage{
tax_fill_unknowns(
  ps,
  min_length = 4,
  unknowns = NA,
  levels = phyloseq::rank_names(ps),
  suffix_rank = "classified",
  sep = " ",
  anon_unique = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ps}{phyloseq or tax_table (taxonomyTable) with at least two ranks}

\item{min_length}{replace strings shorter than this}

\item{unknowns}{also replace strings matching any in this vector, NA default vector shown in details!}

\item{levels}{names of taxonomic levels to modify, defaults to all}

\item{suffix_rank}{"classified" (default) or "current", when replacing an entry, should the suffix be taken from the lowest classified rank for that taxon "classified", or the "current" unclassified rank?}

\item{sep}{character(s) separating new name and taxonomic rank level suffix (see suffix_rank)}

\item{anon_unique}{make anonymous taxa unique by replacing unknowns with taxa_name?
otherwise they are replaced with paste("unknown", rank_names(ps)\link{1}),
which is therefore the same for every anonymous taxon, meaning they will be merged if tax_agg is used.
(anonymous taxa are taxa with all unknown values in their tax_table row, i.e. cannot be classified even at highest rank available)}

\item{verbose}{emit warnings when cannot replace with informative name?}
}
\value{
object same class as ps
}
\description{
Identifies phyloseq tax_table values as unknown or uninformative and replaces them with the first informative value from a higher taxonomic rank.
\itemize{
\item Short values in phyloseq tax_table are typically empty strings or " ", or "g__" etc. so it is helpful to replace them. Set \code{min_length} = 0 to avoid filtering on length.
\item Values in \code{unknowns} are also removed, even if longer than \code{min_length}. It is up to the user to specify sensible values in \code{unknowns} if their dataset has other unwanted values.
\item NA values are also replaced.
}
}
\details{
By default (unknowns = NA), unknowns is set to a vector containing:
's__' 'g__' 'f__' 'o__' 'c__' 'p__' 'k__' 'S__' 'G__' 'F__' 'O__' 'C__' 'P__' 'K__' 'NA' 'NaN' ' ' ''
'unknown' 'Unknown' 's__unknown' 's__Unknown' 's__NA' 'g__unknown' 'g__Unknown' 'g__NA'
'f__unknown' 'f__Unknown' 'f__NA' 'o__unknown' 'o__Unknown' 'o__NA' 'c__unknown' 'c__Unknown' 'c__NA'
'p__unknown' 'p__Unknown' 'p__NA' 'k__unknown' 'k__Unknown' 'k__NA' 'S__unknown' 'S__Unknown' 'S__NA'
'G__unknown' 'G__Unknown' 'G__NA' 'F__unknown' 'F__Unknown' 'F__NA' 'O__unknown' 'O__Unknown' 'O__NA'
'C__unknown' 'C__Unknown' 'C__NA' 'P__unknown' 'P__Unknown' 'P__NA' 'K__unknown' 'K__Unknown' 'K__NA'
}
\examples{
library(dplyr)
library(tibble)
library(phyloseq)
library(microbiome)

data(dietswap)
ps <- dietswap

# create unknowns to test filling
tt <- tax_table(ps)
ntax <- ntaxa(ps)
set.seed(123)
g <- sample(1:ntax, 30)
f <- sample(g, 10)
p <- sample(f, 3)
tt[g, 3] <- "g__"
tt[f, 2] <- "f__"
tt[p, 1] <- "p__"
tt[sample(1:ntax, 10), 3] <- "unknown"
# create a row with only NAs
tt[1, ] <- NA

tax_table(ps) <- tax_table(tt)

ps
tax_table(ps) \%>\% head(50)
# tax_fill_unknowns with defaults should solve most problems

# this will replace `unknown`s as well as short values including "g__" and "f__"
tax_fill_unknowns(ps) \%>\%
  tax_table() \%>\%
  head(50)

# This will only replace values in Genus column,
# and only replace short entries, and so won't replace literal "unknown"
# WARNING:
# only set `levels` arg if you know you won't create an invalid tax_table!
ps \%>\%
  tax_fill_unknowns(unknowns = NULL, levels = "Genus") \%>\%
  tax_table() \%>\%
  head(50)

# Change rank suffix and separator settings
tax_fill_unknowns(ps, suffix_rank = "current", sep = " - ") \%>\%
  tax_table() \%>\%
  head(50)

# larger example tax_table shows 1000s rows still fast, from microbiomeutilities package
# library(microbiomeutilities)
# data("hmp2")
# tax_fill_unknowns(hmp2, min_length = 1)
}
