% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_sort.R
\name{tax_sort}
\alias{tax_sort}
\title{Sort taxa in phyloseq otu_table and tax_table}
\usage{
tax_sort(data, by = "name", ...)
}
\arguments{
\item{data}{ps_extra or phyloseq}

\item{by}{how to sort, see description}

\item{...}{used if summary function given, or pass \code{undetected} arg for tax_transform("binary") if by = "prev" or "prevalence"}
}
\value{
sorted phyloseq or ps_extra
}
\description{
Multiple ways of sorting taxa are possible and determined by the by argument.
\code{by} argument must be one of:
- 'rev' to reverse the current order
- 'name' or 'names' (sort alphabetically by taxa_names)
- a taxonomic rank name (sort alphabetically by this rank)
- a sample name (descending abundance sorting within that sample)
- summary stat. function e.g. \code{sum} or \code{mean}
}
\details{
Don't forget to pass \code{na.rm} to \code{...} if using a summary stat function in by
}
\examples{
library(microbiome)
data("dietswap")
dietswap

# reverse current order
dietswap \%>\%
  tax_sort("rev") \%>\%
  tax_table() \%>\%
  head(30)

# sort alphabetically by a taxonomic rank (or "names" for taxa_names)
dietswap \%>\%
  tax_sort("Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# sequentially sorting by higher ranks
# sets tax_table in nested alphabetical order
dietswap \%>\%
  tax_sort("names") \%>\%
  tax_sort("Genus") \%>\%
  tax_sort("Family") \%>\%
  tax_sort("Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# order by descending abundance in a single named sample
dietswap \%>\%
  tax_sort("Sample-1") \%>\%
  otu_table() \%>\%
  .[1:5, 1:4]

# sort by function e.g. median abundance
dietswap \%>\%
  tax_sort(by = median) \%>\%
  taxa_names() \%>\%
  head(20)

# sum order should always equal mean order if non-negative abundances
# don't forget to add na.rm = TRUE if you expect NAs in otu_table somehow
dietswap \%>\%
  tax_sort(by = sum, na.rm = TRUE) \%>\%
  taxa_names() \%>\%
  head(20)
}
