% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_sort.R
\name{tax_sort}
\alias{tax_sort}
\title{Sort taxa in phyloseq otu_table and tax_table}
\usage{
tax_sort(
  data,
  by = "name",
  at = "names",
  ...,
  tree_warn = TRUE,
  verbose = TRUE,
  trans = "identity",
  use_counts = TRUE
)
}
\arguments{
\item{data}{psExtra or phyloseq}

\item{by}{how to sort, see description}

\item{at}{"names" or a taxonomic rank to apply sorting method to, as specified in \code{by}.}

\item{...}{used if summary function given, or pass \code{undetected} arg
for tax_transform("binary") if by = "prev" or "prevalence"}

\item{tree_warn}{If phylogenetic tree is present in phyloseq phy_tree slot, taxa cannot be reordered.
Default behaviour of tax_sort is to remove the phylogenetic tree and warn about this.
tree_warn = FALSE will suppress the warning message, but still remove the tree!}

\item{verbose}{passed to phyloseq_validate verbose
(if TRUE: message about suspicious values in tax_table, and how to fix)}

\item{trans}{name of transformation to apply to taxa before sorting (taxa are returned un-transformed)}

\item{use_counts}{use count data if available, instead of transformed data}
}
\value{
sorted phyloseq or psExtra
}
\description{
Multiple ways of sorting taxa are possible and determined by the \code{by} argument.
The \code{by} argument must be one of:
\itemize{
\item 'rev' to reverse the current order
\item 'name' (sort alphabetically by \code{at})
\item 'asis' to keep current order as is
\item a sample name (descending abundance sorting within that sample)
\item summary stat. function e.g. \code{sum} or \code{mean}
}

The \code{at} argument must be "names" for sorting unique taxa,
or a rank name, for sorting at that rank.
\code{at} is ignored when \code{by} is "rev".
}
\details{
Don't forget to pass \code{na.rm = TRUE} to \code{...}
if using a summary stat function in \code{by}
}
\examples{
library(phyloseq)
data("dietswap", package = "microbiome")
dietswap

# reverse current order
dietswap \%>\%
  tax_sort("rev") \%>\%
  tax_table() \%>\%
  head(30)

# sort alphabetically by a taxonomic rank (or "names" for taxa_names)
dietswap \%>\%
  tax_sort(by = "name", at = "Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# sequentially sorting by higher ranks
# sets tax_table in nested alphabetical order
dietswap \%>\%
  tax_sort(at = "names") \%>\%
  tax_sort(at = "Genus") \%>\%
  tax_sort(at = "Family") \%>\%
  tax_sort(at = "Phylum") \%>\%
  tax_table() \%>\%
  head(30)

# sort by function e.g. total sum or median abundance
dietswap \%>\%
  tax_sort(by = sum) \%>\%
  taxa_names() \%>\%
  head(20)

# transform to compositional data (proportions) before sorting
# note that abundances are returned untransformed
dietswap \%>\%
  tax_sort(by = sum, trans = "compositional") \%>\%
  taxa_names() \%>\%
  head(20)

# order by descending abundance in a single named sample
dietswap \%>\%
  tax_sort(by = "Sample-1") \%>\%
  otu_table() \%>\%
  .[1:8, 1:4]


# sum order should always equal mean order if non-negative abundances
# don't forget to add na.rm = TRUE if you expect NAs in otu_table somehow
dietswap \%>\%
  tax_sort(by = sum, na.rm = TRUE) \%>\%
  taxa_names() \%>\%
  head(20)

# if your phyloseq object has a phylogenetic tree,
# tax_sort will remove the tree, and warn you about this
# unless you disable that warning with tree_warn = FALSE

# You can sort by abundance at higher taxonomic ranks,
# without losing lower rank info
# e.g. sort (descending) by phyla abundances
dietswap \%>\%
  tax_sort(by = sum, at = "Phylum") \%>\%
  tax_table() \%>\%
  head()

# You can sort by ascending abundance (or prevalence etc) by reversing after
dietswap \%>\%
  tax_sort(by = "prev", at = "Phylum") \%>\%
  tax_sort(by = "rev") \%>\%
  tax_table() \%>\%
  head()
}
