% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_top.R
\name{tax_top}
\alias{tax_top}
\title{Print names of "top" n taxa}
\usage{
tax_top(data, n = 10, by = sum, rank = "unique", ...)
}
\arguments{
\item{data}{phyloseq object or psExtra}

\item{n}{how many taxa names to return, or NA for all
(can return fewer than n values, if there are fewer to return)}

\item{by}{how to sort taxa (see \code{?tax_sort()}),
defaults to \code{sum} which sorts by total abundance across all samples}

\item{rank}{taxonomic rank to aggregate at before calculating
("unique" = no aggregation)}

\item{...}{extra optional args passed to tax_sort}
}
\value{
vector of taxa names at chosen rank
}
\description{
Simple wrapper function that:
\enumerate{
\item optionally aggregates taxa at \code{rank}
\item sorts the aggregated taxa according to \code{by}
\item returns the top \code{n} number of taxa names
}
}
\examples{
data("dietswap", package = "microbiome")
tax_top(dietswap)
tax_top(dietswap, n = 4, by = "prev", rank = "Phylum", undetected = 30)
}
\seealso{
\code{\link{tax_agg}} for more info on taxonomic aggregation

\code{\link{tax_sort}} for more info on sorting taxa
}
