% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_transform.R
\name{tax_transform}
\alias{tax_transform}
\title{Transform taxa in phyloseq object and record transformation}
\usage{
tax_transform(data, transformation, ...)
}
\arguments{
\item{data}{list output from tax_agg, or a phyloseq object}

\item{transformation}{any valid taxa transformation from microbiome::transform}

\item{...}{any extra arguments passed to microbiome::transform()}
}
\value{
list including phyloseq object and level argument value
}
\description{
Pipe the results of tax_agg into this function to transform the (aggregated) taxa features. (Or pass it a phyloseq object).
This function is a simple wrapper around microbiome::transform can perform all the same transformations but returns a list, containing ps (the transformed phyloseq object), tax_transform (a string recording the transformation), and tax_level (a string recording the taxonomic aggregation level if specified in a preceding tax_agg function call.
}
\examples{
library(microbiome)
data("dietswap", package = "microbiome")
tax_agg(ps = dietswap, agg_level = "none") \%>\% tax_transform("identity")
tax_agg(ps = dietswap, agg_level = "Phylum") \%>\% tax_transform("clr")
}
