% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_transform.R
\name{tax_transform}
\alias{tax_transform}
\title{Transform taxa in phyloseq object and record transformation}
\usage{
tax_transform(data, transformation, ...)
}
\arguments{
\item{data}{ps_extra list output from tax_agg, or a phyloseq object}

\item{transformation}{any valid taxa transformation from microbiome::transform}

\item{...}{any extra arguments passed to microbiome::transform()}
}
\value{
ps_extra list including phyloseq object and info
}
\description{
Pipe the results of tax_agg into this function to transform the (aggregated) taxa features.
This function is a simple wrapper around microbiome::transform() that can perform all the same transformations but returns a ps_extra list
containing ps (the transformed phyloseq object) and extra info: tax_transform (a string recording the transformation), and tax_agg
(a string recording the taxonomic aggregation level if specified in a preceding tax_agg function call).
}
\examples{
library(microbiome)
data("dietswap", package = "microbiome")

# aggregate taxa at Phylum level and perform the center log ratio transform on the phyla counts
tax_agg(ps = dietswap, rank = "Phylum") \%>\% tax_transform("clr")

# do nothing except record tax_agg as "unique" and tax_transform as "identity" in ps_extra info
tax_agg(ps = dietswap, rank = NA) \%>\% tax_transform("identity")
}
