% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_models.R
\name{taxatree_models}
\alias{taxatree_models}
\title{Statistical modelling for individual taxa across multiple ranks}
\usage{
taxatree_models(
  ps,
  ranks = NULL,
  type = "lm",
  variables = NULL,
  formula = NULL,
  univariable = FALSE,
  use_future = FALSE,
  checkVars = TRUE,
  checkNA = "warning",
  verbose = "rank",
  ...
)
}
\arguments{
\item{ps}{phyloseq object or ps_extra}

\item{ranks}{vector of rank names at which to aggregate taxa for modelling}

\item{type}{name of regression modelling function, or the function itself}

\item{variables}{vector of variable names, to be used as model formula right hand side
(ignored if formula given)}

\item{formula}{(alternative to variables arg) right hand side of a formula,
as a formula object or character value}

\item{univariable}{should multiple univariable models be run per taxon?
one for each variable named in variables argument}

\item{use_future}{if TRUE parallel processing with future is possible, see details of ?tax_model.}

\item{checkVars}{check variance of variables?}

\item{checkNA}{check variables for NAs?}

\item{verbose}{message about progress: "rank" only notifies which rank is being processed;
TRUE notifies you about each taxon being processed; FALSE for no messages.}

\item{...}{extra arguments are passed directly to modelling function}
}
\description{
\code{taxatree_models} runs \code{tax_model} on every taxon at multiple taxonomic
ranks (you choose which ranks with the plural \code{ranks} argument).
It returns the results as a named nested list of models
attached to a ps_extra object.
One list per rank, one model per taxon at each rank.

The result can then be used with \code{taxatree_models2stats} to extract a
dataframe of statistics for use with \code{taxatree_plots}.
}
