% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_plotkey.R
\name{taxatree_plot_labels}
\alias{taxatree_plot_labels}
\title{Add labels to taxatree plots/key}
\usage{
taxatree_plot_labels(
  p,
  circular = TRUE,
  taxon_renamer = identity,
  fun = ggrepel::geom_text_repel,
  label_var = "label",
  x_nudge = 0.1,
  y_nudge = 0.025,
  fontface = "bold",
  size = 2.5,
  colour = "grey15",
  max.overlaps = Inf,
  min.segment.length = 0,
  segment.size = 0.15,
  segment.color = "grey15",
  point.padding = 0.05,
  box.padding = 0.1,
  seed = NA,
  ...
)
}
\arguments{
\item{p}{taxatree_plotkey or taxatree_plots output plot}

\item{circular}{is the plot layout circular? labels are drawn differently for circular trees}

\item{taxon_renamer}{function that takes taxon names and returns modified names for labels}

\item{fun}{ggrepel labelling function: geom_text_repel or geom_label_repel}

\item{label_var}{name of variable in taxatree_stats that indicates which taxa to label}

\item{x_nudge}{absolute amount by which the initial position of taxon labels is nudged
(relevant only for circular layouts, use nudge_x for other layouts)}

\item{y_nudge}{absolute amount by which the initial position of taxon labels is nudged
(relevant only for circular layouts, use nudge_y for other layouts)}

\item{fontface}{fontface of label text}

\item{size}{size of labels}

\item{colour}{colour of label outlines and text}

\item{max.overlaps}{max number of overlapping labels tolerated}

\item{min.segment.length}{min length of label line segment to bother drawing}

\item{segment.size}{thickness of line segment}

\item{segment.color}{colour of line segment}

\item{point.padding}{padding around node points (for label positioning)}

\item{box.padding}{padding around labels/text (for label positioning)}

\item{seed}{set this for reproducible label positions}

\item{...}{other arguments passed to \code{fun} labelling function}
}
\description{
Finer control over label drawing for \code{taxatree_plotkey}
(with .draw_label = FALSE),
and label drawing for \code{taxatree_plots} output too.
}
