% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxatree_plotkey.R
\name{taxatree_plotkey}
\alias{taxatree_plotkey}
\title{Draw labelled key to accompany taxatree_plots}
\usage{
taxatree_plotkey(
  data,
  ...,
  size_stat = list(prevalence = prev),
  node_size_range = c(1.5, 5),
  edge_width_range = node_size_range * 0.8,
  size_guide = "none",
  size_trans = "identity",
  colour = "lightgrey",
  edge_alpha = 0.7,
  title = "Key",
  title_size = 14,
  taxon_renamer = identity,
  .combine_label = any,
  .draw_label = TRUE,
  .calc_label = TRUE,
  layout = "tree",
  layout_seed = NA,
  circular = identical(layout, "tree"),
  node_sort = NULL,
  add_circles = isTRUE(circular),
  drop_ranks = TRUE
)
}
\arguments{
\item{data}{psExtra (or phyloseq)}

\item{...}{logical conditions for labelling
e.g. rank == "Phylum", p.value < 0.1 | taxon \%in\% listOfTaxa}

\item{size_stat}{named list of length 1, giving function calculated for each taxon,
to determine the size of nodes (and edges). Name used as size legend title.}

\item{node_size_range}{min and max node sizes, decrease to avoid node overlap}

\item{edge_width_range}{min and max edge widths}

\item{size_guide}{guide for node sizes, try "none", "legend" or ggplot2::guide_legend()}

\item{size_trans}{transformation for size scale
you can use (the name of) any transformer from the scales package,
such as "identity", "log1p", or "sqrt"}

\item{colour}{fixed colour and fill of nodes and edges}

\item{edge_alpha}{fixed alpha value for edges}

\item{title}{title of plot (NULL for none)}

\item{title_size}{font size of title}

\item{taxon_renamer}{function that takes taxon names and returns modified names for labels}

\item{.combine_label}{all or any: function to combine multiple logical "label" values for a taxon
(relevant if taxatree_stats already present in data)}

\item{.draw_label}{should labels be drawn, or just the bare tree, set this to FALSE if you want
to draw customised labels with taxatree_plot_labels afterwards}

\item{.calc_label}{if you already set labels with taxatree_label:
set this to FALSE to use / avoid overwriting that data
(ignores \code{...} if FALSE)}

\item{layout}{any ggraph layout, default is "tree"}

\item{layout_seed}{any numeric, required if a stochastic igraph layout is named}

\item{circular}{should the layout be circular?}

\item{node_sort}{sort nodes by "increasing" or "decreasing" size? NULL for no sorting.
Use \code{tax_sort()} before \code{taxatree_plots()} for finer control.}

\item{add_circles}{add faint concentric circles to plot, behind each rank?}

\item{drop_ranks}{drop ranks from tree if not included in stats dataframe}
}
\value{
ggplot
}
\description{
Draw labelled key to accompany taxatree_plots
}
\examples{
# see taxatree_plots() examples
}
