% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_anova}
\alias{test_anova}
\title{Test gene expression differences between multiple groups with one-way ANOVA.}
\usage{
test_anova(
  data,
  split_fct,
  variables,
  adj_method = "BH",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[stats]{aov}}.}
}
\value{
a list of data frames containing the ANOVA and linear model
estimates, respectively.
}
\description{
Compares expression differences between multiple groups with
one-way ANOVA and linear modeling.
}
