% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_nb}
\alias{test_nb}
\title{Test gene expression differences with negative binomial modeling.}
\usage{
test_nb(
  data,
  split_fct,
  variables,
  confounder = NULL,
  adj_method = "BH",
  dev_test = "Chisq",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{confounder}{name of the confounding variable. Defaults to NULL, which
means that no confounder is include in the model.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{dev_test}{name of the test used for analysis of deviance. One of:
'Rao', 'LRT', 'Chisq' (default), 'F', 'Cp'.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[MASS]{glm.nb}}.}
}
\value{
a list of data frames containing the analysis of deviance (AOD)
and linear model estimates.
}
\description{
Compares expression differences between multiple groups with
negative binomial modeling. The method is particularly useful in finding
}
\details{
If a confounding variable is specified it will be included in the
model as a leading fixed term. Statistics for the confounding term are
not shown in the output. Effect size for the analysis of deviance (AOD),
eta-squared statistic, is computed as the ratio of deviance to total deviance.
Effect sizes for NB model coefficients, Cohen's d metrics, are calculated as
ratios of betas to their errors:

\deqn{d_{i} = \frac{\beta{1}}{SEM_{i}}}

As such, these effects size statistics are likely too optimistic, and in
case of models with a confounder, potentially biased. Please treat them with
caution as raw estimates of the effect size.
}
