% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_two_groups}
\alias{test_two_groups}
\title{Test gene expression differences between two groups.}
\usage{
test_two_groups(
  data,
  split_fct,
  variables,
  type = c("t", "wilcox"),
  adj_method = "BH",
  .parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with the expression values.}

\item{split_fct}{name of the group-defining variable, needs to be a factor.}

\item{variables}{a vector with dependent variable names.}

\item{type}{test type, currently 't' for T test (default) or 'wilcox' for
the Wilcoxon/Mann-Whitney test.}

\item{adj_method}{p value adjustment method,
see: \code{\link[stats]{p.adjust}}.}

\item{.parallel}{should the testing be run in parallel?}

\item{...}{additional arguments passed to \code{\link[stats]{t.test}} or
\code{\link[stats]{wilcox.test}}.}
}
\value{
\code{NULL} with a warning.
}
\description{
The testing functions are not active anymore, because we offer a faster and
more versatile tools for high-throughput statistical hypothesis testing.
Please use functions from \code{fastTest} package instead
(https://github.com/PiotrTymoszuk/fastTest/)
}
\details{
These functions are not active anymore and return \code{NULL} with a warning.
}
