% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_functions.R
\name{wrap_vector}
\alias{wrap_vector}
\alias{wrap_text}
\title{Wrap text with a defined word count per line.}
\usage{
wrap_vector(x, len = 5, word_sep = ", ", line_sep = ",\\n")

wrap_text(x, len = 5, word_sep = " ", line_sep = "\\n")
}
\arguments{
\item{x}{a vector or a text}

\item{len}{line length in words.}

\item{word_sep}{word separator, comma (\code{wrap_vec()}) or
space (\code{wrap_text()}) by default.}

\item{line_sep}{line separator, comma + newline (\code{wrap_vector()}) or
newline (\code{wrap_text()}) by default.}
}
\value{
a text value.
}
\description{
Converts a character vector or a text chunk into a text with the defined line
length (word count). If a non-character vector is provided, it will be
coerced to a character vector.
}
